% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsnicarF_2017.R
\name{AsnicarF_2017}
\alias{AsnicarF_2017}
\alias{AsnicarF_2017.genefamilies_relab.milk}
\alias{AsnicarF_2017.genefamilies_relab.stool}
\alias{AsnicarF_2017.marker_abundance.milk}
\alias{AsnicarF_2017.marker_abundance.stool}
\alias{AsnicarF_2017.marker_presence.milk}
\alias{AsnicarF_2017.marker_presence.stool}
\alias{AsnicarF_2017.metaphlan_bugs_list.milk}
\alias{AsnicarF_2017.metaphlan_bugs_list.stool}
\alias{AsnicarF_2017.pathabundance_relab.milk}
\alias{AsnicarF_2017.pathabundance_relab.stool}
\alias{AsnicarF_2017.pathcoverage.milk}
\alias{AsnicarF_2017.pathcoverage.stool}
\title{Data from the AsnicarF_2017 study}
\description{
Data from the AsnicarF_2017 study
}
\section{Datasets}{
 

\subsection{AsnicarF_2017.genefamilies_relab.milk}{
   An ExpressionSet with 8 samples and 752,977 features specific to the milk body site
}

\subsection{AsnicarF_2017.genefamilies_relab.stool}{
   An ExpressionSet with 16 samples and 752,977 features specific to the stool body site
}

\subsection{AsnicarF_2017.marker_abundance.milk}{
   An ExpressionSet with 8 samples and 53,228 features specific to the milk body site
}

\subsection{AsnicarF_2017.marker_abundance.stool}{
   An ExpressionSet with 16 samples and 53,228 features specific to the stool body site
}

\subsection{AsnicarF_2017.marker_presence.milk}{
   An ExpressionSet with 8 samples and 50,192 features specific to the milk body site
}

\subsection{AsnicarF_2017.marker_presence.stool}{
   An ExpressionSet with 16 samples and 50,192 features specific to the stool body site
}

\subsection{AsnicarF_2017.metaphlan_bugs_list.milk}{
   An ExpressionSet with 8 samples and 799 features specific to the milk body site
}

\subsection{AsnicarF_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 16 samples and 799 features specific to the stool body site
}

\subsection{AsnicarF_2017.pathabundance_relab.milk}{
   An ExpressionSet with 8 samples and 9,274 features specific to the milk body site
}

\subsection{AsnicarF_2017.pathabundance_relab.stool}{
   An ExpressionSet with 16 samples and 9,274 features specific to the stool body site
}

\subsection{AsnicarF_2017.pathcoverage.milk}{
   An ExpressionSet with 8 samples and 9,274 features specific to the milk body site
}

\subsection{AsnicarF_2017.pathcoverage.stool}{
   An ExpressionSet with 16 samples and 9,274 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Studying Vertical Microbiome Transmission from Mothers to Infants by Strain-Level Metagenomic Profiling.
}

\subsection{Author}{
   Asnicar F, Manara S, Zolfo M, Truong DT, Scholz M, Armanini F, Ferretti P, Gorfer V, Pedrotti A, Tett A, Segata N
}

\subsection{Lab}{
   [1] Centre for Integrative Biology, University of Trento, Trento, Italy., [2] Azienda Provinciale per i Servizi Sanitari, Trento, Italy.
}

\subsection{PMID}{
   28144631
}
}

\examples{
AsnicarF_2017.metaphlan_bugs_list.milk()

}
