% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bengtsson-PalmeJ_2015.R
\name{Bengtsson-PalmeJ_2015}
\alias{Bengtsson-PalmeJ_2015}
\alias{Bengtsson-PalmeJ_2015.genefamilies_relab.stool}
\alias{Bengtsson-PalmeJ_2015.marker_abundance.stool}
\alias{Bengtsson-PalmeJ_2015.marker_presence.stool}
\alias{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool}
\alias{Bengtsson-PalmeJ_2015.pathabundance_relab.stool}
\alias{Bengtsson-PalmeJ_2015.pathcoverage.stool}
\title{Data from the Bengtsson-PalmeJ_2015 study}
\description{
Data from the Bengtsson-PalmeJ_2015 study
}
\details{
Note that Bengtsson_PalmeJ_2015 is defunct, use Bengtsson-PalmeJ_2015 instead.
}
\section{Datasets}{


\subsection{Bengtsson-PalmeJ_2015.genefamilies_relab.stool}{
   An ExpressionSet with 70 samples and 1,301,172 features specific to the stool body site
}

\subsection{Bengtsson-PalmeJ_2015.marker_abundance.stool}{
   An ExpressionSet with 70 samples and 105,814 features specific to the stool body site
}

\subsection{Bengtsson-PalmeJ_2015.marker_presence.stool}{
   An ExpressionSet with 70 samples and 93,375 features specific to the stool body site
}

\subsection{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool}{
   An ExpressionSet with 70 samples and 1,220 features specific to the stool body site
}

\subsection{Bengtsson-PalmeJ_2015.pathabundance_relab.stool}{
   An ExpressionSet with 70 samples and 10,641 features specific to the stool body site
}

\subsection{Bengtsson-PalmeJ_2015.pathcoverage.stool}{
   An ExpressionSet with 70 samples and 10,641 features specific to the stool body site
}
}

\section{Source}{


\subsection{Title}{
   The Human Gut Microbiome as a Transporter of Antibiotic Resistance Genes between Continents.
}

\subsection{Author}{
   Bengtsson-Palme J, Angelin M, Huss M, Kjellqvist S, Kristiansson E, Palmgren H, Larsson DG, Johansson A
}

\subsection{Lab}{
   [1] Department of Infectious Diseases, Institute of Biomedicine, The Sahlgrenska Academy, University of Gothenburg, Gothenburg, Sweden., [2] Department of Clinical Microbiology, Infectious Diseases, Umea University, Umea, Sweden., [3] Science for Life Laboratory, Department of Biochemistry and Biophysics, Stockholm University, Solna, Sweden., [4] Department of Mathematical Sciences, Chalmers University of Technology, Gothenburg, Sweden., [5] Laboratory for Molecular Infection Medicine Sweden, Department of Clinical Microbiology, Bacteriology, Umea University, Umea, Sweden anders.f.johansson@umu.se.
}

\subsection{PMID}{
   26259788
}
}

\examples{
`Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool`()

}
