\name{ Candela_Africa.stool.eset }
\alias{ Candela_Africa.stool.marker_ab.eset }
\alias{ Candela_Africa.stool.marker_pres.eset }
\alias{ Candela_Africa.stool.abundance.eset }
\alias{ Candela_Africa.stool.pathCov.eset }
\alias{ Candela_Africa.stool.pathAbun.eset }
\alias{ Candela_Africa.stool.geneFam.eset }
\docType{data}
\title{ Metagenome Sequencing of the Hadza Hunter-Gatherer Gut Microbiota }
\description{ Through human microbiome sequencing, we can better understand how host evolutionary and ontogenetic history is reflected in the microbial function. However, there has been no information on the gut metagenome configuration in hunter-gatherer populations, posing a gap in our knowledge of gut microbiota (GM)-host mutualism arising from a lifestyle that describes over 90\% of human evolutionary history. Here, we present the first metagenomic analysis of GM from Hadza hunter-gatherers of Tanzania, showing a unique enrichment in metabolic pathways that aligns with the dietary and environmental factors characteristic of their foraging lifestyle. We found that the Hadza GM is adapted for broad-spectrum carbohydrate metabolism, reflecting the complex polysaccharides in their diet. Furthermore, the Hadza GM is equipped for branched-chain amino acid degradation and aromatic amino acid biosynthesis. Resistome functionality demonstrates the existence of antibiotic resistance genes in a population with little antibiotic exposure, indicating the ubiquitous presence of environmentally derived resistances. Our results demonstrate how the functional specificity of the GM correlates with certain environment and lifestyle factors and how complexity from the exogenous environment can be balanced by endogenous homeostasis. The Hadza gut metagenome structure allows us to appreciate the co-adaptive functional role of the GM in complementing the human physiology, providing a better understanding of the versatility of human life and subsistence. }
\usage{data( Candela_Africa.stool.marker_ab.eset )}
\format{
\preformatted{
experimentData(eset):
Experiment data
  Experimenter name: Rampelli S, Schnorr SL, Consolandi C, Turroni S, Severgnini M, Peano C, Brigidi P, Crittenden AN, Henry AG, Candela M 
  Laboratory: Department of Pharmacy and Biotechnology, University of Bologna, Bologna 40126, Italy. 
  Contact information:  
  Title: Metagenome Sequencing of the Hadza Hunter-Gatherer Gut Microbiota 
  URL:  
  PMIDs: 25981789 

  Abstract: A 213 word abstract is available. Use 'abstract' method.
  notes:
   platform_summary:     
      NA
   platform_accession:     
      NA
   platform_title:     
      GAIIx
   platform_technology:     
      Illumina
   platform_distribution:     
      NA
   platform_manufacturer:     
      Illumina
   processor:     
      MetaPhlAn2 (marker & abundance) / HUMAnN2 (geneFam., pathAbun., & pathCov.)

featureData(eset):
An object of class 'AnnotatedDataFrame': none

}}
\details{
\preformatted{
assayData: 302989 features, 38 samples
Platform type: NA
--------------------------- 
Available sample meta-data: 
--------------------------- 

dataset_name: 
Candela_Africa 
            38 

sampleID: 
   Length     Class      Mode 
       38 character character 

subjectID: 
   Length     Class      Mode 
       38 character character 

bodysite: 
stool 
   38 

disease: 
 n 
38 

age: 
   Length     Class      Mode 
       38 character character 

gender: 
female   male 
    16     22 

country: 
   italy tanzania 
      11       27 

sequencing_technology: 
Illumina 
      38 

pubmedid: 
25981789 
      38 

camp: 
bologna dedauko sengele 
     11      20       7 

}}
\keyword{datasets}
