% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChngKR_2016.R
\name{ChngKR_2016}
\alias{ChngKR_2016}
\alias{ChngKR_2016.genefamilies_relab.skin}
\alias{ChngKR_2016.marker_abundance.skin}
\alias{ChngKR_2016.marker_presence.skin}
\alias{ChngKR_2016.metaphlan_bugs_list.skin}
\alias{ChngKR_2016.pathabundance_relab.skin}
\alias{ChngKR_2016.pathcoverage.skin}
\title{Data from the ChngKR_2016 study}
\description{
Data from the ChngKR_2016 study
}
\section{Datasets}{
 

\subsection{ChngKR_2016.genefamilies_relab.skin}{
   An ExpressionSet with 78 samples and 823,859 features specific to the skin body site
}

\subsection{ChngKR_2016.marker_abundance.skin}{
   An ExpressionSet with 78 samples and 89,589 features specific to the skin body site
}

\subsection{ChngKR_2016.marker_presence.skin}{
   An ExpressionSet with 78 samples and 85,656 features specific to the skin body site
}

\subsection{ChngKR_2016.metaphlan_bugs_list.skin}{
   An ExpressionSet with 78 samples and 1,219 features specific to the skin body site
}

\subsection{ChngKR_2016.pathabundance_relab.skin}{
   An ExpressionSet with 78 samples and 10,995 features specific to the skin body site
}

\subsection{ChngKR_2016.pathcoverage.skin}{
   An ExpressionSet with 78 samples and 10,995 features specific to the skin body site
}
}

\section{Source}{
 

\subsection{Title}{
   Whole metagenome profiling reveals skin microbiome-dependent susceptibility to atopic dermatitis flare.
}

\subsection{Author}{
   Chng KR, Tay AS, Li C, Ng AH, Wang J, Suri BK, Matta SA, McGovern N, Janela B, Wong XF, Sio YY, Au BV, Wilm A, De Sessions PF, Lim TC, Tang MB, Ginhoux F, Connolly JE, Lane EB, Chew FT, Common JE, Nagarajan N
}

\subsection{Lab}{
   [1] Genome Institute of Singapore, Singapore 138672, Singapore., [2] Institute of Medical Biology, Singapore 138648, Singapore., [3] Institute of Molecular and Cell Biology, Singapore 138673, Singapore., [4] Department of Neurology, The First Affiliated Hospital of Zhengzhou University, Zhengzhou, Henan 450001, China., [5] Institute of Biomedical Studies, Baylor University, Waco, Texas 76798, USA., [6] Department of Biological Sciences, National University of Singapore, Singapore 117543., [7] Singapore Immunology Network, Singapore 138648, Singapore., [8] Division of Plastic, Reconstructive &amp;Aesthetic Surgery, National University Health System, Singapore 119074, Singapore., [9] National Skin Centre, Singapore 308205, Singapore., [10] Department of Microbiology and Immunology, National University of Singapore, Singapore 117545, Singapore.
}

\subsection{PMID}{
   27562258
}
}

\examples{
`ChngKR_2016.metaphlan_bugs_list.skin`()

}
