% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{cmdValidVersions}
\alias{cmdValidVersions}
\title{Get valid data versions}
\usage{
cmdValidVersions()
}
\value{
An integer vector of data versions, in the format YYYYMMDD, read from inst/extdata/versions.txt.
}
\description{
Get valid data versions
}
\examples{
cmdValidVersions()
max(cmdValidVersions())  #latest version
stopifnot(is(cmdValidVersions(), "integer"))
}
