% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{cmg_tmpdir}
\alias{cmg_tmpdir}
\title{create a 'globally usable' temporary folder}
\usage{
cmg_tmpdir(path = tempdir())
}
\arguments{
\item{path}{character(1) path to a temporary folder}
}
\description{
create a 'globally usable' temporary folder
}
