% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_metadata.R
\docType{data}
\name{combined_metadata}
\alias{combined_metadata}
\title{Combined metadata of all studies in curatedMetagenomicData}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2875 rows and 128 columns.}
\source{
See dataset specific help functions for source information
}
\usage{
combined_metadata
}
\description{
The combined sample level metadata of all studies in curatedMetagenomicData.
}
\keyword{datasets}
