% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData-defunct.R
\name{curatedMetagenomicData-defunct}
\alias{curatedMetagenomicData-defunct}
\alias{cmdValidVersions}
\alias{getMetaphlanTree}
\alias{ExpressionSet2MRexperiment}
\alias{ExpressionSet2phyloseq}
\alias{AsnicarF_2017.genefamilies_relab.milk}
\alias{AsnicarF_2017.genefamilies_relab.stool}
\alias{AsnicarF_2017.marker_abundance.milk}
\alias{AsnicarF_2017.marker_abundance.stool}
\alias{AsnicarF_2017.marker_presence.milk}
\alias{AsnicarF_2017.marker_presence.stool}
\alias{AsnicarF_2017.metaphlan_bugs_list.milk}
\alias{AsnicarF_2017.metaphlan_bugs_list.stool}
\alias{AsnicarF_2017.pathabundance_relab.milk}
\alias{AsnicarF_2017.pathabundance_relab.stool}
\alias{AsnicarF_2017.pathcoverage.milk}
\alias{AsnicarF_2017.pathcoverage.stool}
\alias{BackhedF_2015.genefamilies_relab.stool}
\alias{BackhedF_2015.marker_abundance.stool}
\alias{BackhedF_2015.marker_presence.stool}
\alias{BackhedF_2015.metaphlan_bugs_list.stool}
\alias{BackhedF_2015.pathabundance_relab.stool}
\alias{BackhedF_2015.pathcoverage.stool}
\alias{Bengtsson-PalmeJ_2015.genefamilies_relab.stool}
\alias{Bengtsson-PalmeJ_2015.marker_abundance.stool}
\alias{Bengtsson-PalmeJ_2015.marker_presence.stool}
\alias{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool}
\alias{Bengtsson-PalmeJ_2015.pathabundance_relab.stool}
\alias{Bengtsson-PalmeJ_2015.pathcoverage.stool}
\alias{BritoIL_2016.genefamilies_relab.oralcavity}
\alias{BritoIL_2016.genefamilies_relab.stool}
\alias{BritoIL_2016.marker_abundance.oralcavity}
\alias{BritoIL_2016.marker_abundance.stool}
\alias{BritoIL_2016.marker_presence.oralcavity}
\alias{BritoIL_2016.marker_presence.stool}
\alias{BritoIL_2016.metaphlan_bugs_list.oralcavity}
\alias{BritoIL_2016.metaphlan_bugs_list.stool}
\alias{BritoIL_2016.pathabundance_relab.oralcavity}
\alias{BritoIL_2016.pathabundance_relab.stool}
\alias{BritoIL_2016.pathcoverage.oralcavity}
\alias{BritoIL_2016.pathcoverage.stool}
\alias{Castro_NallarE_2015.genefamilies_relab.oralcavity}
\alias{Castro_NallarE_2015.marker_abundance.oralcavity}
\alias{Castro_NallarE_2015.marker_presence.oralcavity}
\alias{Castro_NallarE_2015.metaphlan_bugs_list.oralcavity}
\alias{Castro_NallarE_2015.pathabundance_relab.oralcavity}
\alias{Castro_NallarE_2015.pathcoverage.oralcavity}
\alias{Castro-NallarE_2015.genefamilies_relab.oralcavity}
\alias{Castro-NallarE_2015.marker_abundance.oralcavity}
\alias{Castro-NallarE_2015.marker_presence.oralcavity}
\alias{Castro-NallarE_2015.metaphlan_bugs_list.oralcavity}
\alias{Castro-NallarE_2015.pathabundance_relab.oralcavity}
\alias{Castro-NallarE_2015.pathcoverage.oralcavity}
\alias{ChengpingW_2017.genefamilies_relab.stool}
\alias{ChengpingW_2017.marker_abundance.stool}
\alias{ChengpingW_2017.marker_presence.stool}
\alias{ChengpingW_2017.metaphlan_bugs_list.stool}
\alias{ChengpingW_2017.pathabundance_relab.stool}
\alias{ChengpingW_2017.pathcoverage.stool}
\alias{ChngKR_2016.genefamilies_relab.skin}
\alias{ChngKR_2016.marker_abundance.skin}
\alias{ChngKR_2016.marker_presence.skin}
\alias{ChngKR_2016.metaphlan_bugs_list.skin}
\alias{ChngKR_2016.pathabundance_relab.skin}
\alias{ChngKR_2016.pathcoverage.skin}
\alias{CosteaPI_2017.genefamilies_relab.stool}
\alias{CosteaPI_2017.marker_abundance.stool}
\alias{CosteaPI_2017.marker_presence.stool}
\alias{CosteaPI_2017.metaphlan_bugs_list.stool}
\alias{CosteaPI_2017.pathabundance_relab.stool}
\alias{CosteaPI_2017.pathcoverage.stool}
\alias{DavidLA_2015.genefamilies_relab.stool}
\alias{DavidLA_2015.marker_abundance.stool}
\alias{DavidLA_2015.marker_presence.stool}
\alias{DavidLA_2015.metaphlan_bugs_list.stool}
\alias{DavidLA_2015.pathabundance_relab.stool}
\alias{DavidLA_2015.pathcoverage.stool}
\alias{DhakanDB_2019.genefamilies_relab.stool}
\alias{DhakanDB_2019.marker_abundance.stool}
\alias{DhakanDB_2019.marker_presence.stool}
\alias{DhakanDB_2019.metaphlan_bugs_list.stool}
\alias{DhakanDB_2019.pathabundance_relab.stool}
\alias{DhakanDB_2019.pathcoverage.stool}
\alias{FengQ_2015.genefamilies_relab.stool}
\alias{FengQ_2015.marker_abundance.stool}
\alias{FengQ_2015.marker_presence.stool}
\alias{FengQ_2015.metaphlan_bugs_list.stool}
\alias{FengQ_2015.pathabundance_relab.stool}
\alias{FengQ_2015.pathcoverage.stool}
\alias{FerrettiP_2018.genefamilies_relab.oralcavity}
\alias{FerrettiP_2018.genefamilies_relab.skin}
\alias{FerrettiP_2018.genefamilies_relab.stool}
\alias{FerrettiP_2018.genefamilies_relab.vagina}
\alias{FerrettiP_2018.marker_abundance.oralcavity}
\alias{FerrettiP_2018.marker_abundance.skin}
\alias{FerrettiP_2018.marker_abundance.stool}
\alias{FerrettiP_2018.marker_abundance.vagina}
\alias{FerrettiP_2018.marker_presence.oralcavity}
\alias{FerrettiP_2018.marker_presence.skin}
\alias{FerrettiP_2018.marker_presence.stool}
\alias{FerrettiP_2018.marker_presence.vagina}
\alias{FerrettiP_2018.metaphlan_bugs_list.oralcavity}
\alias{FerrettiP_2018.metaphlan_bugs_list.skin}
\alias{FerrettiP_2018.metaphlan_bugs_list.stool}
\alias{FerrettiP_2018.metaphlan_bugs_list.vagina}
\alias{FerrettiP_2018.pathabundance_relab.oralcavity}
\alias{FerrettiP_2018.pathabundance_relab.skin}
\alias{FerrettiP_2018.pathabundance_relab.stool}
\alias{FerrettiP_2018.pathabundance_relab.vagina}
\alias{FerrettiP_2018.pathcoverage.oralcavity}
\alias{FerrettiP_2018.pathcoverage.skin}
\alias{FerrettiP_2018.pathcoverage.stool}
\alias{FerrettiP_2018.pathcoverage.vagina}
\alias{GopalakrishnanV_2018.genefamilies_relab.stool}
\alias{GopalakrishnanV_2018.marker_abundance.stool}
\alias{GopalakrishnanV_2018.marker_presence.stool}
\alias{GopalakrishnanV_2018.metaphlan_bugs_list.stool}
\alias{GopalakrishnanV_2018.pathabundance_relab.stool}
\alias{GopalakrishnanV_2018.pathcoverage.stool}
\alias{HanniganGD_2017.genefamilies_relab.stool}
\alias{HanniganGD_2017.marker_abundance.stool}
\alias{HanniganGD_2017.marker_presence.stool}
\alias{HanniganGD_2017.metaphlan_bugs_list.stool}
\alias{HanniganGD_2017.pathabundance_relab.stool}
\alias{HanniganGD_2017.pathcoverage.stool}
\alias{HansenLBS_2018.genefamilies_relab.stool}
\alias{HansenLBS_2018.marker_abundance.stool}
\alias{HansenLBS_2018.marker_presence.stool}
\alias{HansenLBS_2018.metaphlan_bugs_list.stool}
\alias{HansenLBS_2018.pathabundance_relab.stool}
\alias{HansenLBS_2018.pathcoverage.stool}
\alias{Heitz_BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz_BuschartA_2016.marker_abundance.stool}
\alias{Heitz_BuschartA_2016.marker_presence.stool}
\alias{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz_BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz_BuschartA_2016.pathcoverage.stool}
\alias{Heitz-BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz-BuschartA_2016.marker_abundance.stool}
\alias{Heitz-BuschartA_2016.marker_presence.stool}
\alias{Heitz-BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz-BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz-BuschartA_2016.pathcoverage.stool}
\alias{HMP_2012.genefamilies_relab.nasalcavity}
\alias{HMP_2012.genefamilies_relab.oralcavity}
\alias{HMP_2012.genefamilies_relab.skin}
\alias{HMP_2012.genefamilies_relab.stool}
\alias{HMP_2012.genefamilies_relab.vagina}
\alias{HMP_2012.marker_abundance.nasalcavity}
\alias{HMP_2012.marker_abundance.oralcavity}
\alias{HMP_2012.marker_abundance.skin}
\alias{HMP_2012.marker_abundance.stool}
\alias{HMP_2012.marker_abundance.vagina}
\alias{HMP_2012.marker_presence.nasalcavity}
\alias{HMP_2012.marker_presence.oralcavity}
\alias{HMP_2012.marker_presence.skin}
\alias{HMP_2012.marker_presence.stool}
\alias{HMP_2012.marker_presence.vagina}
\alias{HMP_2012.metaphlan_bugs_list.nasalcavity}
\alias{HMP_2012.metaphlan_bugs_list.oralcavity}
\alias{HMP_2012.metaphlan_bugs_list.skin}
\alias{HMP_2012.metaphlan_bugs_list.stool}
\alias{HMP_2012.metaphlan_bugs_list.vagina}
\alias{HMP_2012.pathabundance_relab.nasalcavity}
\alias{HMP_2012.pathabundance_relab.oralcavity}
\alias{HMP_2012.pathabundance_relab.skin}
\alias{HMP_2012.pathabundance_relab.stool}
\alias{HMP_2012.pathabundance_relab.vagina}
\alias{HMP_2012.pathcoverage.nasalcavity}
\alias{HMP_2012.pathcoverage.oralcavity}
\alias{HMP_2012.pathcoverage.skin}
\alias{HMP_2012.pathcoverage.stool}
\alias{HMP_2012.pathcoverage.vagina}
\alias{JieZ_2017.genefamilies_relab.stool}
\alias{JieZ_2017.marker_abundance.stool}
\alias{JieZ_2017.marker_presence.stool}
\alias{JieZ_2017.metaphlan_bugs_list.stool}
\alias{JieZ_2017.pathabundance_relab.stool}
\alias{JieZ_2017.pathcoverage.stool}
\alias{KarlssonFH_2013.genefamilies_relab.stool}
\alias{KarlssonFH_2013.marker_abundance.stool}
\alias{KarlssonFH_2013.marker_presence.stool}
\alias{KarlssonFH_2013.metaphlan_bugs_list.stool}
\alias{KarlssonFH_2013.pathabundance_relab.stool}
\alias{KarlssonFH_2013.pathcoverage.stool}
\alias{KieserS_2018.genefamilies_relab.stool}
\alias{KieserS_2018.marker_abundance.stool}
\alias{KieserS_2018.marker_presence.stool}
\alias{KieserS_2018.metaphlan_bugs_list.stool}
\alias{KieserS_2018.pathabundance_relab.stool}
\alias{KieserS_2018.pathcoverage.stool}
\alias{KosticAD_2015.genefamilies_relab.stool}
\alias{KosticAD_2015.marker_abundance.stool}
\alias{KosticAD_2015.marker_presence.stool}
\alias{KosticAD_2015.metaphlan_bugs_list.stool}
\alias{KosticAD_2015.pathabundance_relab.stool}
\alias{KosticAD_2015.pathcoverage.stool}
\alias{LeChatelierE_2013.genefamilies_relab.stool}
\alias{LeChatelierE_2013.marker_abundance.stool}
\alias{LeChatelierE_2013.marker_presence.stool}
\alias{LeChatelierE_2013.metaphlan_bugs_list.stool}
\alias{LeChatelierE_2013.pathabundance_relab.stool}
\alias{LeChatelierE_2013.pathcoverage.stool}
\alias{LiJ_2014.genefamilies_relab.stool}
\alias{LiJ_2014.marker_abundance.stool}
\alias{LiJ_2014.marker_presence.stool}
\alias{LiJ_2014.metaphlan_bugs_list.stool}
\alias{LiJ_2014.pathabundance_relab.stool}
\alias{LiJ_2014.pathcoverage.stool}
\alias{LiJ_2017.genefamilies_relab.stool}
\alias{LiJ_2017.marker_abundance.stool}
\alias{LiJ_2017.marker_presence.stool}
\alias{LiJ_2017.metaphlan_bugs_list.stool}
\alias{LiJ_2017.pathabundance_relab.stool}
\alias{LiJ_2017.pathcoverage.stool}
\alias{LiSS_2016.genefamilies_relab.stool}
\alias{LiSS_2016.marker_abundance.stool}
\alias{LiSS_2016.marker_presence.stool}
\alias{LiSS_2016.metaphlan_bugs_list.stool}
\alias{LiSS_2016.pathabundance_relab.stool}
\alias{LiSS_2016.pathcoverage.stool}
\alias{LiuW_2016.genefamilies_relab.stool}
\alias{LiuW_2016.marker_abundance.stool}
\alias{LiuW_2016.marker_presence.stool}
\alias{LiuW_2016.metaphlan_bugs_list.stool}
\alias{LiuW_2016.pathabundance_relab.stool}
\alias{LiuW_2016.pathcoverage.stool}
\alias{LomanNJ_2013.genefamilies_relab.stool}
\alias{LomanNJ_2013.marker_abundance.stool}
\alias{LomanNJ_2013.marker_presence.stool}
\alias{LomanNJ_2013.metaphlan_bugs_list.stool}
\alias{LomanNJ_2013.pathabundance_relab.stool}
\alias{LomanNJ_2013.pathcoverage.stool}
\alias{LoombaR_2017.genefamilies_relab.stool}
\alias{LoombaR_2017.marker_abundance.stool}
\alias{LoombaR_2017.marker_presence.stool}
\alias{LoombaR_2017.metaphlan_bugs_list.stool}
\alias{LoombaR_2017.pathabundance_relab.stool}
\alias{LoombaR_2017.pathcoverage.stool}
\alias{LouisS_2016.genefamilies_relab.stool}
\alias{LouisS_2016.marker_abundance.stool}
\alias{LouisS_2016.marker_presence.stool}
\alias{LouisS_2016.metaphlan_bugs_list.stool}
\alias{LouisS_2016.pathabundance_relab.stool}
\alias{LouisS_2016.pathcoverage.stool}
\alias{MatsonV_2018.genefamilies_relab.stool}
\alias{MatsonV_2018.marker_abundance.stool}
\alias{MatsonV_2018.marker_presence.stool}
\alias{MatsonV_2018.metaphlan_bugs_list.stool}
\alias{MatsonV_2018.pathabundance_relab.stool}
\alias{MatsonV_2018.pathcoverage.stool}
\alias{NielsenHB_2014.genefamilies_relab.stool}
\alias{NielsenHB_2014.marker_abundance.stool}
\alias{NielsenHB_2014.marker_presence.stool}
\alias{NielsenHB_2014.metaphlan_bugs_list.stool}
\alias{NielsenHB_2014.pathabundance_relab.stool}
\alias{NielsenHB_2014.pathcoverage.stool}
\alias{Obregon_TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon_TitoAJ_2015.marker_abundance.stool}
\alias{Obregon_TitoAJ_2015.marker_presence.stool}
\alias{Obregon_TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon_TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon_TitoAJ_2015.pathcoverage.stool}
\alias{Obregon-TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon-TitoAJ_2015.marker_abundance.stool}
\alias{Obregon-TitoAJ_2015.marker_presence.stool}
\alias{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon-TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon-TitoAJ_2015.pathcoverage.stool}
\alias{OhJ_2014.genefamilies_relab.skin}
\alias{OhJ_2014.marker_abundance.skin}
\alias{OhJ_2014.marker_presence.skin}
\alias{OhJ_2014.metaphlan_bugs_list.skin}
\alias{OhJ_2014.pathabundance_relab.skin}
\alias{OhJ_2014.pathcoverage.skin}
\alias{OlmMR_2017.genefamilies_relab.oralcavity}
\alias{OlmMR_2017.genefamilies_relab.skin}
\alias{OlmMR_2017.genefamilies_relab.stool}
\alias{OlmMR_2017.marker_abundance.oralcavity}
\alias{OlmMR_2017.marker_abundance.skin}
\alias{OlmMR_2017.marker_abundance.stool}
\alias{OlmMR_2017.marker_presence.oralcavity}
\alias{OlmMR_2017.marker_presence.skin}
\alias{OlmMR_2017.marker_presence.stool}
\alias{OlmMR_2017.metaphlan_bugs_list.oralcavity}
\alias{OlmMR_2017.metaphlan_bugs_list.skin}
\alias{OlmMR_2017.metaphlan_bugs_list.stool}
\alias{OlmMR_2017.pathabundance_relab.oralcavity}
\alias{OlmMR_2017.pathabundance_relab.skin}
\alias{OlmMR_2017.pathabundance_relab.stool}
\alias{OlmMR_2017.pathcoverage.oralcavity}
\alias{OlmMR_2017.pathcoverage.skin}
\alias{OlmMR_2017.pathcoverage.stool}
\alias{PasolliE_2018.genefamilies_relab.stool}
\alias{PasolliE_2018.marker_abundance.stool}
\alias{PasolliE_2018.marker_presence.stool}
\alias{PasolliE_2018.metaphlan_bugs_list.stool}
\alias{PasolliE_2018.pathabundance_relab.stool}
\alias{PasolliE_2018.pathcoverage.stool}
\alias{PehrssonE_2016.genefamilies_relab.stool}
\alias{PehrssonE_2016.marker_abundance.stool}
\alias{PehrssonE_2016.marker_presence.stool}
\alias{PehrssonE_2016.metaphlan_bugs_list.stool}
\alias{PehrssonE_2016.pathabundance_relab.stool}
\alias{PehrssonE_2016.pathcoverage.stool}
\alias{QinJ_2012.genefamilies_relab.stool}
\alias{QinJ_2012.marker_abundance.stool}
\alias{QinJ_2012.marker_presence.stool}
\alias{QinJ_2012.metaphlan_bugs_list.stool}
\alias{QinJ_2012.pathabundance_relab.stool}
\alias{QinJ_2012.pathcoverage.stool}
\alias{QinN_2014.genefamilies_relab.stool}
\alias{QinN_2014.marker_abundance.stool}
\alias{QinN_2014.marker_presence.stool}
\alias{QinN_2014.metaphlan_bugs_list.stool}
\alias{QinN_2014.pathabundance_relab.stool}
\alias{QinN_2014.pathcoverage.stool}
\alias{RampelliS_2015.genefamilies_relab.stool}
\alias{RampelliS_2015.marker_abundance.stool}
\alias{RampelliS_2015.marker_presence.stool}
\alias{RampelliS_2015.metaphlan_bugs_list.stool}
\alias{RampelliS_2015.pathabundance_relab.stool}
\alias{RampelliS_2015.pathcoverage.stool}
\alias{RaymondF_2016.genefamilies_relab.stool}
\alias{RaymondF_2016.marker_abundance.stool}
\alias{RaymondF_2016.marker_presence.stool}
\alias{RaymondF_2016.metaphlan_bugs_list.stool}
\alias{RaymondF_2016.pathabundance_relab.stool}
\alias{RaymondF_2016.pathcoverage.stool}
\alias{SchirmerM_2016.genefamilies_relab.stool}
\alias{SchirmerM_2016.marker_abundance.stool}
\alias{SchirmerM_2016.marker_presence.stool}
\alias{SchirmerM_2016.metaphlan_bugs_list.stool}
\alias{SchirmerM_2016.pathabundance_relab.stool}
\alias{SchirmerM_2016.pathcoverage.stool}
\alias{ShiB_2015.genefamilies_relab.oralcavity}
\alias{ShiB_2015.marker_abundance.oralcavity}
\alias{ShiB_2015.marker_presence.oralcavity}
\alias{ShiB_2015.metaphlan_bugs_list.oralcavity}
\alias{ShiB_2015.pathabundance_relab.oralcavity}
\alias{ShiB_2015.pathcoverage.oralcavity}
\alias{SmitsSA_2017.genefamilies_relab.stool}
\alias{SmitsSA_2017.marker_abundance.stool}
\alias{SmitsSA_2017.marker_presence.stool}
\alias{SmitsSA_2017.metaphlan_bugs_list.stool}
\alias{SmitsSA_2017.pathabundance_relab.stool}
\alias{SmitsSA_2017.pathcoverage.stool}
\alias{TettAJ_2016.genefamilies_relab.skin}
\alias{TettAJ_2016.marker_abundance.skin}
\alias{TettAJ_2016.marker_presence.skin}
\alias{TettAJ_2016.metaphlan_bugs_list.skin}
\alias{TettAJ_2016.pathabundance_relab.skin}
\alias{TettAJ_2016.pathcoverage.skin}
\alias{TettAJ_2019_a.genefamilies_relab.stool}
\alias{TettAJ_2019_a.marker_abundance.stool}
\alias{TettAJ_2019_a.marker_presence.stool}
\alias{TettAJ_2019_a.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_a.pathabundance_relab.stool}
\alias{TettAJ_2019_a.pathcoverage.stool}
\alias{TettAJ_2019_b.genefamilies_relab.stool}
\alias{TettAJ_2019_b.marker_abundance.stool}
\alias{TettAJ_2019_b.marker_presence.stool}
\alias{TettAJ_2019_b.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_b.pathabundance_relab.stool}
\alias{TettAJ_2019_b.pathcoverage.stool}
\alias{TettAJ_2019_c.genefamilies_relab.stool}
\alias{TettAJ_2019_c.marker_abundance.stool}
\alias{TettAJ_2019_c.marker_presence.stool}
\alias{TettAJ_2019_c.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_c.pathabundance_relab.stool}
\alias{TettAJ_2019_c.pathcoverage.stool}
\alias{ThomasAM_2018a.genefamilies_relab.stool}
\alias{ThomasAM_2018a.marker_abundance.stool}
\alias{ThomasAM_2018a.marker_presence.stool}
\alias{ThomasAM_2018a.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018a.pathabundance_relab.stool}
\alias{ThomasAM_2018a.pathcoverage.stool}
\alias{ThomasAM_2018b.genefamilies_relab.stool}
\alias{ThomasAM_2018b.marker_abundance.stool}
\alias{ThomasAM_2018b.marker_presence.stool}
\alias{ThomasAM_2018b.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018b.pathabundance_relab.stool}
\alias{ThomasAM_2018b.pathcoverage.stool}
\alias{VatanenT_2016.genefamilies_relab.stool}
\alias{VatanenT_2016.marker_abundance.stool}
\alias{VatanenT_2016.marker_presence.stool}
\alias{VatanenT_2016.metaphlan_bugs_list.stool}
\alias{VatanenT_2016.pathabundance_relab.stool}
\alias{VatanenT_2016.pathcoverage.stool}
\alias{VincentC_2016.genefamilies_relab.stool}
\alias{VincentC_2016.marker_abundance.stool}
\alias{VincentC_2016.marker_presence.stool}
\alias{VincentC_2016.metaphlan_bugs_list.stool}
\alias{VincentC_2016.pathabundance_relab.stool}
\alias{VincentC_2016.pathcoverage.stool}
\alias{VogtmannE_2016.genefamilies_relab.stool}
\alias{VogtmannE_2016.marker_abundance.stool}
\alias{VogtmannE_2016.marker_presence.stool}
\alias{VogtmannE_2016.metaphlan_bugs_list.stool}
\alias{VogtmannE_2016.pathabundance_relab.stool}
\alias{VogtmannE_2016.pathcoverage.stool}
\alias{XieH_2016.genefamilies_relab.stool}
\alias{XieH_2016.marker_abundance.stool}
\alias{XieH_2016.marker_presence.stool}
\alias{XieH_2016.metaphlan_bugs_list.stool}
\alias{XieH_2016.pathabundance_relab.stool}
\alias{XieH_2016.pathcoverage.stool}
\alias{YeZ_2018.genefamilies_relab.stool}
\alias{YeZ_2018.marker_abundance.stool}
\alias{YeZ_2018.marker_presence.stool}
\alias{YeZ_2018.metaphlan_bugs_list.stool}
\alias{YeZ_2018.pathabundance_relab.stool}
\alias{YeZ_2018.pathcoverage.stool}
\alias{YuJ_2015.genefamilies_relab.stool}
\alias{YuJ_2015.marker_abundance.stool}
\alias{YuJ_2015.marker_presence.stool}
\alias{YuJ_2015.metaphlan_bugs_list.stool}
\alias{YuJ_2015.pathabundance_relab.stool}
\alias{YuJ_2015.pathcoverage.stool}
\alias{ZeeviD_2015.genefamilies_relab.stool}
\alias{ZeeviD_2015.marker_abundance.stool}
\alias{ZeeviD_2015.marker_presence.stool}
\alias{ZeeviD_2015.metaphlan_bugs_list.stool}
\alias{ZeeviD_2015.pathabundance_relab.stool}
\alias{ZeeviD_2015.pathcoverage.stool}
\alias{ZellerG_2014.genefamilies_relab.stool}
\alias{ZellerG_2014.marker_abundance.stool}
\alias{ZellerG_2014.marker_presence.stool}
\alias{ZellerG_2014.metaphlan_bugs_list.stool}
\alias{ZellerG_2014.pathabundance_relab.stool}
\alias{ZellerG_2014.pathcoverage.stool}
\title{Defunct functions in package ‘curatedMetagenomicData’}
\value{
\itemize{
\item 
}
}
\description{
These functions are defunct and no longer available.
}
\details{
The following functions (methods) are defunct:
\itemize{
\item \code{cmdValidVersions()}
\item \code{getMetaphlanTree()}
\item \code{ExpressionSet2MRexperiment()}
\item \code{ExpressionSet2phyloseq()}
}

The following functions (accessors) are defunct:
\itemize{
\item \code{AsnicarF_2017.genefamilies_relab.milk()}
\item \code{AsnicarF_2017.genefamilies_relab.stool()}
\item \code{AsnicarF_2017.marker_abundance.milk()}
\item \code{AsnicarF_2017.marker_abundance.stool()}
\item \code{AsnicarF_2017.marker_presence.milk()}
\item \code{AsnicarF_2017.marker_presence.stool()}
\item \code{AsnicarF_2017.metaphlan_bugs_list.milk()}
\item \code{AsnicarF_2017.metaphlan_bugs_list.stool()}
\item \code{AsnicarF_2017.pathabundance_relab.milk()}
\item \code{AsnicarF_2017.pathabundance_relab.stool()}
\item \code{AsnicarF_2017.pathcoverage.milk()}
\item \code{AsnicarF_2017.pathcoverage.stool()}
\item \code{BackhedF_2015.genefamilies_relab.stool()}
\item \code{BackhedF_2015.marker_abundance.stool()}
\item \code{BackhedF_2015.marker_presence.stool()}
\item \code{BackhedF_2015.metaphlan_bugs_list.stool()}
\item \code{BackhedF_2015.pathabundance_relab.stool()}
\item \code{BackhedF_2015.pathcoverage.stool()}
\item \code{Bengtsson-PalmeJ_2015.genefamilies_relab.stool()}
\item \code{Bengtsson-PalmeJ_2015.marker_abundance.stool()}
\item \code{Bengtsson-PalmeJ_2015.marker_presence.stool()}
\item \code{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool()}
\item \code{Bengtsson-PalmeJ_2015.pathabundance_relab.stool()}
\item \code{Bengtsson-PalmeJ_2015.pathcoverage.stool()}
\item \code{BritoIL_2016.genefamilies_relab.oralcavity()}
\item \code{BritoIL_2016.genefamilies_relab.stool()}
\item \code{BritoIL_2016.marker_abundance.oralcavity()}
\item \code{BritoIL_2016.marker_abundance.stool()}
\item \code{BritoIL_2016.marker_presence.oralcavity()}
\item \code{BritoIL_2016.marker_presence.stool()}
\item \code{BritoIL_2016.metaphlan_bugs_list.oralcavity()}
\item \code{BritoIL_2016.metaphlan_bugs_list.stool()}
\item \code{BritoIL_2016.pathabundance_relab.oralcavity()}
\item \code{BritoIL_2016.pathabundance_relab.stool()}
\item \code{BritoIL_2016.pathcoverage.oralcavity()}
\item \code{BritoIL_2016.pathcoverage.stool()}
\item \code{Castro_NallarE_2015.genefamilies_relab.oralcavity()}
\item \code{Castro_NallarE_2015.marker_abundance.oralcavity()}
\item \code{Castro_NallarE_2015.marker_presence.oralcavity()}
\item \code{Castro_NallarE_2015.metaphlan_bugs_list.oralcavity()}
\item \code{Castro_NallarE_2015.pathabundance_relab.oralcavity()}
\item \code{Castro_NallarE_2015.pathcoverage.oralcavity()}
\item \code{Castro-NallarE_2015.genefamilies_relab.oralcavity()}
\item \code{Castro-NallarE_2015.marker_abundance.oralcavity()}
\item \code{Castro-NallarE_2015.marker_presence.oralcavity()}
\item \code{Castro-NallarE_2015.metaphlan_bugs_list.oralcavity()}
\item \code{Castro-NallarE_2015.pathabundance_relab.oralcavity()}
\item \code{Castro-NallarE_2015.pathcoverage.oralcavity()}
\item \code{ChengpingW_2017.genefamilies_relab.stool()}
\item \code{ChengpingW_2017.marker_abundance.stool()}
\item \code{ChengpingW_2017.marker_presence.stool()}
\item \code{ChengpingW_2017.metaphlan_bugs_list.stool()}
\item \code{ChengpingW_2017.pathabundance_relab.stool()}
\item \code{ChengpingW_2017.pathcoverage.stool()}
\item \code{ChngKR_2016.genefamilies_relab.skin()}
\item \code{ChngKR_2016.marker_abundance.skin()}
\item \code{ChngKR_2016.marker_presence.skin()}
\item \code{ChngKR_2016.metaphlan_bugs_list.skin()}
\item \code{ChngKR_2016.pathabundance_relab.skin()}
\item \code{ChngKR_2016.pathcoverage.skin()}
\item \code{CosteaPI_2017.genefamilies_relab.stool()}
\item \code{CosteaPI_2017.marker_abundance.stool()}
\item \code{CosteaPI_2017.marker_presence.stool()}
\item \code{CosteaPI_2017.metaphlan_bugs_list.stool()}
\item \code{CosteaPI_2017.pathabundance_relab.stool()}
\item \code{CosteaPI_2017.pathcoverage.stool()}
\item \code{DavidLA_2015.genefamilies_relab.stool()}
\item \code{DavidLA_2015.marker_abundance.stool()}
\item \code{DavidLA_2015.marker_presence.stool()}
\item \code{DavidLA_2015.metaphlan_bugs_list.stool()}
\item \code{DavidLA_2015.pathabundance_relab.stool()}
\item \code{DavidLA_2015.pathcoverage.stool()}
\item \code{DhakanDB_2019.genefamilies_relab.stool()}
\item \code{DhakanDB_2019.marker_abundance.stool()}
\item \code{DhakanDB_2019.marker_presence.stool()}
\item \code{DhakanDB_2019.metaphlan_bugs_list.stool()}
\item \code{DhakanDB_2019.pathabundance_relab.stool()}
\item \code{DhakanDB_2019.pathcoverage.stool()}
\item \code{FengQ_2015.genefamilies_relab.stool()}
\item \code{FengQ_2015.marker_abundance.stool()}
\item \code{FengQ_2015.marker_presence.stool()}
\item \code{FengQ_2015.metaphlan_bugs_list.stool()}
\item \code{FengQ_2015.pathabundance_relab.stool()}
\item \code{FengQ_2015.pathcoverage.stool()}
\item \code{FerrettiP_2018.genefamilies_relab.oralcavity()}
\item \code{FerrettiP_2018.genefamilies_relab.skin()}
\item \code{FerrettiP_2018.genefamilies_relab.stool()}
\item \code{FerrettiP_2018.genefamilies_relab.vagina()}
\item \code{FerrettiP_2018.marker_abundance.oralcavity()}
\item \code{FerrettiP_2018.marker_abundance.skin()}
\item \code{FerrettiP_2018.marker_abundance.stool()}
\item \code{FerrettiP_2018.marker_abundance.vagina()}
\item \code{FerrettiP_2018.marker_presence.oralcavity()}
\item \code{FerrettiP_2018.marker_presence.skin()}
\item \code{FerrettiP_2018.marker_presence.stool()}
\item \code{FerrettiP_2018.marker_presence.vagina()}
\item \code{FerrettiP_2018.metaphlan_bugs_list.oralcavity()}
\item \code{FerrettiP_2018.metaphlan_bugs_list.skin()}
\item \code{FerrettiP_2018.metaphlan_bugs_list.stool()}
\item \code{FerrettiP_2018.metaphlan_bugs_list.vagina()}
\item \code{FerrettiP_2018.pathabundance_relab.oralcavity()}
\item \code{FerrettiP_2018.pathabundance_relab.skin()}
\item \code{FerrettiP_2018.pathabundance_relab.stool()}
\item \code{FerrettiP_2018.pathabundance_relab.vagina()}
\item \code{FerrettiP_2018.pathcoverage.oralcavity()}
\item \code{FerrettiP_2018.pathcoverage.skin()}
\item \code{FerrettiP_2018.pathcoverage.stool()}
\item \code{FerrettiP_2018.pathcoverage.vagina()}
\item \code{GopalakrishnanV_2018.genefamilies_relab.stool()}
\item \code{GopalakrishnanV_2018.marker_abundance.stool()}
\item \code{GopalakrishnanV_2018.marker_presence.stool()}
\item \code{GopalakrishnanV_2018.metaphlan_bugs_list.stool()}
\item \code{GopalakrishnanV_2018.pathabundance_relab.stool()}
\item \code{GopalakrishnanV_2018.pathcoverage.stool()}
\item \code{HanniganGD_2017.genefamilies_relab.stool()}
\item \code{HanniganGD_2017.marker_abundance.stool()}
\item \code{HanniganGD_2017.marker_presence.stool()}
\item \code{HanniganGD_2017.metaphlan_bugs_list.stool()}
\item \code{HanniganGD_2017.pathabundance_relab.stool()}
\item \code{HanniganGD_2017.pathcoverage.stool()}
\item \code{HansenLBS_2018.genefamilies_relab.stool()}
\item \code{HansenLBS_2018.marker_abundance.stool()}
\item \code{HansenLBS_2018.marker_presence.stool()}
\item \code{HansenLBS_2018.metaphlan_bugs_list.stool()}
\item \code{HansenLBS_2018.pathabundance_relab.stool()}
\item \code{HansenLBS_2018.pathcoverage.stool()}
\item \code{Heitz_BuschartA_2016.genefamilies_relab.stool()}
\item \code{Heitz_BuschartA_2016.marker_abundance.stool()}
\item \code{Heitz_BuschartA_2016.marker_presence.stool()}
\item \code{Heitz_BuschartA_2016.metaphlan_bugs_list.stool()}
\item \code{Heitz_BuschartA_2016.pathabundance_relab.stool()}
\item \code{Heitz_BuschartA_2016.pathcoverage.stool()}
\item \code{Heitz-BuschartA_2016.genefamilies_relab.stool()}
\item \code{Heitz-BuschartA_2016.marker_abundance.stool()}
\item \code{Heitz-BuschartA_2016.marker_presence.stool()}
\item \code{Heitz-BuschartA_2016.metaphlan_bugs_list.stool()}
\item \code{Heitz-BuschartA_2016.pathabundance_relab.stool()}
\item \code{Heitz-BuschartA_2016.pathcoverage.stool()}
\item \code{HMP_2012.genefamilies_relab.nasalcavity()}
\item \code{HMP_2012.genefamilies_relab.oralcavity()}
\item \code{HMP_2012.genefamilies_relab.skin()}
\item \code{HMP_2012.genefamilies_relab.stool()}
\item \code{HMP_2012.genefamilies_relab.vagina()}
\item \code{HMP_2012.marker_abundance.nasalcavity()}
\item \code{HMP_2012.marker_abundance.oralcavity()}
\item \code{HMP_2012.marker_abundance.skin()}
\item \code{HMP_2012.marker_abundance.stool()}
\item \code{HMP_2012.marker_abundance.vagina()}
\item \code{HMP_2012.marker_presence.nasalcavity()}
\item \code{HMP_2012.marker_presence.oralcavity()}
\item \code{HMP_2012.marker_presence.skin()}
\item \code{HMP_2012.marker_presence.stool()}
\item \code{HMP_2012.marker_presence.vagina()}
\item \code{HMP_2012.metaphlan_bugs_list.nasalcavity()}
\item \code{HMP_2012.metaphlan_bugs_list.oralcavity()}
\item \code{HMP_2012.metaphlan_bugs_list.skin()}
\item \code{HMP_2012.metaphlan_bugs_list.stool()}
\item \code{HMP_2012.metaphlan_bugs_list.vagina()}
\item \code{HMP_2012.pathabundance_relab.nasalcavity()}
\item \code{HMP_2012.pathabundance_relab.oralcavity()}
\item \code{HMP_2012.pathabundance_relab.skin()}
\item \code{HMP_2012.pathabundance_relab.stool()}
\item \code{HMP_2012.pathabundance_relab.vagina()}
\item \code{HMP_2012.pathcoverage.nasalcavity()}
\item \code{HMP_2012.pathcoverage.oralcavity()}
\item \code{HMP_2012.pathcoverage.skin()}
\item \code{HMP_2012.pathcoverage.stool()}
\item \code{HMP_2012.pathcoverage.vagina()}
\item \code{JieZ_2017.genefamilies_relab.stool()}
\item \code{JieZ_2017.marker_abundance.stool()}
\item \code{JieZ_2017.marker_presence.stool()}
\item \code{JieZ_2017.metaphlan_bugs_list.stool()}
\item \code{JieZ_2017.pathabundance_relab.stool()}
\item \code{JieZ_2017.pathcoverage.stool()}
\item \code{KarlssonFH_2013.genefamilies_relab.stool()}
\item \code{KarlssonFH_2013.marker_abundance.stool()}
\item \code{KarlssonFH_2013.marker_presence.stool()}
\item \code{KarlssonFH_2013.metaphlan_bugs_list.stool()}
\item \code{KarlssonFH_2013.pathabundance_relab.stool()}
\item \code{KarlssonFH_2013.pathcoverage.stool()}
\item \code{KieserS_2018.genefamilies_relab.stool()}
\item \code{KieserS_2018.marker_abundance.stool()}
\item \code{KieserS_2018.marker_presence.stool()}
\item \code{KieserS_2018.metaphlan_bugs_list.stool()}
\item \code{KieserS_2018.pathabundance_relab.stool()}
\item \code{KieserS_2018.pathcoverage.stool()}
\item \code{KosticAD_2015.genefamilies_relab.stool()}
\item \code{KosticAD_2015.marker_abundance.stool()}
\item \code{KosticAD_2015.marker_presence.stool()}
\item \code{KosticAD_2015.metaphlan_bugs_list.stool()}
\item \code{KosticAD_2015.pathabundance_relab.stool()}
\item \code{KosticAD_2015.pathcoverage.stool()}
\item \code{LeChatelierE_2013.genefamilies_relab.stool()}
\item \code{LeChatelierE_2013.marker_abundance.stool()}
\item \code{LeChatelierE_2013.marker_presence.stool()}
\item \code{LeChatelierE_2013.metaphlan_bugs_list.stool()}
\item \code{LeChatelierE_2013.pathabundance_relab.stool()}
\item \code{LeChatelierE_2013.pathcoverage.stool()}
\item \code{LiJ_2014.genefamilies_relab.stool()}
\item \code{LiJ_2014.marker_abundance.stool()}
\item \code{LiJ_2014.marker_presence.stool()}
\item \code{LiJ_2014.metaphlan_bugs_list.stool()}
\item \code{LiJ_2014.pathabundance_relab.stool()}
\item \code{LiJ_2014.pathcoverage.stool()}
\item \code{LiJ_2017.genefamilies_relab.stool()}
\item \code{LiJ_2017.marker_abundance.stool()}
\item \code{LiJ_2017.marker_presence.stool()}
\item \code{LiJ_2017.metaphlan_bugs_list.stool()}
\item \code{LiJ_2017.pathabundance_relab.stool()}
\item \code{LiJ_2017.pathcoverage.stool()}
\item \code{LiSS_2016.genefamilies_relab.stool()}
\item \code{LiSS_2016.marker_abundance.stool()}
\item \code{LiSS_2016.marker_presence.stool()}
\item \code{LiSS_2016.metaphlan_bugs_list.stool()}
\item \code{LiSS_2016.pathabundance_relab.stool()}
\item \code{LiSS_2016.pathcoverage.stool()}
\item \code{LiuW_2016.genefamilies_relab.stool()}
\item \code{LiuW_2016.marker_abundance.stool()}
\item \code{LiuW_2016.marker_presence.stool()}
\item \code{LiuW_2016.metaphlan_bugs_list.stool()}
\item \code{LiuW_2016.pathabundance_relab.stool()}
\item \code{LiuW_2016.pathcoverage.stool()}
\item \code{LomanNJ_2013.genefamilies_relab.stool()}
\item \code{LomanNJ_2013.marker_abundance.stool()}
\item \code{LomanNJ_2013.marker_presence.stool()}
\item \code{LomanNJ_2013.metaphlan_bugs_list.stool()}
\item \code{LomanNJ_2013.pathabundance_relab.stool()}
\item \code{LomanNJ_2013.pathcoverage.stool()}
\item \code{LoombaR_2017.genefamilies_relab.stool()}
\item \code{LoombaR_2017.marker_abundance.stool()}
\item \code{LoombaR_2017.marker_presence.stool()}
\item \code{LoombaR_2017.metaphlan_bugs_list.stool()}
\item \code{LoombaR_2017.pathabundance_relab.stool()}
\item \code{LoombaR_2017.pathcoverage.stool()}
\item \code{LouisS_2016.genefamilies_relab.stool()}
\item \code{LouisS_2016.marker_abundance.stool()}
\item \code{LouisS_2016.marker_presence.stool()}
\item \code{LouisS_2016.metaphlan_bugs_list.stool()}
\item \code{LouisS_2016.pathabundance_relab.stool()}
\item \code{LouisS_2016.pathcoverage.stool()}
\item \code{MatsonV_2018.genefamilies_relab.stool()}
\item \code{MatsonV_2018.marker_abundance.stool()}
\item \code{MatsonV_2018.marker_presence.stool()}
\item \code{MatsonV_2018.metaphlan_bugs_list.stool()}
\item \code{MatsonV_2018.pathabundance_relab.stool()}
\item \code{MatsonV_2018.pathcoverage.stool()}
\item \code{NielsenHB_2014.genefamilies_relab.stool()}
\item \code{NielsenHB_2014.marker_abundance.stool()}
\item \code{NielsenHB_2014.marker_presence.stool()}
\item \code{NielsenHB_2014.metaphlan_bugs_list.stool()}
\item \code{NielsenHB_2014.pathabundance_relab.stool()}
\item \code{NielsenHB_2014.pathcoverage.stool()}
\item \code{Obregon_TitoAJ_2015.genefamilies_relab.stool()}
\item \code{Obregon_TitoAJ_2015.marker_abundance.stool()}
\item \code{Obregon_TitoAJ_2015.marker_presence.stool()}
\item \code{Obregon_TitoAJ_2015.metaphlan_bugs_list.stool()}
\item \code{Obregon_TitoAJ_2015.pathabundance_relab.stool()}
\item \code{Obregon_TitoAJ_2015.pathcoverage.stool()}
\item \code{Obregon-TitoAJ_2015.genefamilies_relab.stool()}
\item \code{Obregon-TitoAJ_2015.marker_abundance.stool()}
\item \code{Obregon-TitoAJ_2015.marker_presence.stool()}
\item \code{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool()}
\item \code{Obregon-TitoAJ_2015.pathabundance_relab.stool()}
\item \code{Obregon-TitoAJ_2015.pathcoverage.stool()}
\item \code{OhJ_2014.genefamilies_relab.skin()}
\item \code{OhJ_2014.marker_abundance.skin()}
\item \code{OhJ_2014.marker_presence.skin()}
\item \code{OhJ_2014.metaphlan_bugs_list.skin()}
\item \code{OhJ_2014.pathabundance_relab.skin()}
\item \code{OhJ_2014.pathcoverage.skin()}
\item \code{OlmMR_2017.genefamilies_relab.oralcavity()}
\item \code{OlmMR_2017.genefamilies_relab.skin()}
\item \code{OlmMR_2017.genefamilies_relab.stool()}
\item \code{OlmMR_2017.marker_abundance.oralcavity()}
\item \code{OlmMR_2017.marker_abundance.skin()}
\item \code{OlmMR_2017.marker_abundance.stool()}
\item \code{OlmMR_2017.marker_presence.oralcavity()}
\item \code{OlmMR_2017.marker_presence.skin()}
\item \code{OlmMR_2017.marker_presence.stool()}
\item \code{OlmMR_2017.metaphlan_bugs_list.oralcavity()}
\item \code{OlmMR_2017.metaphlan_bugs_list.skin()}
\item \code{OlmMR_2017.metaphlan_bugs_list.stool()}
\item \code{OlmMR_2017.pathabundance_relab.oralcavity()}
\item \code{OlmMR_2017.pathabundance_relab.skin()}
\item \code{OlmMR_2017.pathabundance_relab.stool()}
\item \code{OlmMR_2017.pathcoverage.oralcavity()}
\item \code{OlmMR_2017.pathcoverage.skin()}
\item \code{OlmMR_2017.pathcoverage.stool()}
\item \code{PasolliE_2018.genefamilies_relab.stool()}
\item \code{PasolliE_2018.marker_abundance.stool()}
\item \code{PasolliE_2018.marker_presence.stool()}
\item \code{PasolliE_2018.metaphlan_bugs_list.stool()}
\item \code{PasolliE_2018.pathabundance_relab.stool()}
\item \code{PasolliE_2018.pathcoverage.stool()}
\item \code{PehrssonE_2016.genefamilies_relab.stool()}
\item \code{PehrssonE_2016.marker_abundance.stool()}
\item \code{PehrssonE_2016.marker_presence.stool()}
\item \code{PehrssonE_2016.metaphlan_bugs_list.stool()}
\item \code{PehrssonE_2016.pathabundance_relab.stool()}
\item \code{PehrssonE_2016.pathcoverage.stool()}
\item \code{QinJ_2012.genefamilies_relab.stool()}
\item \code{QinJ_2012.marker_abundance.stool()}
\item \code{QinJ_2012.marker_presence.stool()}
\item \code{QinJ_2012.metaphlan_bugs_list.stool()}
\item \code{QinJ_2012.pathabundance_relab.stool()}
\item \code{QinJ_2012.pathcoverage.stool()}
\item \code{QinN_2014.genefamilies_relab.stool()}
\item \code{QinN_2014.marker_abundance.stool()}
\item \code{QinN_2014.marker_presence.stool()}
\item \code{QinN_2014.metaphlan_bugs_list.stool()}
\item \code{QinN_2014.pathabundance_relab.stool()}
\item \code{QinN_2014.pathcoverage.stool()}
\item \code{RampelliS_2015.genefamilies_relab.stool()}
\item \code{RampelliS_2015.marker_abundance.stool()}
\item \code{RampelliS_2015.marker_presence.stool()}
\item \code{RampelliS_2015.metaphlan_bugs_list.stool()}
\item \code{RampelliS_2015.pathabundance_relab.stool()}
\item \code{RampelliS_2015.pathcoverage.stool()}
\item \code{RaymondF_2016.genefamilies_relab.stool()}
\item \code{RaymondF_2016.marker_abundance.stool()}
\item \code{RaymondF_2016.marker_presence.stool()}
\item \code{RaymondF_2016.metaphlan_bugs_list.stool()}
\item \code{RaymondF_2016.pathabundance_relab.stool()}
\item \code{RaymondF_2016.pathcoverage.stool()}
\item \code{SchirmerM_2016.genefamilies_relab.stool()}
\item \code{SchirmerM_2016.marker_abundance.stool()}
\item \code{SchirmerM_2016.marker_presence.stool()}
\item \code{SchirmerM_2016.metaphlan_bugs_list.stool()}
\item \code{SchirmerM_2016.pathabundance_relab.stool()}
\item \code{SchirmerM_2016.pathcoverage.stool()}
\item \code{ShiB_2015.genefamilies_relab.oralcavity()}
\item \code{ShiB_2015.marker_abundance.oralcavity()}
\item \code{ShiB_2015.marker_presence.oralcavity()}
\item \code{ShiB_2015.metaphlan_bugs_list.oralcavity()}
\item \code{ShiB_2015.pathabundance_relab.oralcavity()}
\item \code{ShiB_2015.pathcoverage.oralcavity()}
\item \code{SmitsSA_2017.genefamilies_relab.stool()}
\item \code{SmitsSA_2017.marker_abundance.stool()}
\item \code{SmitsSA_2017.marker_presence.stool()}
\item \code{SmitsSA_2017.metaphlan_bugs_list.stool()}
\item \code{SmitsSA_2017.pathabundance_relab.stool()}
\item \code{SmitsSA_2017.pathcoverage.stool()}
\item \code{TettAJ_2016.genefamilies_relab.skin()}
\item \code{TettAJ_2016.marker_abundance.skin()}
\item \code{TettAJ_2016.marker_presence.skin()}
\item \code{TettAJ_2016.metaphlan_bugs_list.skin()}
\item \code{TettAJ_2016.pathabundance_relab.skin()}
\item \code{TettAJ_2016.pathcoverage.skin()}
\item \code{TettAJ_2019_a.genefamilies_relab.stool()}
\item \code{TettAJ_2019_a.marker_abundance.stool()}
\item \code{TettAJ_2019_a.marker_presence.stool()}
\item \code{TettAJ_2019_a.metaphlan_bugs_list.stool()}
\item \code{TettAJ_2019_a.pathabundance_relab.stool()}
\item \code{TettAJ_2019_a.pathcoverage.stool()}
\item \code{TettAJ_2019_b.genefamilies_relab.stool()}
\item \code{TettAJ_2019_b.marker_abundance.stool()}
\item \code{TettAJ_2019_b.marker_presence.stool()}
\item \code{TettAJ_2019_b.metaphlan_bugs_list.stool()}
\item \code{TettAJ_2019_b.pathabundance_relab.stool()}
\item \code{TettAJ_2019_b.pathcoverage.stool()}
\item \code{TettAJ_2019_c.genefamilies_relab.stool()}
\item \code{TettAJ_2019_c.marker_abundance.stool()}
\item \code{TettAJ_2019_c.marker_presence.stool()}
\item \code{TettAJ_2019_c.metaphlan_bugs_list.stool()}
\item \code{TettAJ_2019_c.pathabundance_relab.stool()}
\item \code{TettAJ_2019_c.pathcoverage.stool()}
\item \code{ThomasAM_2018a.genefamilies_relab.stool()}
\item \code{ThomasAM_2018a.marker_abundance.stool()}
\item \code{ThomasAM_2018a.marker_presence.stool()}
\item \code{ThomasAM_2018a.metaphlan_bugs_list.stool()}
\item \code{ThomasAM_2018a.pathabundance_relab.stool()}
\item \code{ThomasAM_2018a.pathcoverage.stool()}
\item \code{ThomasAM_2018b.genefamilies_relab.stool()}
\item \code{ThomasAM_2018b.marker_abundance.stool()}
\item \code{ThomasAM_2018b.marker_presence.stool()}
\item \code{ThomasAM_2018b.metaphlan_bugs_list.stool()}
\item \code{ThomasAM_2018b.pathabundance_relab.stool()}
\item \code{ThomasAM_2018b.pathcoverage.stool()}
\item \code{VatanenT_2016.genefamilies_relab.stool()}
\item \code{VatanenT_2016.marker_abundance.stool()}
\item \code{VatanenT_2016.marker_presence.stool()}
\item \code{VatanenT_2016.metaphlan_bugs_list.stool()}
\item \code{VatanenT_2016.pathabundance_relab.stool()}
\item \code{VatanenT_2016.pathcoverage.stool()}
\item \code{VincentC_2016.genefamilies_relab.stool()}
\item \code{VincentC_2016.marker_abundance.stool()}
\item \code{VincentC_2016.marker_presence.stool()}
\item \code{VincentC_2016.metaphlan_bugs_list.stool()}
\item \code{VincentC_2016.pathabundance_relab.stool()}
\item \code{VincentC_2016.pathcoverage.stool()}
\item \code{VogtmannE_2016.genefamilies_relab.stool()}
\item \code{VogtmannE_2016.marker_abundance.stool()}
\item \code{VogtmannE_2016.marker_presence.stool()}
\item \code{VogtmannE_2016.metaphlan_bugs_list.stool()}
\item \code{VogtmannE_2016.pathabundance_relab.stool()}
\item \code{VogtmannE_2016.pathcoverage.stool()}
\item \code{XieH_2016.genefamilies_relab.stool()}
\item \code{XieH_2016.marker_abundance.stool()}
\item \code{XieH_2016.marker_presence.stool()}
\item \code{XieH_2016.metaphlan_bugs_list.stool()}
\item \code{XieH_2016.pathabundance_relab.stool()}
\item \code{XieH_2016.pathcoverage.stool()}
\item \code{YeZ_2018.genefamilies_relab.stool()}
\item \code{YeZ_2018.marker_abundance.stool()}
\item \code{YeZ_2018.marker_presence.stool()}
\item \code{YeZ_2018.metaphlan_bugs_list.stool()}
\item \code{YeZ_2018.pathabundance_relab.stool()}
\item \code{YeZ_2018.pathcoverage.stool()}
\item \code{YuJ_2015.genefamilies_relab.stool()}
\item \code{YuJ_2015.marker_abundance.stool()}
\item \code{YuJ_2015.marker_presence.stool()}
\item \code{YuJ_2015.metaphlan_bugs_list.stool()}
\item \code{YuJ_2015.pathabundance_relab.stool()}
\item \code{YuJ_2015.pathcoverage.stool()}
\item \code{ZeeviD_2015.genefamilies_relab.stool()}
\item \code{ZeeviD_2015.marker_abundance.stool()}
\item \code{ZeeviD_2015.marker_presence.stool()}
\item \code{ZeeviD_2015.metaphlan_bugs_list.stool()}
\item \code{ZeeviD_2015.pathabundance_relab.stool()}
\item \code{ZeeviD_2015.pathcoverage.stool()}
\item \code{ZellerG_2014.genefamilies_relab.stool()}
\item \code{ZellerG_2014.marker_abundance.stool()}
\item \code{ZellerG_2014.marker_presence.stool()}
\item \code{ZellerG_2014.metaphlan_bugs_list.stool()}
\item \code{ZellerG_2014.pathabundance_relab.stool()}
\item \code{ZellerG_2014.pathcoverage.stool()}
}
}
\keyword{internal}
