% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DavidLA_2015.R
\name{DavidLA_2015}
\alias{DavidLA_2015}
\alias{DavidLA_2015.genefamilies_relab.stool}
\alias{DavidLA_2015.marker_abundance.stool}
\alias{DavidLA_2015.marker_presence.stool}
\alias{DavidLA_2015.metaphlan_bugs_list.stool}
\alias{DavidLA_2015.pathabundance_relab.stool}
\alias{DavidLA_2015.pathcoverage.stool}
\title{Data from the DavidLA_2015 study}
\description{
Data from the DavidLA_2015 study
}
\section{Datasets}{
 

\subsection{DavidLA_2015.genefamilies_relab.stool}{
   An ExpressionSet with 49 samples and 694,925 features specific to the stool body site
}

\subsection{DavidLA_2015.marker_abundance.stool}{
   An ExpressionSet with 49 samples and 57,862 features specific to the stool body site
}

\subsection{DavidLA_2015.marker_presence.stool}{
   An ExpressionSet with 49 samples and 55,487 features specific to the stool body site
}

\subsection{DavidLA_2015.metaphlan_bugs_list.stool}{
   An ExpressionSet with 49 samples and 892 features specific to the stool body site
}

\subsection{DavidLA_2015.pathabundance_relab.stool}{
   An ExpressionSet with 49 samples and 10,437 features specific to the stool body site
}

\subsection{DavidLA_2015.pathcoverage.stool}{
   An ExpressionSet with 49 samples and 10,437 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Gut microbial succession follows acute secretory diarrhea in humans.
}

\subsection{Author}{
   David LA, Weil A, Ryan ET, Calderwood SB, Harris JB, Chowdhury F, Begum Y, Qadri F, LaRocque RC, Turnbaugh PJ
}

\subsection{Lab}{
   [1] Society of Fellows, Harvard University, Cambridge, Massachusetts, USA FAS Center for Systems Biology, Harvard University, Cambridge, Massachusetts, USA., [2] Center for Vaccine Sciences, International Centre for Diarrhoeal Disease Research, Dhaka, Bangladesh., [3] rclarocque@partners.org Peter.Turnbaugh@ucsf.edu., [4] FAS Center for Systems Biology, Harvard University, Cambridge, Massachusetts, USA rclarocque@partners.org Peter.Turnbaugh@ucsf.edu.
}

\subsection{PMID}{
   25991682
}
}

\examples{
`DavidLA_2015.metaphlan_bugs_list.stool`()

}
