\name{defunct}
\alias{defunct}
\alias{AsnicarF_2017}
\alias{AsnicarF_2017.genefamilies_relab.milk}
\alias{AsnicarF_2017.genefamilies_relab.stool}
\alias{AsnicarF_2017.marker_abundance.milk}
\alias{AsnicarF_2017.marker_abundance.stool}
\alias{AsnicarF_2017.marker_presence.milk}
\alias{AsnicarF_2017.marker_presence.stool}
\alias{AsnicarF_2017.metaphlan_bugs_list.milk}
\alias{AsnicarF_2017.metaphlan_bugs_list.stool}
\alias{AsnicarF_2017.pathabundance_relab.milk}
\alias{AsnicarF_2017.pathabundance_relab.stool}
\alias{AsnicarF_2017.pathcoverage.milk}
\alias{AsnicarF_2017.pathcoverage.stool}
\alias{BackhedF_2015}
\alias{BackhedF_2015.genefamilies_relab.stool}
\alias{BackhedF_2015.marker_abundance.stool}
\alias{BackhedF_2015.marker_presence.stool}
\alias{BackhedF_2015.metaphlan_bugs_list.stool}
\alias{BackhedF_2015.pathabundance_relab.stool}
\alias{BackhedF_2015.pathcoverage.stool}
\alias{Bengtsson-PalmeJ_2015}
\alias{Bengtsson-PalmeJ_2015.genefamilies_relab.stool}
\alias{Bengtsson-PalmeJ_2015.marker_abundance.stool}
\alias{Bengtsson-PalmeJ_2015.marker_presence.stool}
\alias{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool}
\alias{Bengtsson-PalmeJ_2015.pathabundance_relab.stool}
\alias{Bengtsson-PalmeJ_2015.pathcoverage.stool}
\alias{BritoIL_2016}
\alias{BritoIL_2016.genefamilies_relab.oralcavity}
\alias{BritoIL_2016.genefamilies_relab.stool}
\alias{BritoIL_2016.marker_abundance.oralcavity}
\alias{BritoIL_2016.marker_abundance.stool}
\alias{BritoIL_2016.marker_presence.oralcavity}
\alias{BritoIL_2016.marker_presence.stool}
\alias{BritoIL_2016.metaphlan_bugs_list.oralcavity}
\alias{BritoIL_2016.metaphlan_bugs_list.stool}
\alias{BritoIL_2016.pathabundance_relab.oralcavity}
\alias{BritoIL_2016.pathabundance_relab.stool}
\alias{BritoIL_2016.pathcoverage.oralcavity}
\alias{BritoIL_2016.pathcoverage.stool}
\alias{Castro-NallarE_2015}
\alias{Castro-NallarE_2015.genefamilies_relab.oralcavity}
\alias{Castro-NallarE_2015.marker_abundance.oralcavity}
\alias{Castro-NallarE_2015.marker_presence.oralcavity}
\alias{Castro-NallarE_2015.metaphlan_bugs_list.oralcavity}
\alias{Castro-NallarE_2015.pathabundance_relab.oralcavity}
\alias{Castro-NallarE_2015.pathcoverage.oralcavity}
\alias{Castro_NallarE_2015.genefamilies_relab.oralcavity}
\alias{Castro_NallarE_2015}
\alias{Castro_NallarE_2015.marker_abundance.oralcavity}
\alias{Castro_NallarE_2015.marker_presence.oralcavity}
\alias{Castro_NallarE_2015.metaphlan_bugs_list.oralcavity}
\alias{Castro_NallarE_2015.pathabundance_relab.oralcavity}
\alias{Castro_NallarE_2015.pathcoverage.oralcavity}
\alias{ChengpingW_2017}
\alias{ChengpingW_2017.genefamilies_relab.stool}
\alias{ChengpingW_2017.marker_abundance.stool}
\alias{ChengpingW_2017.marker_presence.stool}
\alias{ChengpingW_2017.metaphlan_bugs_list.stool}
\alias{ChengpingW_2017.pathabundance_relab.stool}
\alias{ChengpingW_2017.pathcoverage.stool}
\alias{ChngKR_2016}
\alias{ChngKR_2016.genefamilies_relab.skin}
\alias{ChngKR_2016.marker_abundance.skin}
\alias{ChngKR_2016.marker_presence.skin}
\alias{ChngKR_2016.metaphlan_bugs_list.skin}
\alias{ChngKR_2016.pathabundance_relab.skin}
\alias{ChngKR_2016.pathcoverage.skin}
\alias{CosteaPI_2017}
\alias{CosteaPI_2017.genefamilies_relab.stool}
\alias{CosteaPI_2017.marker_abundance.stool}
\alias{CosteaPI_2017.marker_presence.stool}
\alias{CosteaPI_2017.metaphlan_bugs_list.stool}
\alias{CosteaPI_2017.pathabundance_relab.stool}
\alias{CosteaPI_2017.pathcoverage.stool}
\alias{DavidLA_2015}
\alias{DavidLA_2015.genefamilies_relab.stool}
\alias{DavidLA_2015.marker_abundance.stool}
\alias{DavidLA_2015.marker_presence.stool}
\alias{DavidLA_2015.metaphlan_bugs_list.stool}
\alias{DavidLA_2015.pathabundance_relab.stool}
\alias{DavidLA_2015.pathcoverage.stool}
\alias{DhakanDB_2019}
\alias{DhakanDB_2019.genefamilies_relab.stool}
\alias{DhakanDB_2019.marker_abundance.stool}
\alias{DhakanDB_2019.marker_presence.stool}
\alias{DhakanDB_2019.metaphlan_bugs_list.stool}
\alias{DhakanDB_2019.pathabundance_relab.stool}
\alias{DhakanDB_2019.pathcoverage.stool}
\alias{FengQ_2015}
\alias{FengQ_2015.genefamilies_relab.stool}
\alias{FengQ_2015.marker_abundance.stool}
\alias{FengQ_2015.marker_presence.stool}
\alias{FengQ_2015.metaphlan_bugs_list.stool}
\alias{FengQ_2015.pathabundance_relab.stool}
\alias{FengQ_2015.pathcoverage.stool}
\alias{FerrettiP_2018}
\alias{FerrettiP_2018.genefamilies_relab.oralcavity}
\alias{FerrettiP_2018.genefamilies_relab.skin}
\alias{FerrettiP_2018.genefamilies_relab.stool}
\alias{FerrettiP_2018.genefamilies_relab.vagina}
\alias{FerrettiP_2018.marker_abundance.oralcavity}
\alias{FerrettiP_2018.marker_abundance.skin}
\alias{FerrettiP_2018.marker_abundance.stool}
\alias{FerrettiP_2018.marker_abundance.vagina}
\alias{FerrettiP_2018.marker_presence.oralcavity}
\alias{FerrettiP_2018.marker_presence.skin}
\alias{FerrettiP_2018.marker_presence.stool}
\alias{FerrettiP_2018.marker_presence.vagina}
\alias{FerrettiP_2018.metaphlan_bugs_list.oralcavity}
\alias{FerrettiP_2018.metaphlan_bugs_list.skin}
\alias{FerrettiP_2018.metaphlan_bugs_list.stool}
\alias{FerrettiP_2018.metaphlan_bugs_list.vagina}
\alias{FerrettiP_2018.pathabundance_relab.oralcavity}
\alias{FerrettiP_2018.pathabundance_relab.skin}
\alias{FerrettiP_2018.pathabundance_relab.stool}
\alias{FerrettiP_2018.pathabundance_relab.vagina}
\alias{FerrettiP_2018.pathcoverage.oralcavity}
\alias{FerrettiP_2018.pathcoverage.skin}
\alias{FerrettiP_2018.pathcoverage.stool}
\alias{FerrettiP_2018.pathcoverage.vagina}
\alias{GopalakrishnanV_2018}
\alias{GopalakrishnanV_2018.genefamilies_relab.stool}
\alias{GopalakrishnanV_2018.marker_abundance.stool}
\alias{GopalakrishnanV_2018.marker_presence.stool}
\alias{GopalakrishnanV_2018.metaphlan_bugs_list.stool}
\alias{GopalakrishnanV_2018.pathabundance_relab.stool}
\alias{GopalakrishnanV_2018.pathcoverage.stool}
\alias{HMP_2012}
\alias{HMP_2012.genefamilies_relab.nasalcavity}
\alias{HMP_2012.genefamilies_relab.oralcavity}
\alias{HMP_2012.genefamilies_relab.skin}
\alias{HMP_2012.genefamilies_relab.stool}
\alias{HMP_2012.genefamilies_relab.vagina}
\alias{HMP_2012.marker_abundance.nasalcavity}
\alias{HMP_2012.marker_abundance.oralcavity}
\alias{HMP_2012.marker_abundance.skin}
\alias{HMP_2012.marker_abundance.stool}
\alias{HMP_2012.marker_abundance.vagina}
\alias{HMP_2012.marker_presence.nasalcavity}
\alias{HMP_2012.marker_presence.oralcavity}
\alias{HMP_2012.marker_presence.skin}
\alias{HMP_2012.marker_presence.stool}
\alias{HMP_2012.marker_presence.vagina}
\alias{HMP_2012.metaphlan_bugs_list.nasalcavity}
\alias{HMP_2012.metaphlan_bugs_list.oralcavity}
\alias{HMP_2012.metaphlan_bugs_list.skin}
\alias{HMP_2012.metaphlan_bugs_list.stool}
\alias{HMP_2012.metaphlan_bugs_list.vagina}
\alias{HMP_2012.pathabundance_relab.nasalcavity}
\alias{HMP_2012.pathabundance_relab.oralcavity}
\alias{HMP_2012.pathabundance_relab.skin}
\alias{HMP_2012.pathabundance_relab.stool}
\alias{HMP_2012.pathabundance_relab.vagina}
\alias{HMP_2012.pathcoverage.nasalcavity}
\alias{HMP_2012.pathcoverage.oralcavity}
\alias{HMP_2012.pathcoverage.skin}
\alias{HMP_2012.pathcoverage.stool}
\alias{HMP_2012.pathcoverage.vagina}
\alias{HanniganGD_2017}
\alias{HanniganGD_2017.genefamilies_relab.stool}
\alias{HanniganGD_2017.marker_abundance.stool}
\alias{HanniganGD_2017.marker_presence.stool}
\alias{HanniganGD_2017.metaphlan_bugs_list.stool}
\alias{HanniganGD_2017.pathabundance_relab.stool}
\alias{HanniganGD_2017.pathcoverage.stool}
\alias{HansenLBS_2018}
\alias{HansenLBS_2018.genefamilies_relab.stool}
\alias{HansenLBS_2018.marker_abundance.stool}
\alias{HansenLBS_2018.marker_presence.stool}
\alias{HansenLBS_2018.metaphlan_bugs_list.stool}
\alias{HansenLBS_2018.pathabundance_relab.stool}
\alias{HansenLBS_2018.pathcoverage.stool}
\alias{Heitz-BuschartA_2016}
\alias{Heitz-BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz-BuschartA_2016.marker_abundance.stool}
\alias{Heitz-BuschartA_2016.marker_presence.stool}
\alias{Heitz-BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz-BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz-BuschartA_2016.pathcoverage.stool}
\alias{Heitz_BuschartA_2016}
\alias{Heitz_BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz_BuschartA_2016.marker_abundance.stool}
\alias{Heitz_BuschartA_2016.marker_presence.stool}
\alias{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz_BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz_BuschartA_2016.pathcoverage.stool}
\alias{JieZ_2017}
\alias{JieZ_2017.genefamilies_relab.stool}
\alias{JieZ_2017.marker_abundance.stool}
\alias{JieZ_2017.marker_presence.stool}
\alias{JieZ_2017.metaphlan_bugs_list.stool}
\alias{JieZ_2017.pathabundance_relab.stool}
\alias{JieZ_2017.pathcoverage.stool}
\alias{KarlssonFH_2013}
\alias{KarlssonFH_2013.genefamilies_relab.stool}
\alias{KarlssonFH_2013.marker_abundance.stool}
\alias{KarlssonFH_2013.marker_presence.stool}
\alias{KarlssonFH_2013.metaphlan_bugs_list.stool}
\alias{KarlssonFH_2013.pathabundance_relab.stool}
\alias{KarlssonFH_2013.pathcoverage.stool}
\alias{KieserS_2018}
\alias{KieserS_2018.genefamilies_relab.stool}
\alias{KieserS_2018.marker_abundance.stool}
\alias{KieserS_2018.marker_presence.stool}
\alias{KieserS_2018.metaphlan_bugs_list.stool}
\alias{KieserS_2018.pathabundance_relab.stool}
\alias{KieserS_2018.pathcoverage.stool}
\alias{KosticAD_2015}
\alias{KosticAD_2015.genefamilies_relab.stool}
\alias{KosticAD_2015.marker_abundance.stool}
\alias{KosticAD_2015.marker_presence.stool}
\alias{KosticAD_2015.metaphlan_bugs_list.stool}
\alias{KosticAD_2015.pathabundance_relab.stool}
\alias{KosticAD_2015.pathcoverage.stool}
\alias{LeChatelierE_2013}
\alias{LeChatelierE_2013.genefamilies_relab.stool}
\alias{LeChatelierE_2013.marker_abundance.stool}
\alias{LeChatelierE_2013.marker_presence.stool}
\alias{LeChatelierE_2013.metaphlan_bugs_list.stool}
\alias{LeChatelierE_2013.pathabundance_relab.stool}
\alias{LeChatelierE_2013.pathcoverage.stool}
\alias{LiJ_2014}
\alias{LiJ_2014.genefamilies_relab.stool}
\alias{LiJ_2014.marker_abundance.stool}
\alias{LiJ_2014.marker_presence.stool}
\alias{LiJ_2014.metaphlan_bugs_list.stool}
\alias{LiJ_2014.pathabundance_relab.stool}
\alias{LiJ_2014.pathcoverage.stool}
\alias{LiJ_2017}
\alias{LiJ_2017.genefamilies_relab.stool}
\alias{LiJ_2017.marker_abundance.stool}
\alias{LiJ_2017.marker_presence.stool}
\alias{LiJ_2017.metaphlan_bugs_list.stool}
\alias{LiJ_2017.pathabundance_relab.stool}
\alias{LiJ_2017.pathcoverage.stool}
\alias{LiSS_2016}
\alias{LiSS_2016.genefamilies_relab.stool}
\alias{LiSS_2016.marker_abundance.stool}
\alias{LiSS_2016.marker_presence.stool}
\alias{LiSS_2016.metaphlan_bugs_list.stool}
\alias{LiSS_2016.pathabundance_relab.stool}
\alias{LiSS_2016.pathcoverage.stool}
\alias{LiuW_2016}
\alias{LiuW_2016.genefamilies_relab.stool}
\alias{LiuW_2016.marker_abundance.stool}
\alias{LiuW_2016.marker_presence.stool}
\alias{LiuW_2016.metaphlan_bugs_list.stool}
\alias{LiuW_2016.pathabundance_relab.stool}
\alias{LiuW_2016.pathcoverage.stool}
\alias{LomanNJ_2013}
\alias{LomanNJ_2013.genefamilies_relab.stool}
\alias{LomanNJ_2013.marker_abundance.stool}
\alias{LomanNJ_2013.marker_presence.stool}
\alias{LomanNJ_2013.metaphlan_bugs_list.stool}
\alias{LomanNJ_2013.pathabundance_relab.stool}
\alias{LomanNJ_2013.pathcoverage.stool}
\alias{LoombaR_2017}
\alias{LoombaR_2017.genefamilies_relab.stool}
\alias{LoombaR_2017.marker_abundance.stool}
\alias{LoombaR_2017.marker_presence.stool}
\alias{LoombaR_2017.metaphlan_bugs_list.stool}
\alias{LoombaR_2017.pathabundance_relab.stool}
\alias{LoombaR_2017.pathcoverage.stool}
\alias{LouisS_2016}
\alias{LouisS_2016.genefamilies_relab.stool}
\alias{LouisS_2016.marker_abundance.stool}
\alias{LouisS_2016.marker_presence.stool}
\alias{LouisS_2016.metaphlan_bugs_list.stool}
\alias{LouisS_2016.pathabundance_relab.stool}
\alias{LouisS_2016.pathcoverage.stool}
\alias{MatsonV_2018}
\alias{MatsonV_2018.genefamilies_relab.stool}
\alias{MatsonV_2018.marker_abundance.stool}
\alias{MatsonV_2018.marker_presence.stool}
\alias{MatsonV_2018.metaphlan_bugs_list.stool}
\alias{MatsonV_2018.pathabundance_relab.stool}
\alias{MatsonV_2018.pathcoverage.stool}
\alias{NielsenHB_2014}
\alias{NielsenHB_2014.genefamilies_relab.stool}
\alias{NielsenHB_2014.marker_abundance.stool}
\alias{NielsenHB_2014.marker_presence.stool}
\alias{NielsenHB_2014.metaphlan_bugs_list.stool}
\alias{NielsenHB_2014.pathabundance_relab.stool}
\alias{NielsenHB_2014.pathcoverage.stool}
\alias{Obregon-TitoAJ_2015}
\alias{Obregon-TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon-TitoAJ_2015.marker_abundance.stool}
\alias{Obregon-TitoAJ_2015.marker_presence.stool}
\alias{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon-TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon-TitoAJ_2015.pathcoverage.stool}
\alias{Obregon_TitoAJ_2015}
\alias{Obregon_TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon_TitoAJ_2015.marker_abundance.stool}
\alias{Obregon_TitoAJ_2015.marker_presence.stool}
\alias{Obregon_TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon_TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon_TitoAJ_2015.pathcoverage.stool}
\alias{OhJ_2014}
\alias{OhJ_2014.genefamilies_relab.skin}
\alias{OhJ_2014.marker_abundance.skin}
\alias{OhJ_2014.marker_presence.skin}
\alias{OhJ_2014.metaphlan_bugs_list.skin}
\alias{OhJ_2014.pathabundance_relab.skin}
\alias{OhJ_2014.pathcoverage.skin}
\alias{OlmMR_2017}
\alias{OlmMR_2017.genefamilies_relab.oralcavity}
\alias{OlmMR_2017.genefamilies_relab.skin}
\alias{OlmMR_2017.genefamilies_relab.stool}
\alias{OlmMR_2017.marker_abundance.oralcavity}
\alias{OlmMR_2017.marker_abundance.skin}
\alias{OlmMR_2017.marker_abundance.stool}
\alias{OlmMR_2017.marker_presence.oralcavity}
\alias{OlmMR_2017.marker_presence.skin}
\alias{OlmMR_2017.marker_presence.stool}
\alias{OlmMR_2017.metaphlan_bugs_list.oralcavity}
\alias{OlmMR_2017.metaphlan_bugs_list.skin}
\alias{OlmMR_2017.metaphlan_bugs_list.stool}
\alias{OlmMR_2017.pathabundance_relab.oralcavity}
\alias{OlmMR_2017.pathabundance_relab.skin}
\alias{OlmMR_2017.pathabundance_relab.stool}
\alias{OlmMR_2017.pathcoverage.oralcavity}
\alias{OlmMR_2017.pathcoverage.skin}
\alias{OlmMR_2017.pathcoverage.stool}
\alias{PasolliE_2018}
\alias{PasolliE_2018.genefamilies_relab.stool}
\alias{PasolliE_2018.marker_abundance.stool}
\alias{PasolliE_2018.marker_presence.stool}
\alias{PasolliE_2018.metaphlan_bugs_list.stool}
\alias{PasolliE_2018.pathabundance_relab.stool}
\alias{PasolliE_2018.pathcoverage.stool}
\alias{PehrssonE_2016}
\alias{PehrssonE_2016.genefamilies_relab.stool}
\alias{PehrssonE_2016.marker_abundance.stool}
\alias{PehrssonE_2016.marker_presence.stool}
\alias{PehrssonE_2016.metaphlan_bugs_list.stool}
\alias{PehrssonE_2016.pathabundance_relab.stool}
\alias{PehrssonE_2016.pathcoverage.stool}
\alias{QinJ_2012}
\alias{QinJ_2012.genefamilies_relab.stool}
\alias{QinJ_2012.marker_abundance.stool}
\alias{QinJ_2012.marker_presence.stool}
\alias{QinJ_2012.metaphlan_bugs_list.stool}
\alias{QinJ_2012.pathabundance_relab.stool}
\alias{QinJ_2012.pathcoverage.stool}
\alias{QinN_2014}
\alias{QinN_2014.genefamilies_relab.stool}
\alias{QinN_2014.marker_abundance.stool}
\alias{QinN_2014.marker_presence.stool}
\alias{QinN_2014.metaphlan_bugs_list.stool}
\alias{QinN_2014.pathabundance_relab.stool}
\alias{QinN_2014.pathcoverage.stool}
\alias{RampelliS_2015}
\alias{RampelliS_2015.genefamilies_relab.stool}
\alias{RampelliS_2015.marker_abundance.stool}
\alias{RampelliS_2015.marker_presence.stool}
\alias{RampelliS_2015.metaphlan_bugs_list.stool}
\alias{RampelliS_2015.pathabundance_relab.stool}
\alias{RampelliS_2015.pathcoverage.stool}
\alias{RaymondF_2016}
\alias{RaymondF_2016.genefamilies_relab.stool}
\alias{RaymondF_2016.marker_abundance.stool}
\alias{RaymondF_2016.marker_presence.stool}
\alias{RaymondF_2016.metaphlan_bugs_list.stool}
\alias{RaymondF_2016.pathabundance_relab.stool}
\alias{RaymondF_2016.pathcoverage.stool}
\alias{SchirmerM_2016}
\alias{SchirmerM_2016.genefamilies_relab.stool}
\alias{SchirmerM_2016.marker_abundance.stool}
\alias{SchirmerM_2016.marker_presence.stool}
\alias{SchirmerM_2016.metaphlan_bugs_list.stool}
\alias{SchirmerM_2016.pathabundance_relab.stool}
\alias{SchirmerM_2016.pathcoverage.stool}
\alias{ShiB_2015}
\alias{ShiB_2015.genefamilies_relab.oralcavity}
\alias{ShiB_2015.marker_abundance.oralcavity}
\alias{ShiB_2015.marker_presence.oralcavity}
\alias{ShiB_2015.metaphlan_bugs_list.oralcavity}
\alias{ShiB_2015.pathabundance_relab.oralcavity}
\alias{ShiB_2015.pathcoverage.oralcavity}
\alias{SmitsSA_2017}
\alias{SmitsSA_2017.genefamilies_relab.stool}
\alias{SmitsSA_2017.marker_abundance.stool}
\alias{SmitsSA_2017.marker_presence.stool}
\alias{SmitsSA_2017.metaphlan_bugs_list.stool}
\alias{SmitsSA_2017.pathabundance_relab.stool}
\alias{SmitsSA_2017.pathcoverage.stool}
\alias{TettAJ_2016}
\alias{TettAJ_2016.genefamilies_relab.skin}
\alias{TettAJ_2016.marker_abundance.skin}
\alias{TettAJ_2016.marker_presence.skin}
\alias{TettAJ_2016.metaphlan_bugs_list.skin}
\alias{TettAJ_2016.pathabundance_relab.skin}
\alias{TettAJ_2016.pathcoverage.skin}
\alias{TettAJ_2019_a}
\alias{TettAJ_2019_a.genefamilies_relab.stool}
\alias{TettAJ_2019_a.marker_abundance.stool}
\alias{TettAJ_2019_a.marker_presence.stool}
\alias{TettAJ_2019_a.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_a.pathabundance_relab.stool}
\alias{TettAJ_2019_a.pathcoverage.stool}
\alias{TettAJ_2019_b}
\alias{TettAJ_2019_b.genefamilies_relab.stool}
\alias{TettAJ_2019_b.marker_abundance.stool}
\alias{TettAJ_2019_b.marker_presence.stool}
\alias{TettAJ_2019_b.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_b.pathabundance_relab.stool}
\alias{TettAJ_2019_b.pathcoverage.stool}
\alias{TettAJ_2019_c}
\alias{TettAJ_2019_c.genefamilies_relab.stool}
\alias{TettAJ_2019_c.marker_abundance.stool}
\alias{TettAJ_2019_c.marker_presence.stool}
\alias{TettAJ_2019_c.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_c.pathabundance_relab.stool}
\alias{TettAJ_2019_c.pathcoverage.stool}
\alias{ThomasAM_2018a}
\alias{ThomasAM_2018a.genefamilies_relab.stool}
\alias{ThomasAM_2018a.marker_abundance.stool}
\alias{ThomasAM_2018a.marker_presence.stool}
\alias{ThomasAM_2018a.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018a.pathabundance_relab.stool}
\alias{ThomasAM_2018a.pathcoverage.stool}
\alias{ThomasAM_2018b.genefamilies_relab.stool}
\alias{ThomasAM_2018b}
\alias{ThomasAM_2018b.marker_abundance.stool}
\alias{ThomasAM_2018b.marker_presence.stool}
\alias{ThomasAM_2018b.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018b.pathabundance_relab.stool}
\alias{ThomasAM_2018b.pathcoverage.stool}
\alias{VatanenT_2016}
\alias{VatanenT_2016.genefamilies_relab.stool}
\alias{VatanenT_2016.marker_abundance.stool}
\alias{VatanenT_2016.marker_presence.stool}
\alias{VatanenT_2016.metaphlan_bugs_list.stool}
\alias{VatanenT_2016.pathabundance_relab.stool}
\alias{VatanenT_2016.pathcoverage.stool}
\alias{VincentC_2016}
\alias{VincentC_2016.genefamilies_relab.stool}
\alias{VincentC_2016.marker_abundance.stool}
\alias{VincentC_2016.marker_presence.stool}
\alias{VincentC_2016.metaphlan_bugs_list.stool}
\alias{VincentC_2016.pathabundance_relab.stool}
\alias{VincentC_2016.pathcoverage.stool}
\alias{VogtmannE_2016}
\alias{VogtmannE_2016.genefamilies_relab.stool}
\alias{VogtmannE_2016.marker_abundance.stool}
\alias{VogtmannE_2016.marker_presence.stool}
\alias{VogtmannE_2016.metaphlan_bugs_list.stool}
\alias{VogtmannE_2016.pathabundance_relab.stool}
\alias{VogtmannE_2016.pathcoverage.stool}
\alias{WenC_2017.genefamilies_relab.stool}
\alias{WenC_2017.marker_abundance.stool}
\alias{WenC_2017.marker_presence.stool}
\alias{WenC_2017.metaphlan_bugs_list.stool}
\alias{WenC_2017.pathabundance_relab.stool}
\alias{WenC_2017.pathcoverage.stool}
\alias{XieH_2016}
\alias{XieH_2016.genefamilies_relab.stool}
\alias{XieH_2016.marker_abundance.stool}
\alias{XieH_2016.marker_presence.stool}
\alias{XieH_2016.metaphlan_bugs_list.stool}
\alias{XieH_2016.pathabundance_relab.stool}
\alias{XieH_2016.pathcoverage.stool}
\alias{YeZ_2018}
\alias{YeZ_2018.genefamilies_relab.stool}
\alias{YeZ_2018.marker_abundance.stool}
\alias{YeZ_2018.marker_presence.stool}
\alias{YeZ_2018.metaphlan_bugs_list.stool}
\alias{YeZ_2018.pathabundance_relab.stool}
\alias{YeZ_2018.pathcoverage.stool}
\alias{YuJ_2015}
\alias{YuJ_2015.genefamilies_relab.stool}
\alias{YuJ_2015.marker_abundance.stool}
\alias{YuJ_2015.marker_presence.stool}
\alias{YuJ_2015.metaphlan_bugs_list.stool}
\alias{YuJ_2015.pathabundance_relab.stool}
\alias{YuJ_2015.pathcoverage.stool}
\alias{ZeeviD_2015}
\alias{ZeeviD_2015.genefamilies_relab.stool}
\alias{ZeeviD_2015.marker_abundance.stool}
\alias{ZeeviD_2015.marker_presence.stool}
\alias{ZeeviD_2015.metaphlan_bugs_list.stool}
\alias{ZeeviD_2015.pathabundance_relab.stool}
\alias{ZeeviD_2015.pathcoverage.stool}
\alias{ZellerG_2014}
\alias{ZellerG_2014.genefamilies_relab.stool}
\alias{ZellerG_2014.marker_abundance.stool}
\alias{ZellerG_2014.marker_presence.stool}
\alias{ZellerG_2014.metaphlan_bugs_list.stool}
\alias{ZellerG_2014.pathabundance_relab.stool}
\alias{ZellerG_2014.pathcoverage.stool}
\title{Defunct Functions}
\description{
The following functions are defunct and are no longer part of
the curatedMetagenomicData package. Use \code{curatedMetagenomicData()}.
}
\section{Defunct Functions}{
Use \code{curatedMetagenomicData()} instead of the following defunct functions:
\itemize{
\item{AsnicarF_2017}
\item{AsnicarF_2017.genefamilies_relab.milk}
\item{AsnicarF_2017.genefamilies_relab.stool}
\item{AsnicarF_2017.marker_abundance.milk}
\item{AsnicarF_2017.marker_abundance.stool}
\item{AsnicarF_2017.marker_presence.milk}
\item{AsnicarF_2017.marker_presence.stool}
\item{AsnicarF_2017.metaphlan_bugs_list.milk}
\item{AsnicarF_2017.metaphlan_bugs_list.stool}
\item{AsnicarF_2017.pathabundance_relab.milk}
\item{AsnicarF_2017.pathabundance_relab.stool}
\item{AsnicarF_2017.pathcoverage.milk}
\item{AsnicarF_2017.pathcoverage.stool}
\item{BackhedF_2015}
\item{BackhedF_2015.genefamilies_relab.stool}
\item{BackhedF_2015.marker_abundance.stool}
\item{BackhedF_2015.marker_presence.stool}
\item{BackhedF_2015.metaphlan_bugs_list.stool}
\item{BackhedF_2015.pathabundance_relab.stool}
\item{BackhedF_2015.pathcoverage.stool}
\item{Bengtsson-PalmeJ_2015}
\item{Bengtsson-PalmeJ_2015.genefamilies_relab.stool}
\item{Bengtsson-PalmeJ_2015.marker_abundance.stool}
\item{Bengtsson-PalmeJ_2015.marker_presence.stool}
\item{Bengtsson-PalmeJ_2015.metaphlan_bugs_list.stool}
\item{Bengtsson-PalmeJ_2015.pathabundance_relab.stool}
\item{Bengtsson-PalmeJ_2015.pathcoverage.stool}
\item{BritoIL_2016}
\item{BritoIL_2016.genefamilies_relab.oralcavity}
\item{BritoIL_2016.genefamilies_relab.stool}
\item{BritoIL_2016.marker_abundance.oralcavity}
\item{BritoIL_2016.marker_abundance.stool}
\item{BritoIL_2016.marker_presence.oralcavity}
\item{BritoIL_2016.marker_presence.stool}
\item{BritoIL_2016.metaphlan_bugs_list.oralcavity}
\item{BritoIL_2016.metaphlan_bugs_list.stool}
\item{BritoIL_2016.pathabundance_relab.oralcavity}
\item{BritoIL_2016.pathabundance_relab.stool}
\item{BritoIL_2016.pathcoverage.oralcavity}
\item{BritoIL_2016.pathcoverage.stool}
\item{Castro-NallarE_2015}
\item{Castro-NallarE_2015.genefamilies_relab.oralcavity}
\item{Castro-NallarE_2015.marker_abundance.oralcavity}
\item{Castro-NallarE_2015.marker_presence.oralcavity}
\item{Castro-NallarE_2015.metaphlan_bugs_list.oralcavity}
\item{Castro-NallarE_2015.pathabundance_relab.oralcavity}
\item{Castro-NallarE_2015.pathcoverage.oralcavity}
\item{Castro_NallarE_2015.genefamilies_relab.oralcavity}
\item{Castro_NallarE_2015}
\item{Castro_NallarE_2015.marker_abundance.oralcavity}
\item{Castro_NallarE_2015.marker_presence.oralcavity}
\item{Castro_NallarE_2015.metaphlan_bugs_list.oralcavity}
\item{Castro_NallarE_2015.pathabundance_relab.oralcavity}
\item{Castro_NallarE_2015.pathcoverage.oralcavity}
\item{ChengpingW_2017}
\item{ChengpingW_2017.genefamilies_relab.stool}
\item{ChengpingW_2017.marker_abundance.stool}
\item{ChengpingW_2017.marker_presence.stool}
\item{ChengpingW_2017.metaphlan_bugs_list.stool}
\item{ChengpingW_2017.pathabundance_relab.stool}
\item{ChengpingW_2017.pathcoverage.stool}
\item{ChngKR_2016}
\item{ChngKR_2016.genefamilies_relab.skin}
\item{ChngKR_2016.marker_abundance.skin}
\item{ChngKR_2016.marker_presence.skin}
\item{ChngKR_2016.metaphlan_bugs_list.skin}
\item{ChngKR_2016.pathabundance_relab.skin}
\item{ChngKR_2016.pathcoverage.skin}
\item{CosteaPI_2017}
\item{CosteaPI_2017.genefamilies_relab.stool}
\item{CosteaPI_2017.marker_abundance.stool}
\item{CosteaPI_2017.marker_presence.stool}
\item{CosteaPI_2017.metaphlan_bugs_list.stool}
\item{CosteaPI_2017.pathabundance_relab.stool}
\item{CosteaPI_2017.pathcoverage.stool}
\item{DavidLA_2015}
\item{DavidLA_2015.genefamilies_relab.stool}
\item{DavidLA_2015.marker_abundance.stool}
\item{DavidLA_2015.marker_presence.stool}
\item{DavidLA_2015.metaphlan_bugs_list.stool}
\item{DavidLA_2015.pathabundance_relab.stool}
\item{DavidLA_2015.pathcoverage.stool}
\item{DhakanDB_2019}
\item{DhakanDB_2019.genefamilies_relab.stool}
\item{DhakanDB_2019.marker_abundance.stool}
\item{DhakanDB_2019.marker_presence.stool}
\item{DhakanDB_2019.metaphlan_bugs_list.stool}
\item{DhakanDB_2019.pathabundance_relab.stool}
\item{DhakanDB_2019.pathcoverage.stool}
\item{FengQ_2015}
\item{FengQ_2015.genefamilies_relab.stool}
\item{FengQ_2015.marker_abundance.stool}
\item{FengQ_2015.marker_presence.stool}
\item{FengQ_2015.metaphlan_bugs_list.stool}
\item{FengQ_2015.pathabundance_relab.stool}
\item{FengQ_2015.pathcoverage.stool}
\item{FerrettiP_2018}
\item{FerrettiP_2018.genefamilies_relab.oralcavity}
\item{FerrettiP_2018.genefamilies_relab.skin}
\item{FerrettiP_2018.genefamilies_relab.stool}
\item{FerrettiP_2018.genefamilies_relab.vagina}
\item{FerrettiP_2018.marker_abundance.oralcavity}
\item{FerrettiP_2018.marker_abundance.skin}
\item{FerrettiP_2018.marker_abundance.stool}
\item{FerrettiP_2018.marker_abundance.vagina}
\item{FerrettiP_2018.marker_presence.oralcavity}
\item{FerrettiP_2018.marker_presence.skin}
\item{FerrettiP_2018.marker_presence.stool}
\item{FerrettiP_2018.marker_presence.vagina}
\item{FerrettiP_2018.metaphlan_bugs_list.oralcavity}
\item{FerrettiP_2018.metaphlan_bugs_list.skin}
\item{FerrettiP_2018.metaphlan_bugs_list.stool}
\item{FerrettiP_2018.metaphlan_bugs_list.vagina}
\item{FerrettiP_2018.pathabundance_relab.oralcavity}
\item{FerrettiP_2018.pathabundance_relab.skin}
\item{FerrettiP_2018.pathabundance_relab.stool}
\item{FerrettiP_2018.pathabundance_relab.vagina}
\item{FerrettiP_2018.pathcoverage.oralcavity}
\item{FerrettiP_2018.pathcoverage.skin}
\item{FerrettiP_2018.pathcoverage.stool}
\item{FerrettiP_2018.pathcoverage.vagina}
\item{GopalakrishnanV_2018}
\item{GopalakrishnanV_2018.genefamilies_relab.stool}
\item{GopalakrishnanV_2018.marker_abundance.stool}
\item{GopalakrishnanV_2018.marker_presence.stool}
\item{GopalakrishnanV_2018.metaphlan_bugs_list.stool}
\item{GopalakrishnanV_2018.pathabundance_relab.stool}
\item{GopalakrishnanV_2018.pathcoverage.stool}
\item{HMP_2012}
\item{HMP_2012.genefamilies_relab.nasalcavity}
\item{HMP_2012.genefamilies_relab.oralcavity}
\item{HMP_2012.genefamilies_relab.skin}
\item{HMP_2012.genefamilies_relab.stool}
\item{HMP_2012.genefamilies_relab.vagina}
\item{HMP_2012.marker_abundance.nasalcavity}
\item{HMP_2012.marker_abundance.oralcavity}
\item{HMP_2012.marker_abundance.skin}
\item{HMP_2012.marker_abundance.stool}
\item{HMP_2012.marker_abundance.vagina}
\item{HMP_2012.marker_presence.nasalcavity}
\item{HMP_2012.marker_presence.oralcavity}
\item{HMP_2012.marker_presence.skin}
\item{HMP_2012.marker_presence.stool}
\item{HMP_2012.marker_presence.vagina}
\item{HMP_2012.metaphlan_bugs_list.nasalcavity}
\item{HMP_2012.metaphlan_bugs_list.oralcavity}
\item{HMP_2012.metaphlan_bugs_list.skin}
\item{HMP_2012.metaphlan_bugs_list.stool}
\item{HMP_2012.metaphlan_bugs_list.vagina}
\item{HMP_2012.pathabundance_relab.nasalcavity}
\item{HMP_2012.pathabundance_relab.oralcavity}
\item{HMP_2012.pathabundance_relab.skin}
\item{HMP_2012.pathabundance_relab.stool}
\item{HMP_2012.pathabundance_relab.vagina}
\item{HMP_2012.pathcoverage.nasalcavity}
\item{HMP_2012.pathcoverage.oralcavity}
\item{HMP_2012.pathcoverage.skin}
\item{HMP_2012.pathcoverage.stool}
\item{HMP_2012.pathcoverage.vagina}
\item{HanniganGD_2017}
\item{HanniganGD_2017.genefamilies_relab.stool}
\item{HanniganGD_2017.marker_abundance.stool}
\item{HanniganGD_2017.marker_presence.stool}
\item{HanniganGD_2017.metaphlan_bugs_list.stool}
\item{HanniganGD_2017.pathabundance_relab.stool}
\item{HanniganGD_2017.pathcoverage.stool}
\item{HansenLBS_2018}
\item{HansenLBS_2018.genefamilies_relab.stool}
\item{HansenLBS_2018.marker_abundance.stool}
\item{HansenLBS_2018.marker_presence.stool}
\item{HansenLBS_2018.metaphlan_bugs_list.stool}
\item{HansenLBS_2018.pathabundance_relab.stool}
\item{HansenLBS_2018.pathcoverage.stool}
\item{Heitz-BuschartA_2016}
\item{Heitz-BuschartA_2016.genefamilies_relab.stool}
\item{Heitz-BuschartA_2016.marker_abundance.stool}
\item{Heitz-BuschartA_2016.marker_presence.stool}
\item{Heitz-BuschartA_2016.metaphlan_bugs_list.stool}
\item{Heitz-BuschartA_2016.pathabundance_relab.stool}
\item{Heitz-BuschartA_2016.pathcoverage.stool}
\item{Heitz_BuschartA_2016}
\item{Heitz_BuschartA_2016.genefamilies_relab.stool}
\item{Heitz_BuschartA_2016.marker_abundance.stool}
\item{Heitz_BuschartA_2016.marker_presence.stool}
\item{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}
\item{Heitz_BuschartA_2016.pathabundance_relab.stool}
\item{Heitz_BuschartA_2016.pathcoverage.stool}
\item{JieZ_2017}
\item{JieZ_2017.genefamilies_relab.stool}
\item{JieZ_2017.marker_abundance.stool}
\item{JieZ_2017.marker_presence.stool}
\item{JieZ_2017.metaphlan_bugs_list.stool}
\item{JieZ_2017.pathabundance_relab.stool}
\item{JieZ_2017.pathcoverage.stool}
\item{KarlssonFH_2013}
\item{KarlssonFH_2013.genefamilies_relab.stool}
\item{KarlssonFH_2013.marker_abundance.stool}
\item{KarlssonFH_2013.marker_presence.stool}
\item{KarlssonFH_2013.metaphlan_bugs_list.stool}
\item{KarlssonFH_2013.pathabundance_relab.stool}
\item{KarlssonFH_2013.pathcoverage.stool}
\item{KieserS_2018}
\item{KieserS_2018.genefamilies_relab.stool}
\item{KieserS_2018.marker_abundance.stool}
\item{KieserS_2018.marker_presence.stool}
\item{KieserS_2018.metaphlan_bugs_list.stool}
\item{KieserS_2018.pathabundance_relab.stool}
\item{KieserS_2018.pathcoverage.stool}
\item{KosticAD_2015}
\item{KosticAD_2015.genefamilies_relab.stool}
\item{KosticAD_2015.marker_abundance.stool}
\item{KosticAD_2015.marker_presence.stool}
\item{KosticAD_2015.metaphlan_bugs_list.stool}
\item{KosticAD_2015.pathabundance_relab.stool}
\item{KosticAD_2015.pathcoverage.stool}
\item{LeChatelierE_2013}
\item{LeChatelierE_2013.genefamilies_relab.stool}
\item{LeChatelierE_2013.marker_abundance.stool}
\item{LeChatelierE_2013.marker_presence.stool}
\item{LeChatelierE_2013.metaphlan_bugs_list.stool}
\item{LeChatelierE_2013.pathabundance_relab.stool}
\item{LeChatelierE_2013.pathcoverage.stool}
\item{LiJ_2014}
\item{LiJ_2014.genefamilies_relab.stool}
\item{LiJ_2014.marker_abundance.stool}
\item{LiJ_2014.marker_presence.stool}
\item{LiJ_2014.metaphlan_bugs_list.stool}
\item{LiJ_2014.pathabundance_relab.stool}
\item{LiJ_2014.pathcoverage.stool}
\item{LiJ_2017}
\item{LiJ_2017.genefamilies_relab.stool}
\item{LiJ_2017.marker_abundance.stool}
\item{LiJ_2017.marker_presence.stool}
\item{LiJ_2017.metaphlan_bugs_list.stool}
\item{LiJ_2017.pathabundance_relab.stool}
\item{LiJ_2017.pathcoverage.stool}
\item{LiSS_2016}
\item{LiSS_2016.genefamilies_relab.stool}
\item{LiSS_2016.marker_abundance.stool}
\item{LiSS_2016.marker_presence.stool}
\item{LiSS_2016.metaphlan_bugs_list.stool}
\item{LiSS_2016.pathabundance_relab.stool}
\item{LiSS_2016.pathcoverage.stool}
\item{LiuW_2016}
\item{LiuW_2016.genefamilies_relab.stool}
\item{LiuW_2016.marker_abundance.stool}
\item{LiuW_2016.marker_presence.stool}
\item{LiuW_2016.metaphlan_bugs_list.stool}
\item{LiuW_2016.pathabundance_relab.stool}
\item{LiuW_2016.pathcoverage.stool}
\item{LomanNJ_2013}
\item{LomanNJ_2013.genefamilies_relab.stool}
\item{LomanNJ_2013.marker_abundance.stool}
\item{LomanNJ_2013.marker_presence.stool}
\item{LomanNJ_2013.metaphlan_bugs_list.stool}
\item{LomanNJ_2013.pathabundance_relab.stool}
\item{LomanNJ_2013.pathcoverage.stool}
\item{LoombaR_2017}
\item{LoombaR_2017.genefamilies_relab.stool}
\item{LoombaR_2017.marker_abundance.stool}
\item{LoombaR_2017.marker_presence.stool}
\item{LoombaR_2017.metaphlan_bugs_list.stool}
\item{LoombaR_2017.pathabundance_relab.stool}
\item{LoombaR_2017.pathcoverage.stool}
\item{LouisS_2016}
\item{LouisS_2016.genefamilies_relab.stool}
\item{LouisS_2016.marker_abundance.stool}
\item{LouisS_2016.marker_presence.stool}
\item{LouisS_2016.metaphlan_bugs_list.stool}
\item{LouisS_2016.pathabundance_relab.stool}
\item{LouisS_2016.pathcoverage.stool}
\item{MatsonV_2018}
\item{MatsonV_2018.genefamilies_relab.stool}
\item{MatsonV_2018.marker_abundance.stool}
\item{MatsonV_2018.marker_presence.stool}
\item{MatsonV_2018.metaphlan_bugs_list.stool}
\item{MatsonV_2018.pathabundance_relab.stool}
\item{MatsonV_2018.pathcoverage.stool}
\item{NielsenHB_2014}
\item{NielsenHB_2014.genefamilies_relab.stool}
\item{NielsenHB_2014.marker_abundance.stool}
\item{NielsenHB_2014.marker_presence.stool}
\item{NielsenHB_2014.metaphlan_bugs_list.stool}
\item{NielsenHB_2014.pathabundance_relab.stool}
\item{NielsenHB_2014.pathcoverage.stool}
\item{Obregon-TitoAJ_2015}
\item{Obregon-TitoAJ_2015.genefamilies_relab.stool}
\item{Obregon-TitoAJ_2015.marker_abundance.stool}
\item{Obregon-TitoAJ_2015.marker_presence.stool}
\item{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool}
\item{Obregon-TitoAJ_2015.pathabundance_relab.stool}
\item{Obregon-TitoAJ_2015.pathcoverage.stool}
\item{Obregon_TitoAJ_2015}
\item{Obregon_TitoAJ_2015.genefamilies_relab.stool}
\item{Obregon_TitoAJ_2015.marker_abundance.stool}
\item{Obregon_TitoAJ_2015.marker_presence.stool}
\item{Obregon_TitoAJ_2015.metaphlan_bugs_list.stool}
\item{Obregon_TitoAJ_2015.pathabundance_relab.stool}
\item{Obregon_TitoAJ_2015.pathcoverage.stool}
\item{OhJ_2014}
\item{OhJ_2014.genefamilies_relab.skin}
\item{OhJ_2014.marker_abundance.skin}
\item{OhJ_2014.marker_presence.skin}
\item{OhJ_2014.metaphlan_bugs_list.skin}
\item{OhJ_2014.pathabundance_relab.skin}
\item{OhJ_2014.pathcoverage.skin}
\item{OlmMR_2017}
\item{OlmMR_2017.genefamilies_relab.oralcavity}
\item{OlmMR_2017.genefamilies_relab.skin}
\item{OlmMR_2017.genefamilies_relab.stool}
\item{OlmMR_2017.marker_abundance.oralcavity}
\item{OlmMR_2017.marker_abundance.skin}
\item{OlmMR_2017.marker_abundance.stool}
\item{OlmMR_2017.marker_presence.oralcavity}
\item{OlmMR_2017.marker_presence.skin}
\item{OlmMR_2017.marker_presence.stool}
\item{OlmMR_2017.metaphlan_bugs_list.oralcavity}
\item{OlmMR_2017.metaphlan_bugs_list.skin}
\item{OlmMR_2017.metaphlan_bugs_list.stool}
\item{OlmMR_2017.pathabundance_relab.oralcavity}
\item{OlmMR_2017.pathabundance_relab.skin}
\item{OlmMR_2017.pathabundance_relab.stool}
\item{OlmMR_2017.pathcoverage.oralcavity}
\item{OlmMR_2017.pathcoverage.skin}
\item{OlmMR_2017.pathcoverage.stool}
\item{PasolliE_2018}
\item{PasolliE_2018.genefamilies_relab.stool}
\item{PasolliE_2018.marker_abundance.stool}
\item{PasolliE_2018.marker_presence.stool}
\item{PasolliE_2018.metaphlan_bugs_list.stool}
\item{PasolliE_2018.pathabundance_relab.stool}
\item{PasolliE_2018.pathcoverage.stool}
\item{PehrssonE_2016}
\item{PehrssonE_2016.genefamilies_relab.stool}
\item{PehrssonE_2016.marker_abundance.stool}
\item{PehrssonE_2016.marker_presence.stool}
\item{PehrssonE_2016.metaphlan_bugs_list.stool}
\item{PehrssonE_2016.pathabundance_relab.stool}
\item{PehrssonE_2016.pathcoverage.stool}
\item{QinJ_2012}
\item{QinJ_2012.genefamilies_relab.stool}
\item{QinJ_2012.marker_abundance.stool}
\item{QinJ_2012.marker_presence.stool}
\item{QinJ_2012.metaphlan_bugs_list.stool}
\item{QinJ_2012.pathabundance_relab.stool}
\item{QinJ_2012.pathcoverage.stool}
\item{QinN_2014}
\item{QinN_2014.genefamilies_relab.stool}
\item{QinN_2014.marker_abundance.stool}
\item{QinN_2014.marker_presence.stool}
\item{QinN_2014.metaphlan_bugs_list.stool}
\item{QinN_2014.pathabundance_relab.stool}
\item{QinN_2014.pathcoverage.stool}
\item{RampelliS_2015}
\item{RampelliS_2015.genefamilies_relab.stool}
\item{RampelliS_2015.marker_abundance.stool}
\item{RampelliS_2015.marker_presence.stool}
\item{RampelliS_2015.metaphlan_bugs_list.stool}
\item{RampelliS_2015.pathabundance_relab.stool}
\item{RampelliS_2015.pathcoverage.stool}
\item{RaymondF_2016}
\item{RaymondF_2016.genefamilies_relab.stool}
\item{RaymondF_2016.marker_abundance.stool}
\item{RaymondF_2016.marker_presence.stool}
\item{RaymondF_2016.metaphlan_bugs_list.stool}
\item{RaymondF_2016.pathabundance_relab.stool}
\item{RaymondF_2016.pathcoverage.stool}
\item{SchirmerM_2016}
\item{SchirmerM_2016.genefamilies_relab.stool}
\item{SchirmerM_2016.marker_abundance.stool}
\item{SchirmerM_2016.marker_presence.stool}
\item{SchirmerM_2016.metaphlan_bugs_list.stool}
\item{SchirmerM_2016.pathabundance_relab.stool}
\item{SchirmerM_2016.pathcoverage.stool}
\item{ShiB_2015}
\item{ShiB_2015.genefamilies_relab.oralcavity}
\item{ShiB_2015.marker_abundance.oralcavity}
\item{ShiB_2015.marker_presence.oralcavity}
\item{ShiB_2015.metaphlan_bugs_list.oralcavity}
\item{ShiB_2015.pathabundance_relab.oralcavity}
\item{ShiB_2015.pathcoverage.oralcavity}
\item{SmitsSA_2017}
\item{SmitsSA_2017.genefamilies_relab.stool}
\item{SmitsSA_2017.marker_abundance.stool}
\item{SmitsSA_2017.marker_presence.stool}
\item{SmitsSA_2017.metaphlan_bugs_list.stool}
\item{SmitsSA_2017.pathabundance_relab.stool}
\item{SmitsSA_2017.pathcoverage.stool}
\item{TettAJ_2016}
\item{TettAJ_2016.genefamilies_relab.skin}
\item{TettAJ_2016.marker_abundance.skin}
\item{TettAJ_2016.marker_presence.skin}
\item{TettAJ_2016.metaphlan_bugs_list.skin}
\item{TettAJ_2016.pathabundance_relab.skin}
\item{TettAJ_2016.pathcoverage.skin}
\item{TettAJ_2019_a}
\item{TettAJ_2019_a.genefamilies_relab.stool}
\item{TettAJ_2019_a.marker_abundance.stool}
\item{TettAJ_2019_a.marker_presence.stool}
\item{TettAJ_2019_a.metaphlan_bugs_list.stool}
\item{TettAJ_2019_a.pathabundance_relab.stool}
\item{TettAJ_2019_a.pathcoverage.stool}
\item{TettAJ_2019_b}
\item{TettAJ_2019_b.genefamilies_relab.stool}
\item{TettAJ_2019_b.marker_abundance.stool}
\item{TettAJ_2019_b.marker_presence.stool}
\item{TettAJ_2019_b.metaphlan_bugs_list.stool}
\item{TettAJ_2019_b.pathabundance_relab.stool}
\item{TettAJ_2019_b.pathcoverage.stool}
\item{TettAJ_2019_c}
\item{TettAJ_2019_c.genefamilies_relab.stool}
\item{TettAJ_2019_c.marker_abundance.stool}
\item{TettAJ_2019_c.marker_presence.stool}
\item{TettAJ_2019_c.metaphlan_bugs_list.stool}
\item{TettAJ_2019_c.pathabundance_relab.stool}
\item{TettAJ_2019_c.pathcoverage.stool}
\item{ThomasAM_2018a}
\item{ThomasAM_2018a.genefamilies_relab.stool}
\item{ThomasAM_2018a.marker_abundance.stool}
\item{ThomasAM_2018a.marker_presence.stool}
\item{ThomasAM_2018a.metaphlan_bugs_list.stool}
\item{ThomasAM_2018a.pathabundance_relab.stool}
\item{ThomasAM_2018a.pathcoverage.stool}
\item{ThomasAM_2018b.genefamilies_relab.stool}
\item{ThomasAM_2018b}
\item{ThomasAM_2018b.marker_abundance.stool}
\item{ThomasAM_2018b.marker_presence.stool}
\item{ThomasAM_2018b.metaphlan_bugs_list.stool}
\item{ThomasAM_2018b.pathabundance_relab.stool}
\item{ThomasAM_2018b.pathcoverage.stool}
\item{VatanenT_2016}
\item{VatanenT_2016.genefamilies_relab.stool}
\item{VatanenT_2016.marker_abundance.stool}
\item{VatanenT_2016.marker_presence.stool}
\item{VatanenT_2016.metaphlan_bugs_list.stool}
\item{VatanenT_2016.pathabundance_relab.stool}
\item{VatanenT_2016.pathcoverage.stool}
\item{VincentC_2016}
\item{VincentC_2016.genefamilies_relab.stool}
\item{VincentC_2016.marker_abundance.stool}
\item{VincentC_2016.marker_presence.stool}
\item{VincentC_2016.metaphlan_bugs_list.stool}
\item{VincentC_2016.pathabundance_relab.stool}
\item{VincentC_2016.pathcoverage.stool}
\item{VogtmannE_2016}
\item{VogtmannE_2016.genefamilies_relab.stool}
\item{VogtmannE_2016.marker_abundance.stool}
\item{VogtmannE_2016.marker_presence.stool}
\item{VogtmannE_2016.metaphlan_bugs_list.stool}
\item{VogtmannE_2016.pathabundance_relab.stool}
\item{VogtmannE_2016.pathcoverage.stool}
\item{WenC_2017.genefamilies_relab.stool}
\item{WenC_2017.marker_abundance.stool}
\item{WenC_2017.marker_presence.stool}
\item{WenC_2017.metaphlan_bugs_list.stool}
\item{WenC_2017.pathabundance_relab.stool}
\item{WenC_2017.pathcoverage.stool}
\item{XieH_2016}
\item{XieH_2016.genefamilies_relab.stool}
\item{XieH_2016.marker_abundance.stool}
\item{XieH_2016.marker_presence.stool}
\item{XieH_2016.metaphlan_bugs_list.stool}
\item{XieH_2016.pathabundance_relab.stool}
\item{XieH_2016.pathcoverage.stool}
\item{YeZ_2018}
\item{YeZ_2018.genefamilies_relab.stool}
\item{YeZ_2018.marker_abundance.stool}
\item{YeZ_2018.marker_presence.stool}
\item{YeZ_2018.metaphlan_bugs_list.stool}
\item{YeZ_2018.pathabundance_relab.stool}
\item{YeZ_2018.pathcoverage.stool}
\item{YuJ_2015}
\item{YuJ_2015.genefamilies_relab.stool}
\item{YuJ_2015.marker_abundance.stool}
\item{YuJ_2015.marker_presence.stool}
\item{YuJ_2015.metaphlan_bugs_list.stool}
\item{YuJ_2015.pathabundance_relab.stool}
\item{YuJ_2015.pathcoverage.stool}
\item{ZeeviD_2015}
\item{ZeeviD_2015.genefamilies_relab.stool}
\item{ZeeviD_2015.marker_abundance.stool}
\item{ZeeviD_2015.marker_presence.stool}
\item{ZeeviD_2015.metaphlan_bugs_list.stool}
\item{ZeeviD_2015.pathabundance_relab.stool}
\item{ZeeviD_2015.pathcoverage.stool}
\item{ZellerG_2014}
\item{ZellerG_2014.genefamilies_relab.stool}
\item{ZellerG_2014.marker_abundance.stool}
\item{ZellerG_2014.marker_presence.stool}
\item{ZellerG_2014.metaphlan_bugs_list.stool}
\item{ZellerG_2014.pathabundance_relab.stool}
\item{ZellerG_2014.pathcoverage.stool}
}
}

