% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet2MRexperiment.R
\name{ExpressionSet2MRexperiment}
\alias{ExpressionSet2MRexperiment}
\title{Convert an ExpressionSet object to a metagenomeSeq::MRexperiment-class object}
\usage{
ExpressionSet2MRexperiment(eset, simplify = TRUE)
}
\arguments{
\item{eset}{An eset object}

\item{simplify}{if TRUE the most detailed clade name is used, instead of the original
metaPhlAn2 names which contain the full taxonomy.}
}
\value{
A metagenomeSeq::MRexperiment-class object
}
\description{
Convert an ExpressionSet object to a metagenomeSeq::MRexperiment-class object
}
\examples{
eset <- LomanNJ_2013_Mi.metaphlan_bugs_list.stool()
ExpressionSet2MRexperiment(eset)

}
