% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metadata_documentation.R
\name{extdata_dir}
\alias{extdata_dir}
\title{define path to a template-holding folder subordinate to cmg_tmpdir}
\usage{
extdata_dir()
}
\description{
define path to a template-holding folder subordinate to cmg_tmpdir
}
