% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaphlanTree.R
\name{getMetaphlanTree}
\alias{getMetaphlanTree}
\title{Title Return a phylogenetic tree for MetaPhlAn2 bugs}
\usage{
getMetaphlanTree(removeGCF = TRUE, simplify = TRUE)
}
\arguments{
\item{removeGCF}{remove "|GCF_nnnnnnnnn" from the end of tip labels. Default is TRUE.}

\item{simplify}{if TRUE, only the most detailed level of the taxonomy is kept in the names,
for example species or strain.  Default is TRUE.}
}
\value{
a phylogenetic tree of class ape::phylo
}
\description{
Title Return a phylogenetic tree for MetaPhlAn2 bugs
}
\details{
The phylogenetic tree was built with PhyloPhlAn, using all the genomes
from MetaPhlAn2. Clades that had more than one leaf per species were
cleaned and a new tree generated with these selected genomes.
Labels are in the form: "taxonomy|genome_ID".
The Newick file of the tree is stored in the package as
inst/extdata/metaphlan2_selected.tree.reroot.nwk.bz2.
Thanks to Francesco Asnicar <f.asnicar@unitn.it> for generating this tree.
}
\examples{
tree <- getMetaphlanTree()
summary(tree)
getMetaphlanTree(simplify = FALSE)
getMetaphlanTree(simplify = FALSE, removeGCF = FALSE)

}
