% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HanniganGD_2017.R
\name{HanniganGD_2017}
\alias{HanniganGD_2017}
\alias{HanniganGD_2017.genefamilies_relab.stool}
\alias{HanniganGD_2017.marker_abundance.stool}
\alias{HanniganGD_2017.marker_presence.stool}
\alias{HanniganGD_2017.metaphlan_bugs_list.stool}
\alias{HanniganGD_2017.pathabundance_relab.stool}
\alias{HanniganGD_2017.pathcoverage.stool}
\title{Data from the HanniganGD_2017 study}
\description{
Data from the HanniganGD_2017 study
}
\section{Datasets}{
 

\subsection{HanniganGD_2017.genefamilies_relab.stool}{
   An ExpressionSet with 82 samples and 709,894 features specific to the stool body site
}

\subsection{HanniganGD_2017.marker_abundance.stool}{
   An ExpressionSet with 82 samples and 57,511 features specific to the stool body site
}

\subsection{HanniganGD_2017.marker_presence.stool}{
   An ExpressionSet with 82 samples and 52,996 features specific to the stool body site
}

\subsection{HanniganGD_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 82 samples and 716 features specific to the stool body site
}

\subsection{HanniganGD_2017.pathabundance_relab.stool}{
   An ExpressionSet with 82 samples and 8,500 features specific to the stool body site
}

\subsection{HanniganGD_2017.pathcoverage.stool}{
   An ExpressionSet with 82 samples and 8,500 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
`HanniganGD_2017.metaphlan_bugs_list.stool`()

}
