% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HansenLBS_2018.R
\name{HansenLBS_2018}
\alias{HansenLBS_2018}
\alias{HansenLBS_2018.genefamilies_relab.stool}
\alias{HansenLBS_2018.marker_abundance.stool}
\alias{HansenLBS_2018.marker_presence.stool}
\alias{HansenLBS_2018.metaphlan_bugs_list.stool}
\alias{HansenLBS_2018.pathabundance_relab.stool}
\alias{HansenLBS_2018.pathcoverage.stool}
\title{Data from the HansenLBS_2018 study}
\description{
Data from the HansenLBS_2018 study
}
\section{Datasets}{
 

\subsection{HansenLBS_2018.genefamilies_relab.stool}{
   An ExpressionSet with 208 samples and 1,624,505 features specific to the stool body site
}

\subsection{HansenLBS_2018.marker_abundance.stool}{
   An ExpressionSet with 208 samples and 106,894 features specific to the stool body site
}

\subsection{HansenLBS_2018.marker_presence.stool}{
   An ExpressionSet with 208 samples and 100,724 features specific to the stool body site
}

\subsection{HansenLBS_2018.metaphlan_bugs_list.stool}{
   An ExpressionSet with 208 samples and 1,221 features specific to the stool body site
}

\subsection{HansenLBS_2018.pathabundance_relab.stool}{
   An ExpressionSet with 208 samples and 8,797 features specific to the stool body site
}

\subsection{HansenLBS_2018.pathcoverage.stool}{
   An ExpressionSet with 208 samples and 8,797 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   A low-gluten diet induces changes in the intestinal microbiome of healthy Danish adults.
}

\subsection{Author}{
   Hansen LBS, Roager HM, Søndertoft NB, Gøbel RJ, Kristensen M, Vallès-Colomer M, Vieira-Silva S, Ibrügger S, Lind MV, Mærkedahl RB, Bahl MI, Madsen ML, Havelund J, Falony G, Tetens I, Nielsen T, Allin KH, Frandsen HL, Hartmann B, Holst JJ, Sparholt MH, Holck J, Blennow A, Moll JM, Meyer AS, Hoppe C, Poulsen JH, Carvalho V, Sagnelli D, Dalgaard MD, Christensen AF, Lydolph MC, Ross AB, Villas-Bôas S, Brix S, Sicheritz-Pontén T, Buschard K, Linneberg A, Rumessen JJ, Ekstrøm CT, Ritz C, Kristiansen K, Nielsen HB, Vestergaard H, Færgeman NJ, Raes J, Frøkiær H, Hansen T, Lauritzen L, Gupta R, Licht TR, Pedersen O
}

\subsection{Lab}{
   [1] Department of Bio and Health Informatics, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark., [2] National Food Institute, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark., [3] Department of Nutrition, Exercise and Sports, Faculty of Science, University of Copenhagen, DK-1958, Frederiksberg, Denmark., [4] The Novo Nordisk Foundation Center for Basic Metabolic Research, University of Copenhagen, DK-2200, Copenhagen, Denmark., [5] Department of Microbiology and Immunology, KU Leuven-University of Leuven, Rega Institute, 3000, Leuven, Belgium., [6] VIB, Center for Microbiology, 3000, Leuven, Belgium., [7] Department of Veterinary Disease Biology, Faculty of Science, University of Copenhagen, DK-1958, Frederiksberg, Denmark., [8] Department of Biochemistry and Molecular Biology, University of Southern Denmark, DK-5230, Odense, Denmark., [9] Department of Biomedical Sciences, University of Copenhagen, Copenhagen, DK-2200, Denmark., [10] Department of Radiology, Bispebjerg Hospital, Copenhagen, DK-2400, Denmark., [11] Department of Chemical and Biochemical Engineering, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark., [12] Department of Plant and Environmental Sciences, University of Copenhagen, DK-1958, Frederiksberg, Denmark., [13] Department of Biotechnology and Biomedicine, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark., [14] Department of Clinical Biochemistry, Copenhagen University Hospital Hvidovre, DK-2650, Hvidovre, Denmark., [15] Department of Autoimmunology &amp; Biomarkers, Statens Serum Institut, DK-2300, Copenhagen, Denmark., [16] Department of Biology and Biological Engineering, Chalmers University of Technology, 412 96, Gothenburg, Sweden., [17] School of Biological Sciences, The University of Auckland, 1010, Auckland, New Zealand., [18] Bartholin Institute, Rigshospitalet, DK-2200, Copenhagen, Denmark., [19] Research Centre for Prevention and Health, The Capital Region of Denmark, DK-2000, Frederiksberg, Denmark., [20] Research Unit and Department of Gastroenterology, Herlev and Gentofte Hospital, the Capital Region of Denmark, 2730, Herlev, Denmark., [21] Biostatistics, Department of Public Health, University of Copenhagen, DK-1014, Copenhagen, Denmark., [22] Laboratory of Genomics and Molecular Biomedicine, Department of Biology, University of Copenhagen, DK-2100, Copenhagen, Denmark., [23] Clinical-Microbiomics A/S, DK-2200, Copenhagen, Denmark., [24] Department of Bio and Health Informatics, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark. ramneek@bioinformatics.dtu.dk., [25] National Food Institute, Technical University of Denmark, DK-2800, Kgs. Lyngby, Denmark. trli@food.dtu.dk., [26] The Novo Nordisk Foundation Center for Basic Metabolic Research, University of Copenhagen, DK-2200, Copenhagen, Denmark. oluf@sund.ku.dk.
}

\subsection{PMID}{
   30425247
}
}

\examples{
HansenLBS_2018.metaphlan_bugs_list.stool()

}
