% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heitz_BuschartA_2016.R
\name{Heitz_BuschartA_2016}
\alias{Heitz_BuschartA_2016}
\alias{Heitz_BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz_BuschartA_2016.marker_abundance.stool}
\alias{Heitz_BuschartA_2016.marker_presence.stool}
\alias{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz_BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz_BuschartA_2016.pathcoverage.stool}
\title{Data from the Heitz_BuschartA_2016 study}
\description{
Data from the Heitz_BuschartA_2016 study
}
\section{Datasets}{
 

\subsection{Heitz_BuschartA_2016.genefamilies_relab.stool}{
   An ExpressionSet with 53 samples and 1,110,454 features specific to the stool body site
}

\subsection{Heitz_BuschartA_2016.marker_abundance.stool}{
   An ExpressionSet with 53 samples and 80,864 features specific to the stool body site
}

\subsection{Heitz_BuschartA_2016.marker_presence.stool}{
   An ExpressionSet with 53 samples and 77,488 features specific to the stool body site
}

\subsection{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 53 samples and 1,011 features specific to the stool body site
}

\subsection{Heitz_BuschartA_2016.pathabundance_relab.stool}{
   An ExpressionSet with 53 samples and 10,281 features specific to the stool body site
}

\subsection{Heitz_BuschartA_2016.pathcoverage.stool}{
   An ExpressionSet with 53 samples and 10,281 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Integrated multi-omics of the human gut microbiome in a case study of familial type 1 diabetes.
}

\subsection{Author}{
   Heintz-Buschart A, May P, Laczny CC, Lebrun LA, Bellora C, Krishna A, Wampach L, Schneider JG, Hogan A, de Beaufort C, Wilmes P
}

\subsection{Lab}{
   [1] Luxembourg Centre for Systems Biomedicine, 7 avenue des Hauts-Fourneaux, 4362 Esch-sur-Alzette, Luxembourg., [2] Integrated BioBank of Luxembourg, 6 rue Nicolas Ernest Barble, 1210 Luxembourg, Luxembourg., [3] Department of Internal Medicine II, Saarland University Medical Center, 66421 Homburg, Germany., [4] Centre Hospitalier Emile Mayrisch, Rue Emile Mayrisch, 4240 Esch-sur-Alzette, Luxembourg., [5] Clinique Pediatrique - Centre Hospitalier de Luxembourg, 4 rue Nicolas Ernest Barble, 1210 Luxembourg.
}

\subsection{PMID}{
   27723761
}
}

\examples{
Heitz_BuschartA_2016.metaphlan_bugs_list.stool()

}
