% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMP_2012.pathcoverage.R
\name{HMP_2012.pathcoverage}
\alias{HMP_2012.pathcoverage}
\alias{HMP_2012.pathcoverage.anterior_nares}
\title{Pathcoverage data from the HMP_2012 dataset}
\source{
Department of Psychology, Abdul Haq Campus, Federal Urdu University for Arts, Science and Technology, Karachi, Pakistan. shahiq_psy@yahoo.com
}
\description{
Pathcoverage data from the HMP_2012 dataset
}
\section{Datasets}{
 

\subsection{HMP_2012.pathcoverage.anterior_nares}{
   An ExpressionSet with 93 samples and 23,271 features specific to the anterior_nares bodysite
}
}
\seealso{
\code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}, \code{\link{HMP_2012.pathabundance_relab}}
}

