% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMP_2012.R
\name{HMP_2012}
\alias{HMP_2012}
\alias{HMP_2012.genefamilies_relab.nasalcavity}
\alias{HMP_2012.genefamilies_relab.oralcavity}
\alias{HMP_2012.genefamilies_relab.skin}
\alias{HMP_2012.genefamilies_relab.stool}
\alias{HMP_2012.genefamilies_relab.vagina}
\alias{HMP_2012.marker_abundance.nasalcavity}
\alias{HMP_2012.marker_abundance.oralcavity}
\alias{HMP_2012.marker_abundance.skin}
\alias{HMP_2012.marker_abundance.stool}
\alias{HMP_2012.marker_abundance.vagina}
\alias{HMP_2012.marker_presence.nasalcavity}
\alias{HMP_2012.marker_presence.oralcavity}
\alias{HMP_2012.marker_presence.skin}
\alias{HMP_2012.marker_presence.stool}
\alias{HMP_2012.marker_presence.vagina}
\alias{HMP_2012.metaphlan_bugs_list.nasalcavity}
\alias{HMP_2012.metaphlan_bugs_list.oralcavity}
\alias{HMP_2012.metaphlan_bugs_list.skin}
\alias{HMP_2012.metaphlan_bugs_list.stool}
\alias{HMP_2012.metaphlan_bugs_list.vagina}
\alias{HMP_2012.pathabundance_relab.nasalcavity}
\alias{HMP_2012.pathabundance_relab.oralcavity}
\alias{HMP_2012.pathabundance_relab.skin}
\alias{HMP_2012.pathabundance_relab.stool}
\alias{HMP_2012.pathabundance_relab.vagina}
\alias{HMP_2012.pathcoverage.nasalcavity}
\alias{HMP_2012.pathcoverage.oralcavity}
\alias{HMP_2012.pathcoverage.skin}
\alias{HMP_2012.pathcoverage.stool}
\alias{HMP_2012.pathcoverage.vagina}
\title{Data from the HMP_2012 study}
\description{
Data from the HMP_2012 study
}
\section{Datasets}{
 

\subsection{HMP_2012.genefamilies_relab.nasalcavity}{
   An ExpressionSet with 93 samples and 1,964,480 features specific to the nasalcavity body site
}

\subsection{HMP_2012.genefamilies_relab.oralcavity}{
   An ExpressionSet with 415 samples and 1,964,480 features specific to the oralcavity body site
}

\subsection{HMP_2012.genefamilies_relab.skin}{
   An ExpressionSet with 27 samples and 1,964,480 features specific to the skin body site
}

\subsection{HMP_2012.genefamilies_relab.stool}{
   An ExpressionSet with 147 samples and 1,964,480 features specific to the stool body site
}

\subsection{HMP_2012.genefamilies_relab.vagina}{
   An ExpressionSet with 67 samples and 1,964,480 features specific to the vagina body site
}

\subsection{HMP_2012.marker_abundance.nasalcavity}{
   An ExpressionSet with 93 samples and 162,107 features specific to the nasalcavity body site
}

\subsection{HMP_2012.marker_abundance.oralcavity}{
   An ExpressionSet with 415 samples and 162,107 features specific to the oralcavity body site
}

\subsection{HMP_2012.marker_abundance.skin}{
   An ExpressionSet with 27 samples and 162,107 features specific to the skin body site
}

\subsection{HMP_2012.marker_abundance.stool}{
   An ExpressionSet with 147 samples and 162,107 features specific to the stool body site
}

\subsection{HMP_2012.marker_abundance.vagina}{
   An ExpressionSet with 67 samples and 162,107 features specific to the vagina body site
}

\subsection{HMP_2012.marker_presence.nasalcavity}{
   An ExpressionSet with 93 samples and 158,645 features specific to the nasalcavity body site
}

\subsection{HMP_2012.marker_presence.oralcavity}{
   An ExpressionSet with 415 samples and 158,645 features specific to the oralcavity body site
}

\subsection{HMP_2012.marker_presence.skin}{
   An ExpressionSet with 27 samples and 158,645 features specific to the skin body site
}

\subsection{HMP_2012.marker_presence.stool}{
   An ExpressionSet with 147 samples and 158,645 features specific to the stool body site
}

\subsection{HMP_2012.marker_presence.vagina}{
   An ExpressionSet with 67 samples and 158,645 features specific to the vagina body site
}

\subsection{HMP_2012.metaphlan_bugs_list.nasalcavity}{
   An ExpressionSet with 93 samples and 1,988 features specific to the nasalcavity body site
}

\subsection{HMP_2012.metaphlan_bugs_list.oralcavity}{
   An ExpressionSet with 415 samples and 1,988 features specific to the oralcavity body site
}

\subsection{HMP_2012.metaphlan_bugs_list.skin}{
   An ExpressionSet with 27 samples and 1,988 features specific to the skin body site
}

\subsection{HMP_2012.metaphlan_bugs_list.stool}{
   An ExpressionSet with 147 samples and 1,988 features specific to the stool body site
}

\subsection{HMP_2012.metaphlan_bugs_list.vagina}{
   An ExpressionSet with 67 samples and 1,988 features specific to the vagina body site
}

\subsection{HMP_2012.pathabundance_relab.nasalcavity}{
   An ExpressionSet with 93 samples and 23,271 features specific to the nasalcavity body site
}

\subsection{HMP_2012.pathabundance_relab.oralcavity}{
   An ExpressionSet with 415 samples and 23,271 features specific to the oralcavity body site
}

\subsection{HMP_2012.pathabundance_relab.skin}{
   An ExpressionSet with 27 samples and 23,271 features specific to the skin body site
}

\subsection{HMP_2012.pathabundance_relab.stool}{
   An ExpressionSet with 147 samples and 23,271 features specific to the stool body site
}

\subsection{HMP_2012.pathabundance_relab.vagina}{
   An ExpressionSet with 67 samples and 23,271 features specific to the vagina body site
}

\subsection{HMP_2012.pathcoverage.nasalcavity}{
   An ExpressionSet with 93 samples and 23,271 features specific to the nasalcavity body site
}

\subsection{HMP_2012.pathcoverage.oralcavity}{
   An ExpressionSet with 415 samples and 23,271 features specific to the oralcavity body site
}

\subsection{HMP_2012.pathcoverage.skin}{
   An ExpressionSet with 27 samples and 23,271 features specific to the skin body site
}

\subsection{HMP_2012.pathcoverage.stool}{
   An ExpressionSet with 147 samples and 23,271 features specific to the stool body site
}

\subsection{HMP_2012.pathcoverage.vagina}{
   An ExpressionSet with 67 samples and 23,271 features specific to the vagina body site
}
}

\section{Source}{
 

\subsection{Title}{
   Structure, function and diversity of the healthy human microbiome.
}

\subsection{Author}{
   Huttenhower C, Gevers D, Knight R, Abubucker S, Badger JH, Chinwalla AT, Creasy HH, Earl AM, FitzGerald MG, Fulton RS, Giglio MG, Hallsworth-Pepin K, Lobos EA, Madupu R, Magrini V, Martin JC, Mitreva M, Muzny DM, Sodergren EJ, Versalovic J, Wollam AM, Worley KC, Wortman JR, Young SK, Zeng Q, Aagaard KM, Abolude OO, Allen-Vercoe E, Alm EJ, Alvarado L, Andersen GL, Anderson S, Appelbaum E, Arachchi HM, Armitage G, Arze CA, Ayvaz T, Baker CC, Begg L, Belachew T, Bhonagiri V, Bihan M, Blaser MJ, Bloom T, Bonazzi V, Brooks J, Buck GA, Buhay CJ, Busam DA, Campbell JL, Canon SR, Cantarel BL, Chain PS, Chen IM, Chen L, Chhibba S, Chu K, Ciulla DM, Clemente JC, Clifton SW, Conlan S, Crabtree J, Cutting MA, Davidovics NJ, Davis CC, DeSantis TZ, Deal C, Delehaunty KD, Dewhirst FE, Deych E, Ding Y, Dooling DJ, Dugan SP, Dunne WM, Durkin A, Edgar RC, Erlich RL, Farmer CN, Farrell RM, Faust K, Feldgarden M, Felix VM, Fisher S, Fodor AA, Forney LJ, Foster L, Di Francesco V, Friedman J, Friedrich DC, Fronick CC, Fulton LL, Gao H, Garcia N, Giannoukos G, Giblin C, Giovanni MY, Goldberg JM, Goll J, Gonzalez A, Griggs A, Gujja S, Haake SK, Haas BJ, Hamilton HA, Harris EL, Hepburn TA, Herter B, Hoffmann DE, Holder ME, Howarth C, Huang KH, Huse SM, Izard J, Jansson JK, Jiang H, Jordan C, Joshi V, Katancik JA, Keitel WA, Kelley ST, Kells C, King NB, Knights D, Kong HH, Koren O, Koren S, Kota KC, Kovar CL, Kyrpides NC, La Rosa PS, Lee SL, Lemon KP, Lennon N, Lewis CM, Lewis L, Ley RE, Li K, Liolios K, Liu B, Liu Y, Lo CC, Lozupone CA, Lunsford R, Madden T, Mahurkar AA, Mannon PJ, Mardis ER, Markowitz VM, Mavromatis K, McCorrison JM, McDonald D, McEwen J, McGuire AL, McInnes P, Mehta T, Mihindukulasuriya KA, Miller JR, Minx PJ, Newsham I, Nusbaum C, O'Laughlin M, Orvis J, Pagani I, Palaniappan K, Patel SM, Pearson M, Peterson J, Podar M, Pohl C, Pollard KS, Pop M, Priest ME, Proctor LM, Qin X, Raes J, Ravel J, Reid JG, Rho M, Rhodes R, Riehle KP, Rivera MC, Rodriguez-Mueller B, Rogers YH, Ross MC, Russ C, Sanka RK, Sankar P, Sathirapongsasuti J, Schloss JA, Schloss PD, Schmidt TM, Scholz M, Schriml L, Schubert AM, Segata N, Segre JA, Shannon WD, Sharp RR, Sharpton TJ, Shenoy N, Sheth NU, Simone GA, Singh I, Smillie CS, Sobel JD, Sommer DD, Spicer P, Sutton GG, Sykes SM, Tabbaa DG, Thiagarajan M, Tomlinson CM, Torralba M, Treangen TJ, Truty RM, Vishnivetskaya TA, Walker J, Wang L, Wang Z, Ward DV, Warren W, Watson MA, Wellington C, Wetterstrand KA, White JR, Wilczek-Boney K, Wu Y, Wylie KM, Wylie T, Yandava C, Ye L, Ye Y, Yooseph S, Youmans BP, Zhang L, Zhou Y, Zhu Y, Zoloth L, Zucker JD, Birren BW, Gibbs RA, Highlander SK, Methé BA, Nelson KE, Petrosino JF, Weinstock GM, Wilson RK, White O
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   22699609
}
}

\examples{
HMP_2012.metaphlan_bugs_list.nasalcavity()

}
