% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JieZ_2017.R
\name{JieZ_2017}
\alias{JieZ_2017}
\alias{JieZ_2017.genefamilies_relab.stool}
\alias{JieZ_2017.marker_abundance.stool}
\alias{JieZ_2017.marker_presence.stool}
\alias{JieZ_2017.metaphlan_bugs_list.stool}
\alias{JieZ_2017.pathabundance_relab.stool}
\alias{JieZ_2017.pathcoverage.stool}
\title{Data from the JieZ_2017 study}
\description{
Data from the JieZ_2017 study
}
\section{Datasets}{
 

\subsection{JieZ_2017.genefamilies_relab.stool}{
   An ExpressionSet with 385 samples and 1,976,093 features specific to the stool body site
}

\subsection{JieZ_2017.marker_abundance.stool}{
   An ExpressionSet with 385 samples and 158,072 features specific to the stool body site
}

\subsection{JieZ_2017.marker_presence.stool}{
   An ExpressionSet with 385 samples and 145,642 features specific to the stool body site
}

\subsection{JieZ_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 385 samples and 1,666 features specific to the stool body site
}

\subsection{JieZ_2017.pathabundance_relab.stool}{
   An ExpressionSet with 385 samples and 13,222 features specific to the stool body site
}

\subsection{JieZ_2017.pathcoverage.stool}{
   An ExpressionSet with 385 samples and 13,222 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   The gut microbiome in atherosclerotic cardiovascular disease.
}

\subsection{Author}{
   Jie Z, Xia H, Zhong SL, Feng Q, Li S, Liang S, Zhong H, Liu Z, Gao Y, Zhao H, Zhang D, Su Z, Fang Z, Lan Z, Li J, Xiao L, Li J, Li R, Li X, Li F, Ren H, Huang Y, Peng Y, Li G, Wen B, Dong B, Chen JY, Geng QS, Zhang ZW, Yang H, Wang J, Wang J, Zhang X, Madsen L, Brix S, Ning G, Xu X, Liu X, Hou Y, Jia H, He K, Kristiansen K
}

\subsection{Lab}{
   [1] BGI-Shenzhen, Shenzhen, 518083, China., [2] China National Genebank, Shenzhen, 518120, China., [3] Shenzhen Key Laboratory of Human Commensal Microorganisms and Health Research, BGI-Shenzhen, Shenzhen, 518083, China., [4] Guangdong Provincial Key Laboratory of Coronary Heart Disease Prevention, Guangdong Cardiovascular Institute, Guangzhou, 510080, China., [5] Medical Research Center of Guangdong General Hospital, Guangdong Academy of Medical Sciences, Guangzhou, 510080, China., [6] Shenzhen Engineering Laboratory of Detection and Intervention of Human Intestinal Microbiome, Shenzhen, 518083, China., [7] Department of Biology, Laboratory of Genomics and Molecular Biomedicine, University of Copenhagen, Universitetsparken 13, 2100, Copenhagen, Denmark., [8] Department of Human Microbiome, School of Stomatology, Shandong University, Shandong Provincial Key Laboratory of Oral Tissue Regeneration, Jinan, 250012, China., [9] BGI Education Center, University of Chinese Academy of Sciences, Shenzhen, 518083, China., [10] School of Bioscience and Biotechnology, South China University of Technology, Guangzhou, 510006, China., [11] Beijing Key Laboratory for Precision Medicine of Chronic Heart Failure, Chinese PLA General Hospital, Beijing, 100853, China., [12] Center for Genome Sciences &amp; Systems Biology, Washington University School of Medicine, St. Louis, MO, 63110, USA., [13] James D. Watson Institute of Genome Sciences, Hangzhou, 310000, China., [14] Macau University of Science and Technology, Macau, 999078, China., [15] iCarbonX, Shenzhen, 518053, China., [16] Department of Rheumatology and Clinical Immunology, Peking Union Medical College Hospital, Chinese Academy of Medical Sciences and Peking Union Medical College, Beijing, 100730, China., [17] National Institute of Nutrition and Seafood Research, (NIFES), Postboks 2029, Nordnes, N-5817, Bergen, Norway., [18] Department of Biotechnology and Biomedicine, Technical University of Denmark (DTU), 2800, Kongens Lyngby, Denmark., [19] Department of Endocrinology and Metabolism, State Key Laboratory of Medical Genomes, National Clinical Research Center for Metabolic Diseases, Shanghai Clinical Center for Endocrine and Metabolic Diseases, Shanghai Institute of Endocrine and Metabolic Diseases, Ruijin Hospital, Shanghai Jiao Tong University School of Medicine, Shanghai, 200025, China., [20] BGI-Shenzhen, Shenzhen, 518083, China. jiahuijue@genomics.cn., [21] China National Genebank, Shenzhen, 518120, China. jiahuijue@genomics.cn., [22] Shenzhen Key Laboratory of Human Commensal Microorganisms and Health Research, BGI-Shenzhen, Shenzhen, 518083, China. jiahuijue@genomics.cn., [23] Macau University of Science and Technology, Macau, 999078, China. jiahuijue@genomics.cn., [24] Beijing Key Laboratory for Precision Medicine of Chronic Heart Failure, Chinese PLA General Hospital, Beijing, 100853, China. hekl301@aliyun.com., [25] BGI-Shenzhen, Shenzhen, 518083, China. kk@bio.ku.dk., [26] China National Genebank, Shenzhen, 518120, China. kk@bio.ku.dk., [27] Department of Biology, Laboratory of Genomics and Molecular Biomedicine, University of Copenhagen, Universitetsparken 13, 2100, Copenhagen, Denmark. kk@bio.ku.dk.
}

\subsection{PMID}{
   29018189
}
}

\examples{
JieZ_2017.metaphlan_bugs_list.stool()

}
