% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KarlssonFH_2013.marker_abundance.R
\name{KarlssonFH_2013.marker_abundance}
\alias{KarlssonFH_2013.marker_abundance}
\alias{KarlssonFH_2013.marker_abundance.stool}
\title{Marker abundance data from the KarlssonFH_2013 dataset}
\source{
Department of Chemical and Biological Engineering, Chalmers University of Technology, SE-412 96 Gothenburg, Sweden.
}
\description{
Marker abundance data from the KarlssonFH_2013 dataset
}
\section{Datasets}{
 

\subsection{KarlssonFH_2013.marker_abundance.stool}{
   An ExpressionSet with 145 samples and 101,168 features specific to the stool bodysite
}
}
\seealso{
\code{\link{KarlssonFH_2013.genefamilies_relab}}, \code{\link{KarlssonFH_2013.marker_presence}}, \code{\link{KarlssonFH_2013.metaphlan_bugs_list}}, \code{\link{KarlssonFH_2013.pathabundance_relab}}, \code{\link{KarlssonFH_2013.pathcoverage}}
}

