% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KarlssonFH_2013.R
\name{KarlssonFH_2013}
\alias{KarlssonFH_2013}
\alias{KarlssonFH_2013.genefamilies_relab.stool}
\alias{KarlssonFH_2013.marker_abundance.stool}
\alias{KarlssonFH_2013.marker_presence.stool}
\alias{KarlssonFH_2013.metaphlan_bugs_list.stool}
\alias{KarlssonFH_2013.pathabundance_relab.stool}
\alias{KarlssonFH_2013.pathcoverage.stool}
\title{Data from the KarlssonFH_2013 study}
\description{
Data from the KarlssonFH_2013 study
}
\section{Datasets}{
 

\subsection{KarlssonFH_2013.genefamilies_relab.stool}{
   An ExpressionSet with 145 samples and 1,415,750 features specific to the stool body site
}

\subsection{KarlssonFH_2013.marker_abundance.stool}{
   An ExpressionSet with 145 samples and 101,166 features specific to the stool body site
}

\subsection{KarlssonFH_2013.marker_presence.stool}{
   An ExpressionSet with 145 samples and 95,324 features specific to the stool body site
}

\subsection{KarlssonFH_2013.metaphlan_bugs_list.stool}{
   An ExpressionSet with 145 samples and 1,140 features specific to the stool body site
}

\subsection{KarlssonFH_2013.pathabundance_relab.stool}{
   An ExpressionSet with 145 samples and 13,392 features specific to the stool body site
}

\subsection{KarlssonFH_2013.pathcoverage.stool}{
   An ExpressionSet with 145 samples and 13,392 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Gut metagenome in European women with normal, impaired and diabetic glucose control.
}

\subsection{Author}{
   Karlsson FH, Tremaroli V, Nookaew I, Bergström G, Behre CJ, Fagerberg B, Nielsen J, Bäckhed F
}

\subsection{Lab}{
   [1] Department of Chemical and Biological Engineering, Chalmers University of Technology, SE-412 96 Gothenburg, Sweden.
}

\subsection{PMID}{
   23719380
}
}

\examples{
KarlssonFH_2013.metaphlan_bugs_list.stool()

}
