% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KieserS_2018.R
\name{KieserS_2018}
\alias{KieserS_2018}
\alias{KieserS_2018.genefamilies_relab.stool}
\alias{KieserS_2018.marker_abundance.stool}
\alias{KieserS_2018.marker_presence.stool}
\alias{KieserS_2018.metaphlan_bugs_list.stool}
\alias{KieserS_2018.pathabundance_relab.stool}
\alias{KieserS_2018.pathcoverage.stool}
\title{Data from the KieserS_2018 study}
\description{
Data from the KieserS_2018 study
}
\section{Datasets}{
 

\subsection{KieserS_2018.genefamilies_relab.stool}{
   An ExpressionSet with 27 samples and 599,293 features specific to the stool body site
}

\subsection{KieserS_2018.marker_abundance.stool}{
   An ExpressionSet with 27 samples and 43,611 features specific to the stool body site
}

\subsection{KieserS_2018.marker_presence.stool}{
   An ExpressionSet with 27 samples and 41,452 features specific to the stool body site
}

\subsection{KieserS_2018.metaphlan_bugs_list.stool}{
   An ExpressionSet with 27 samples and 723 features specific to the stool body site
}

\subsection{KieserS_2018.pathabundance_relab.stool}{
   An ExpressionSet with 27 samples and 9,031 features specific to the stool body site
}

\subsection{KieserS_2018.pathcoverage.stool}{
   An ExpressionSet with 27 samples and 9,031 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Bangladeshi children with acute diarrhoea show faecal microbiomes with increased Streptococcus abundance, irrespective of diarrhoea aetiology.
}

\subsection{Author}{
   Kieser S, Sarker SA, Sakwinska O, Foata F, Sultana S, Khan Z, Islam S, Porta N, Combremont S, Betrisey B, Fournier C, Charpagne A, Descombes P, Mercenier A, Berger B, Brüssow H
}

\subsection{Lab}{
   [1] Gut Ecosystem Department, Institute of Nutritional Science, Nestle Research Centre, Vers-chez-les-Blanc, CH-1000, Lausanne 26, Switzerland., [2] International Centre for Diarrheal Diseases Research, Bangladesh (icddr,b), Nutrition and Clinical Services Division, 68 Shaheed Tajuddin Ahmed Sharani, Mohakhali, Dhaka 1212, Bangladesh., [3] Nestle Institute of Health Sciences, EPFL Innovation Park, CH-1015, Lausanne, Switzerland.
}

\subsection{PMID}{
   29786169
}
}

\examples{
KieserS_2018.metaphlan_bugs_list.stool()

}
