% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KosticAD_2015.R
\name{KosticAD_2015}
\alias{KosticAD_2015}
\alias{KosticAD_2015.genefamilies_relab.stool}
\alias{KosticAD_2015.marker_abundance.stool}
\alias{KosticAD_2015.marker_presence.stool}
\alias{KosticAD_2015.metaphlan_bugs_list.stool}
\alias{KosticAD_2015.pathabundance_relab.stool}
\alias{KosticAD_2015.pathcoverage.stool}
\title{Data from the KosticAD_2015 study}
\description{
Data from the KosticAD_2015 study
}
\section{Datasets}{
 

\subsection{KosticAD_2015.genefamilies_relab.stool}{
   An ExpressionSet with 124 samples and 994,216 features specific to the stool body site
}

\subsection{KosticAD_2015.marker_abundance.stool}{
   An ExpressionSet with 124 samples and 74,501 features specific to the stool body site
}

\subsection{KosticAD_2015.marker_presence.stool}{
   An ExpressionSet with 124 samples and 72,708 features specific to the stool body site
}

\subsection{KosticAD_2015.metaphlan_bugs_list.stool}{
   An ExpressionSet with 124 samples and 869 features specific to the stool body site
}

\subsection{KosticAD_2015.pathabundance_relab.stool}{
   An ExpressionSet with 124 samples and 10,948 features specific to the stool body site
}

\subsection{KosticAD_2015.pathcoverage.stool}{
   An ExpressionSet with 124 samples and 10,948 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   The dynamics of the human infant gut microbiome in development and in progression toward type 1 diabetes.
}

\subsection{Author}{
   Kostic AD, Gevers D, Siljander H, Vatanen T, Hyötyläinen T, Hämäläinen AM, Peet A, Tillmann V, Pöhö P, Mattila I, Lähdesmäki H, Franzosa EA, Vaarala O, de Goffau M, Harmsen H, Ilonen J, Virtanen SM, Clish CB, Orešič M, Huttenhower C, Knip M, Xavier RJ, Knip M, Koski K, Koski M, Härkönen T, Ryhänen S, Siljander H, Hämäläinen A, Ormisson A, Peet A, Tillmann V, Ulich V, Kuzmicheva E, Mokurov S, Markova S, Pylova S, Isakova M, Shakurova E, Petrov V, Dorshakova NV, Karapetyan T, Varlamova T, Ilonen J, Kiviniemi M, Alnek K, Janson H, Uibo R, Salum T, von Mutius E, Weber J, Ahlfors H, Kallionpää H, Laajala E, Lahesmaa R, Lähdesmäki H, Moulder R, Nieminen J, Ruohtula T, Vaarala O, Honkanen H, Hyöty H, Kondrashova A, Oikarinen S, Harmsen HJ, De Goffau MC, Welling G, Alahuhta K, Korhonen T, Virtanen SM
}

\subsection{Lab}{
   [1] Broad Institute of MIT and Harvard, Cambridge, MA 02142, USA; Center for Computational and Integrative Biology, Massachusetts General Hospital and Harvard Medical School, Boston, MA 02114, USA; Department of Biostatistics, Harvard School of Public Health, Boston, MA 02115, USA., [2] Broad Institute of MIT and Harvard, Cambridge, MA 02142, USA., [3] Children's Hospital, University of Helsinki and Helsinki University Hospital, 00290 Helsinki, Finland; Research Program Unit, Diabetes and Obesity, University of Helsinki, 00290 Helsinki, Finland., [4] Broad Institute of MIT and Harvard, Cambridge, MA 02142, USA; Department of Information and Computer Science, Aalto University School of Science, 02150 Espoo, Finland., [5] Steno Diabetes Center, 2820 Gentofte, Denmark; VTT Technical Research Centre of Finland, 02044 Espoo, Finland., [6] Department of Pediatrics, Jorvi Hospital, 02740 Espoo, Finland., [7] Department of Pediatrics, University of Tartu, Estonia and Tartu University Hospital, 51014 Tartu, Estonia., [8] Faculty of Pharmacy, University of Helsinki, 00290 Helsinki, Finland; VTT Technical Research Centre of Finland, 02044 Espoo, Finland., [9] Department of Information and Computer Science, Aalto University School of Science, 02150 Espoo, Finland., [10] Department of Biostatistics, Harvard School of Public Health, Boston, MA 02115, USA., [11] Research Program Unit, Diabetes and Obesity, University of Helsinki, 00290 Helsinki, Finland., [12] Department of Medical Microbiology, University Medical Center Groningen and University of Groningen, 9713 GZ Groningen, the Netherlands., [13] Immunogenetics Laboratory, University of Turku, 20520 Turku, Finland; Department of Clinical Microbiology, University of Eastern Finland, 70211 Kuopio, Finland., [14] Department of Lifestyle and Participation, National Institute for Health and Welfare, 00271 Helsinki, Finland; School of Health Sciences, University of Tampere, 33014 Tampere, Finland; Science Centre, Pirkanmaa Hospital District, 33521 Tampere, Finland., [15] Broad Institute of MIT and Harvard, Cambridge, MA 02142, USA; Department of Biostatistics, Harvard School of Public Health, Boston, MA 02115, USA., [16] Children's Hospital, University of Helsinki and Helsinki University Hospital, 00290 Helsinki, Finland; Research Program Unit, Diabetes and Obesity, University of Helsinki, 00290 Helsinki, Finland; Folkhalsan Research Center, 00290 Helsinki, Finland; Department of Pediatrics, Tampere University Hospital, 33521 Tampere, Finland., [17] Broad Institute of MIT and Harvard, Cambridge, MA 02142, USA; Center for Computational and Integrative Biology, Massachusetts General Hospital and Harvard Medical School, Boston, MA 02114, USA; Gastrointestinal Unit and Center for the Study of Inflammatory Bowel Disease, Massachusetts General Hospital and Harvard Medical School, Boston, MA 02114, USA; Center for Microbiome Informatics and Therapeutics, Massachusetts Institute of Technology, Cambridge, MA 02139, USA. Electronic address: xavier@molbio.mgh.harvard.edu.
}

\subsection{PMID}{
   25662751
}
}

\examples{
`KosticAD_2015.metaphlan_bugs_list.stool`()

}
