% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeChatelierE_2013.marker_abundance.R
\name{LeChatelierE_2013.marker_abundance}
\alias{LeChatelierE_2013.marker_abundance}
\alias{LeChatelierE_2013.marker_abundance.stool}
\title{Marker abundance data from the LeChatelierE_2013 dataset}
\source{
INRA, Institut National de la Recherche Agronomique, US1367 Metagenopolis, 78350 Jouy en Josas, France.
}
\description{
Marker abundance data from the LeChatelierE_2013 dataset
}
\section{Datasets}{
 

\subsection{LeChatelierE_2013.marker_abundance.stool}{
   An ExpressionSet with 292 samples and 130,621 features specific to the stool bodysite
}
}

\examples{
LeChatelierE_2013.marker_abundance.stool()

}
\seealso{
\code{\link{LeChatelierE_2013.genefamilies_relab}}, \code{\link{LeChatelierE_2013.marker_presence}}, \code{\link{LeChatelierE_2013.metaphlan_bugs_list}}, \code{\link{LeChatelierE_2013.pathabundance_relab}}, \code{\link{LeChatelierE_2013.pathcoverage}}
}
