% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeChatelierE_2013.marker_presence.R
\name{LeChatelierE_2013.marker_presence}
\alias{LeChatelierE_2013.marker_presence}
\alias{LeChatelierE_2013.marker_presence.stool}
\title{Marker presence data from the LeChatelierE_2013 dataset}
\source{
INRA, Institut National de la Recherche Agronomique, US1367 Metagenopolis, 78350 Jouy en Josas, France.
}
\description{
Marker presence data from the LeChatelierE_2013 dataset
}
\section{Datasets}{
 

\subsection{LeChatelierE_2013.marker_presence.stool}{
   An ExpressionSet with 292 samples and 117,258 features specific to the stool bodysite
}
}
\seealso{
\code{\link{LeChatelierE_2013.genefamilies_relab}}, \code{\link{LeChatelierE_2013.marker_abundance}}, \code{\link{LeChatelierE_2013.metaphlan_bugs_list}}, \code{\link{LeChatelierE_2013.pathabundance_relab}}, \code{\link{LeChatelierE_2013.pathcoverage}}
}

