% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeChatelierE_2013.pathabundance_relab.R
\name{LeChatelierE_2013.pathabundance_relab}
\alias{LeChatelierE_2013.pathabundance_relab}
\alias{LeChatelierE_2013.pathabundance_relab.stool}
\title{Relabeled pathabundance data from the LeChatelierE_2013 dataset}
\source{
INRA, Institut National de la Recherche Agronomique, US1367 Metagenopolis, 78350 Jouy en Josas, France.
}
\description{
Relabeled pathabundance data from the LeChatelierE_2013 dataset
}
\section{Datasets}{
 

\subsection{LeChatelierE_2013.pathabundance_relab.stool}{
   An ExpressionSet with 292 samples and 13,504 features specific to the stool bodysite
}
}
\examples{
LeChatelierE_2013.pathabundance_relab.stool()

}
\seealso{
\code{\link{LeChatelierE_2013.genefamilies_relab}}, \code{\link{LeChatelierE_2013.marker_abundance}}, \code{\link{LeChatelierE_2013.marker_presence}}, \code{\link{LeChatelierE_2013.metaphlan_bugs_list}}, \code{\link{LeChatelierE_2013.pathcoverage}}
}

