% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiJ_2017.R
\name{LiJ_2017}
\alias{LiJ_2017}
\alias{LiJ_2017.genefamilies_relab.stool}
\alias{LiJ_2017.marker_abundance.stool}
\alias{LiJ_2017.marker_presence.stool}
\alias{LiJ_2017.metaphlan_bugs_list.stool}
\alias{LiJ_2017.pathabundance_relab.stool}
\alias{LiJ_2017.pathcoverage.stool}
\title{Data from the LiJ_2017 study}
\description{
Data from the LiJ_2017 study
}
\section{Datasets}{
 

\subsection{LiJ_2017.genefamilies_relab.stool}{
   An ExpressionSet with 196 samples and 1,393,787 features specific to the stool body site
}

\subsection{LiJ_2017.marker_abundance.stool}{
   An ExpressionSet with 196 samples and 88,651 features specific to the stool body site
}

\subsection{LiJ_2017.marker_presence.stool}{
   An ExpressionSet with 196 samples and 82,679 features specific to the stool body site
}

\subsection{LiJ_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 196 samples and 1,150 features specific to the stool body site
}

\subsection{LiJ_2017.pathabundance_relab.stool}{
   An ExpressionSet with 196 samples and 11,776 features specific to the stool body site
}

\subsection{LiJ_2017.pathcoverage.stool}{
   An ExpressionSet with 196 samples and 11,776 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Gut microbiota dysbiosis contributes to the development of hypertension.
}

\subsection{Author}{
   Li J, Zhao F, Wang Y, Chen J, Tao J, Tian G, Wu S, Liu W, Cui Q, Geng B, Zhang W, Weldon R, Auguste K, Yang L, Liu X, Chen L, Yang X, Zhu B, Cai J
}

\subsection{Lab}{
   [1] Hypertension Center, Fuwai Hospital, State Key Laboratory of Cardiovascular Disease of China, National Center for Cardiovascular Diseases of China, Chinese Academy of Medical Sciences and Peking Union Medical College, Beijing, 100037, China., [2] Department of Cardiology, Beijing ChaoYang Hospital, Capital Medical University, Beijing, 100020, China., [3] Beijing Key Laboratory of Hypertension, Beijing, 100020, China., [4] Computational Genomics Laboratory, Beijing Institutes of Life Science, Chinese Academy of Sciences, Beijing, 100101, China., [5] Novogene Bioinformatics Institute, Beijing, 100000, China., [6] Department of Cardiology, Baoding NO.1 Central Hospital, Baoding, 071000, China., [7] Department of Cardiology, The First Affiliated Hospital, Xi'an Jiaotong University, Xi'an, 710061, China., [8] Department of Cardiology Kailuan General Hospital, Hebei Union University, Tangshan, 063000, China., [9] Department of Biomedical Informatics, Centre for Noncoding RNA Medicine, School of Basic Medical Sciences, Peking University, Beijing, 100191, China., [10] Department of Biology and Biochemistry, University of Houston, Houston, TX, 77204, USA., [11] Medical Research Center, Beijing ChaoYang Hospital, Capital Medical University, Beijing, 100020, China., [12] Department of Stem Cell Engineering, Texas Heart Institute, Houston, TX, 77030, USA., [13] Tongji Hospital, Huazhong University of Science and Technology, Wuhan, Hubei, 430030, China., [14] Department of Cardiology, Beijing ChaoYang Hospital, Capital Medical University, Beijing, 100020, China. yxc6229@sina.com., [15] Beijing Key Laboratory of Hypertension, Beijing, 100020, China. yxc6229@sina.com., [16] CAS Key Laboratory of Pathogenic Microbiology and Immunology, Institute of Microbiology, Chinese Academy of Sciences, Beijing, 100101, China. zhubaoli@im.ac.cn., [17] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, The First Affiliated Hospital, College of Medicine, Zhejiang University, Hangzhou, 310003, China. zhubaoli@im.ac.cn., [18] Hypertension Center, Fuwai Hospital, State Key Laboratory of Cardiovascular Disease of China, National Center for Cardiovascular Diseases of China, Chinese Academy of Medical Sciences and Peking Union Medical College, Beijing, 100037, China. caijun@fuwaihospital.org.
}

\subsection{PMID}{
   28143587
}
}

\examples{
LiJ_2017.metaphlan_bugs_list.stool()

}
