% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiSS_2016.R
\name{LiSS_2016}
\alias{LiSS_2016}
\alias{LiSS_2016.genefamilies_relab.stool}
\alias{LiSS_2016.marker_abundance.stool}
\alias{LiSS_2016.marker_presence.stool}
\alias{LiSS_2016.metaphlan_bugs_list.stool}
\alias{LiSS_2016.pathabundance_relab.stool}
\alias{LiSS_2016.pathcoverage.stool}
\title{Data from the LiSS_2016 study}
\description{
Data from the LiSS_2016 study
}
\section{Datasets}{
 

\subsection{LiSS_2016.genefamilies_relab.stool}{
   An ExpressionSet with 55 samples and 974,539 features specific to the stool body site
}

\subsection{LiSS_2016.marker_abundance.stool}{
   An ExpressionSet with 55 samples and 72,504 features specific to the stool body site
}

\subsection{LiSS_2016.marker_presence.stool}{
   An ExpressionSet with 55 samples and 69,172 features specific to the stool body site
}

\subsection{LiSS_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 55 samples and 854 features specific to the stool body site
}

\subsection{LiSS_2016.pathabundance_relab.stool}{
   An ExpressionSet with 55 samples and 8,888 features specific to the stool body site
}

\subsection{LiSS_2016.pathcoverage.stool}{
   An ExpressionSet with 55 samples and 8,888 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Durable coexistence of donor and recipient strains after fecal microbiota transplantation.
}

\subsection{Author}{
   Li SS, Zhu A, Benes V, Costea PI, Hercog R, Hildebrand F, Huerta-Cepas J, Nieuwdorp M, Salojärvi J, Voigt AY, Zeller G, Sunagawa S, de Vos WM, Bork P
}

\subsection{Lab}{
   [1] Structural and Computational Biology Unit, European Molecular Biology Laboratory, 69117 Heidelberg, Germany. School of Biotechnology and Biomolecular Sciences, University of New South Wales, 2052 Sydney, Australia., [2] Structural and Computational Biology Unit, European Molecular Biology Laboratory, 69117 Heidelberg, Germany., [3] Genomics Core Facility, European Molecular Biology Laboratory, 69117 Heidelberg, Germany., [4] Department of Vascular Medicine, Academic Medical Center, 1105 AZ Amsterdam, Netherlands. Diabetes Center, Vrije University Medical Center, 1018 HV Amsterdam, Netherlands. Wallenberg Laboratory, University of Gothenburg, 41345 Gothenburg, Sweden., [5] Department of Veterinary Biosciences, University of Helsinki, 00014 Helsinki, Finland. Department of Biosciences, University of Helsinki, 00014 Helsinki, Finland., [6] Structural and Computational Biology Unit, European Molecular Biology Laboratory, 69117 Heidelberg, Germany. Department of Applied Tumor Biology, Institute of Pathology, University Hospital Heidelberg, 69120 Heidelberg, Germany. Molecular Medicine Partnership Unit, University of Heidelberg and European Molecular Biology Laboratory, 69120 Heidelberg, Germany., [7] Structural and Computational Biology Unit, European Molecular Biology Laboratory, 69117 Heidelberg, Germany. bork@embl.de willem.devos@wur.nl sunagawa@embl.de., [8] Department of Veterinary Biosciences, University of Helsinki, 00014 Helsinki, Finland. Laboratory of Microbiology, Wageningen University, 6703 HB Wageningen, Netherlands. Immunobiology Research Program, Department of Bacteriology and Immunology, University of Helsinki, 00014 Helsinki, Finland. bork@embl.de willem.devos@wur.nl sunagawa@embl.de., [9] Structural and Computational Biology Unit, European Molecular Biology Laboratory, 69117 Heidelberg, Germany. Molecular Medicine Partnership Unit, University of Heidelberg and European Molecular Biology Laboratory, 69120 Heidelberg, Germany. Max Delbruck Centre for Molecular Medicine, 13125 Berlin, Germany. Department of Bioinformatics, Biocenter, University of Wurzburg, 97074 Wurzburg, Germany. bork@embl.de willem.devos@wur.nl sunagawa@embl.de.
}

\subsection{PMID}{
   27126044
}
}

\examples{
LiSS_2016.metaphlan_bugs_list.stool()

}
