% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiuW_2016.R
\name{LiuW_2016}
\alias{LiuW_2016}
\alias{LiuW_2016.genefamilies_relab.stool}
\alias{LiuW_2016.marker_abundance.stool}
\alias{LiuW_2016.marker_presence.stool}
\alias{LiuW_2016.metaphlan_bugs_list.stool}
\alias{LiuW_2016.pathabundance_relab.stool}
\alias{LiuW_2016.pathcoverage.stool}
\title{Data from the LiuW_2016 study}
\description{
Data from the LiuW_2016 study
}
\section{Datasets}{
 

\subsection{LiuW_2016.genefamilies_relab.stool}{
   An ExpressionSet with 110 samples and 1,178,616 features specific to the stool body site
}

\subsection{LiuW_2016.marker_abundance.stool}{
   An ExpressionSet with 110 samples and 81,028 features specific to the stool body site
}

\subsection{LiuW_2016.marker_presence.stool}{
   An ExpressionSet with 110 samples and 76,593 features specific to the stool body site
}

\subsection{LiuW_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 110 samples and 1,078 features specific to the stool body site
}

\subsection{LiuW_2016.pathabundance_relab.stool}{
   An ExpressionSet with 110 samples and 12,647 features specific to the stool body site
}

\subsection{LiuW_2016.pathcoverage.stool}{
   An ExpressionSet with 110 samples and 12,647 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Unique Features of Ethnic Mongolian Gut Microbiome revealed by metagenomic analysis.
}

\subsection{Author}{
   Liu W, Zhang J, Wu C, Cai S, Huang W, Chen J, Xi X, Liang Z, Hou Q, Zhou B, Qin N, Zhang H
}

\subsection{Lab}{
   [1] Key Laboratory of Dairy Biotechnology and Engineering, Education Ministry of P. R. China, Department of Food Science and Engineering, Inner Mongolia Agricultural University, Hohhot 010018, China., [2] RealBio Genomic Institute, Shanghai 200050, China., [3] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, the First Affiliated Hospital, Zhejiang University, Hangzhou 310003, China.
}

\subsection{PMID}{
   27708392
}
}

\examples{
LiuW_2016.metaphlan_bugs_list.stool()

}
