% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LomanNJ_2013_Hi.metaphlan_bugs_list.R
\name{LomanNJ_2013_Hi.metaphlan_bugs_list}
\alias{LomanNJ_2013_Hi.metaphlan_bugs_list}
\alias{LomanNJ_2013_Hi.metaphlan_bugs_list.stool}
\title{Taxonomic abundance data from the LomanNJ_2013_Hi dataset}
\source{
Institute of Microbiology and Infection, University of Birmingham, Birmingham, England.
}
\description{
Taxonomic abundance data from the LomanNJ_2013_Hi dataset
}
\section{Datasets}{
 

\subsection{LomanNJ_2013_Hi.metaphlan_bugs_list.stool}{
   An ExpressionSet with 44 samples and 736 features specific to the stool bodysite
}
}

\examples{
LomanNJ_2013_Hi.metaphlan_bugs_list.stool()

}
\seealso{
\code{\link{LomanNJ_2013_Hi.genefamilies_relab}}, \code{\link{LomanNJ_2013_Hi.marker_abundance}}, \code{\link{LomanNJ_2013_Hi.marker_presence}}, \code{\link{LomanNJ_2013_Hi.pathabundance_relab}}, \code{\link{LomanNJ_2013_Hi.pathcoverage}}
}
