% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LomanNJ_2013_Hi.pathcoverage.R
\name{LomanNJ_2013_Hi.pathcoverage}
\alias{LomanNJ_2013_Hi.pathcoverage}
\alias{LomanNJ_2013_Hi.pathcoverage.stool}
\title{Pathcoverage data from the LomanNJ_2013_Hi dataset}
\source{
Institute of Microbiology and Infection, University of Birmingham, Birmingham, England.
}
\description{
Pathcoverage data from the LomanNJ_2013_Hi dataset
}
\section{Datasets}{
 

\subsection{LomanNJ_2013_Hi.pathcoverage.stool}{
   An ExpressionSet with 44 samples and 8,716 features specific to the stool bodysite
}
}
\examples{
LomanNJ_2013_Hi.pathcoverage.stool()

}
\seealso{
\code{\link{LomanNJ_2013_Hi.genefamilies_relab}}, \code{\link{LomanNJ_2013_Hi.marker_abundance}}, \code{\link{LomanNJ_2013_Hi.marker_presence}}, \code{\link{LomanNJ_2013_Hi.metaphlan_bugs_list}}, \code{\link{LomanNJ_2013_Hi.pathabundance_relab}}
}

