% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LomanNJ_2013_Mi.genefamilies_relab.R
\name{LomanNJ_2013_Mi.genefamilies_relab}
\alias{LomanNJ_2013_Mi.genefamilies_relab}
\alias{LomanNJ_2013_Mi.genefamilies_relab.stool}
\title{Relabeled genefamilies data from the LomanNJ_2013_Mi dataset}
\source{
Institute of Microbiology and Infection, University of Birmingham, Birmingham, England.
}
\description{
Relabeled genefamilies data from the LomanNJ_2013_Mi dataset
}
\section{Datasets}{
 

\subsection{LomanNJ_2013_Mi.genefamilies_relab.stool}{
   An ExpressionSet with 9 samples and 337,637 features specific to the stool bodysite
}
}
\seealso{
\code{\link{LomanNJ_2013_Mi.marker_abundance}}, \code{\link{LomanNJ_2013_Mi.marker_presence}}, \code{\link{LomanNJ_2013_Mi.metaphlan_bugs_list}}, \code{\link{LomanNJ_2013_Mi.pathabundance_relab}}, \code{\link{LomanNJ_2013_Mi.pathcoverage}}
}

