% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LomanNJ_2013_Mi.marker_presence.R
\name{LomanNJ_2013_Mi.marker_presence}
\alias{LomanNJ_2013_Mi.marker_presence}
\alias{LomanNJ_2013_Mi.marker_presence.stool}
\title{Marker presence data from the LomanNJ_2013_Mi dataset}
\source{
Institute of Microbiology and Infection, University of Birmingham, Birmingham, England.
}
\description{
Marker presence data from the LomanNJ_2013_Mi dataset
}
\section{Datasets}{


\subsection{LomanNJ_2013_Mi.marker_presence.stool}{
   An ExpressionSet with 9 samples and 22,710 features specific to the stool bodysite
}
}
\examples{
query(ExperimentHub(), "curatedMetagenomicData")

}
\seealso{
\code{\link{LomanNJ_2013_Mi.genefamilies_relab}}, \code{\link{LomanNJ_2013_Mi.marker_abundance}}, \code{\link{LomanNJ_2013_Mi.metaphlan_bugs_list}}, \code{\link{LomanNJ_2013_Mi.pathabundance_relab}}, \code{\link{LomanNJ_2013_Mi.pathcoverage}}
}

