% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LomanNJ_2013_Mi.pathcoverage.R
\name{LomanNJ_2013_Mi.pathcoverage}
\alias{LomanNJ_2013_Mi.pathcoverage}
\alias{LomanNJ_2013_Mi.pathcoverage.stool}
\title{Pathcoverage data from the LomanNJ_2013_Mi dataset}
\source{
Institute of Microbiology and Infection, University of Birmingham, Birmingham, England.
}
\description{
Pathcoverage data from the LomanNJ_2013_Mi dataset
}
\section{Datasets}{


\subsection{LomanNJ_2013_Mi.pathcoverage.stool}{
   An ExpressionSet with 9 samples and 4,606 features specific to the stool bodysite
}
}
\examples{
query(ExperimentHub(), "curatedMetagenomicData")

}
\seealso{
\code{\link{LomanNJ_2013_Mi.genefamilies_relab}}, \code{\link{LomanNJ_2013_Mi.marker_abundance}}, \code{\link{LomanNJ_2013_Mi.marker_presence}}, \code{\link{LomanNJ_2013_Mi.metaphlan_bugs_list}}, \code{\link{LomanNJ_2013_Mi.pathabundance_relab}}
}

