% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoombaR_2017.R
\name{LoombaR_2017}
\alias{LoombaR_2017}
\alias{LoombaR_2017.genefamilies_relab.stool}
\alias{LoombaR_2017.marker_abundance.stool}
\alias{LoombaR_2017.marker_presence.stool}
\alias{LoombaR_2017.metaphlan_bugs_list.stool}
\alias{LoombaR_2017.pathabundance_relab.stool}
\alias{LoombaR_2017.pathcoverage.stool}
\title{Data from the LoombaR_2017 study}
\description{
Data from the LoombaR_2017 study
}
\section{Datasets}{
 

\subsection{LoombaR_2017.genefamilies_relab.stool}{
   An ExpressionSet with 86 samples and 1,222,524 features specific to the stool body site
}

\subsection{LoombaR_2017.marker_abundance.stool}{
   An ExpressionSet with 86 samples and 88,649 features specific to the stool body site
}

\subsection{LoombaR_2017.marker_presence.stool}{
   An ExpressionSet with 86 samples and 85,505 features specific to the stool body site
}

\subsection{LoombaR_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 86 samples and 1,034 features specific to the stool body site
}

\subsection{LoombaR_2017.pathabundance_relab.stool}{
   An ExpressionSet with 86 samples and 12,913 features specific to the stool body site
}

\subsection{LoombaR_2017.pathcoverage.stool}{
   An ExpressionSet with 86 samples and 12,913 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Gut Microbiome-Based Metagenomic Signature for Non-invasive Detection of Advanced Fibrosis in Human Nonalcoholic Fatty Liver Disease.
}

\subsection{Author}{
   Loomba R, Seguritan V, Li W, Long T, Klitgord N, Bhatt A, Dulai PS, Caussy C, Bettencourt R, Highlander SK, Jones MB, Sirlin CB, Schnabl B, Brinkac L, Schork N, Chen CH, Brenner DA, Biggs W, Yooseph S, Venter JC, Nelson KE
}

\subsection{Lab}{
   [1] NAFLD Research Center, Department of Medicine, University of California, San Diego, La Jolla, CA 92093, USA; Division of Epidemiology, Department of Family and Preventive Medicine, University of California, San Diego, La Jolla, CA 92093, USA; Division of Gastroenterology, Department of Medicine, University of California, San Diego, La Jolla, CA 92093, USA. Electronic address: roloomba@ucsd.edu., [2] Human Longevity, San Diego, CA 92121, USA., [3] Human Longevity, San Diego, CA 92121, USA; J. Craig Venter Institute, La Jolla, CA 92037, USA., [4] NAFLD Research Center, Department of Medicine, University of California, San Diego, La Jolla, CA 92093, USA., [5] NAFLD Research Center, Department of Medicine, University of California, San Diego, La Jolla, CA 92093, USA; Division of Gastroenterology, Department of Medicine, University of California, San Diego, La Jolla, CA 92093, USA., [6] J. Craig Venter Institute, La Jolla, CA 92037, USA., [7] Liver Imaging Group, Department of Radiology, University of California, San Diego, La Jolla, CA 92093, USA., [8] J. Craig Venter Institute, Rockville, MD 20850, USA.
}

\subsection{PMID}{
   28467925
}
}

\examples{
`LoombaR_2017.metaphlan_bugs_list.stool`()

}
