% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LouisS_2016.R
\name{LouisS_2016}
\alias{LouisS_2016}
\alias{LouisS_2016.genefamilies_relab.stool}
\alias{LouisS_2016.marker_abundance.stool}
\alias{LouisS_2016.marker_presence.stool}
\alias{LouisS_2016.metaphlan_bugs_list.stool}
\alias{LouisS_2016.pathabundance_relab.stool}
\alias{LouisS_2016.pathcoverage.stool}
\title{Data from the LouisS_2016 study}
\description{
Data from the LouisS_2016 study
}
\section{Datasets}{
 

\subsection{LouisS_2016.genefamilies_relab.stool}{
   An ExpressionSet with 92 samples and 814,252 features specific to the stool body site
}

\subsection{LouisS_2016.marker_abundance.stool}{
   An ExpressionSet with 92 samples and 53,320 features specific to the stool body site
}

\subsection{LouisS_2016.marker_presence.stool}{
   An ExpressionSet with 92 samples and 51,680 features specific to the stool body site
}

\subsection{LouisS_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 92 samples and 657 features specific to the stool body site
}

\subsection{LouisS_2016.pathabundance_relab.stool}{
   An ExpressionSet with 92 samples and 7,212 features specific to the stool body site
}

\subsection{LouisS_2016.pathcoverage.stool}{
   An ExpressionSet with 92 samples and 7,212 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Characterization of the Gut Microbial Community of Obese Patients Following a Weight-Loss Intervention Using Whole Metagenome Shotgun Sequencing.
}

\subsection{Author}{
   Louis S, Tappu RM, Damms-Machado A, Huson DH, Bischoff SC
}

\subsection{Lab}{
   [1] Institute of Clinical Nutrition, University of Hohenheim, Stuttgart, Germany., [2] Algorithms in Bioinformatics, University of Tubingen, Tubingen, Germany.
}

\subsection{PMID}{
   26919743
}
}

\examples{
`LouisS_2016.metaphlan_bugs_list.stool`()

}
