% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{revise make_data program to behave in a conventional way}
\usage{
make_data(tar_gz_file, tmpdir = cmg_tmpdir())
}
\arguments{
\item{tar_gz_file}{character(1)}

\item{tmpdir}{character(1) path to folder where tgz file is extracted}
}
\description{
revise make_data program to behave in a conventional way
}
