% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatsonV_2018.R
\name{MatsonV_2018}
\alias{MatsonV_2018}
\alias{MatsonV_2018.genefamilies_relab.stool}
\alias{MatsonV_2018.marker_abundance.stool}
\alias{MatsonV_2018.marker_presence.stool}
\alias{MatsonV_2018.metaphlan_bugs_list.stool}
\alias{MatsonV_2018.pathabundance_relab.stool}
\alias{MatsonV_2018.pathcoverage.stool}
\title{Data from the MatsonV_2018 study}
\description{
Data from the MatsonV_2018 study
}
\section{Datasets}{
 

\subsection{MatsonV_2018.genefamilies_relab.stool}{
   An ExpressionSet with 39 samples and 1,198,303 features specific to the stool body site
}

\subsection{MatsonV_2018.marker_abundance.stool}{
   An ExpressionSet with 39 samples and 75,666 features specific to the stool body site
}

\subsection{MatsonV_2018.marker_presence.stool}{
   An ExpressionSet with 39 samples and 73,435 features specific to the stool body site
}

\subsection{MatsonV_2018.metaphlan_bugs_list.stool}{
   An ExpressionSet with 39 samples and 921 features specific to the stool body site
}

\subsection{MatsonV_2018.pathabundance_relab.stool}{
   An ExpressionSet with 39 samples and 7,058 features specific to the stool body site
}

\subsection{MatsonV_2018.pathcoverage.stool}{
   An ExpressionSet with 39 samples and 7,058 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   The commensal microbiome is associated with anti-PD-1 efficacy in metastatic melanoma patients.
}

\subsection{Author}{
   Matson V, Fessler J, Bao R, Chongsuwat T, Zha Y, Alegre ML, Luke JJ, Gajewski TF
}

\subsection{Lab}{
   [1] Department of Pathology, University of Chicago, Chicago, IL 60637, USA., [2] Center for Research Informatics, University of Chicago, IL 60637, USA., [3] Department of Pediatrics, University of Chicago, IL 60637, USA., [4] Department of Medicine, University of Chicago, Chicago, IL 60637, USA.
}

\subsection{PMID}{
   29302014
}
}

\examples{
MatsonV_2018.metaphlan_bugs_list.stool()

}
