% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Manually Curated Metadata for All Studies in curatedMetagenomicData}
\format{
An object of class \code{data.frame} with 20494 rows and 132 columns.
}
\usage{
metadata
}
\description{
The combined sample level metadata of all studies in curatedMetagenomicData.
}
\keyword{datasets}
