% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NielsenHB_2014.pathcoverage.R
\name{NielsenHB_2014.pathcoverage}
\alias{NielsenHB_2014.pathcoverage}
\alias{NielsenHB_2014.pathcoverage.stool}
\title{Pathcoverage data from the NielsenHB_2014 dataset}
\source{
1] Center for Biological Sequence Analysis, Technical University of Denmark, Kongens Lyngby, Denmark. [2] Novo Nordisk Foundation Center for Biosustainability, Technical University of Denmark, Kongens Lyngby, Denmark. [3].
}
\description{
Pathcoverage data from the NielsenHB_2014 dataset
}
\section{Datasets}{
 

\subsection{NielsenHB_2014.pathcoverage.stool}{
   An ExpressionSet with 396 samples and 17,280 features specific to the stool bodysite
}
}
\examples{
NielsenHB_2014.pathcoverage.stool()

}
\seealso{
\code{\link{NielsenHB_2014.genefamilies_relab}}, \code{\link{NielsenHB_2014.marker_abundance}}, \code{\link{NielsenHB_2014.marker_presence}}, \code{\link{NielsenHB_2014.metaphlan_bugs_list}}, \code{\link{NielsenHB_2014.pathabundance_relab}}
}

