% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obregon-TitoAJ_2015.genefamilies_relab.R
\name{Obregon-TitoAJ_2015.genefamilies_relab}
\alias{Obregon-TitoAJ_2015.genefamilies_relab}
\alias{Obregon-TitoAJ_2015.genefamilies_relab.stool}
\title{Relabeled genefamilies data from the Obregon-TitoAJ_2015 dataset}
\source{
1] Department of Anthropology, University of Oklahoma, Dale Hall Tower, 521 Norman, Oklahoma 73019, USA [2] Universidad Cient??fica del Sur, Lima 18, Per?? [3] City of Hope, NCI-designated Comprehensive Cancer Center, Duarte, California 91010, USA.
}
\description{
Relabeled genefamilies data from the Obregon-TitoAJ_2015 dataset
}
\section{Datasets}{


\subsection{Obregon-TitoAJ_2015.genefamilies_relab.stool}{
   An ExpressionSet with 58 samples and 1,192,382 features specific to the stool bodysite
}
}
\seealso{
\code{\link{Obregon-TitoAJ_2015.marker_abundance}}, \code{\link{Obregon-TitoAJ_2015.marker_presence}}, \code{\link{Obregon-TitoAJ_2015.metaphlan_bugs_list}}, \code{\link{Obregon-TitoAJ_2015.pathabundance_relab}}, \code{\link{Obregon-TitoAJ_2015.pathcoverage}}
}

