% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obregon-TitoAJ_2015.R
\name{Obregon-TitoAJ_2015}
\alias{Obregon-TitoAJ_2015}
\alias{Obregon-TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon-TitoAJ_2015.marker_abundance.stool}
\alias{Obregon-TitoAJ_2015.marker_presence.stool}
\alias{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon-TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon-TitoAJ_2015.pathcoverage.stool}
\title{Data from the Obregon-TitoAJ_2015 study}
\description{
Data from the Obregon-TitoAJ_2015 study
}
\section{Datasets}{
 

\subsection{Obregon-TitoAJ_2015.genefamilies_relab.stool}{
   An ExpressionSet with 58 samples and 1,185,621 features specific to the stool body site
}

\subsection{Obregon-TitoAJ_2015.marker_abundance.stool}{
   An ExpressionSet with 58 samples and 96,336 features specific to the stool body site
}

\subsection{Obregon-TitoAJ_2015.marker_presence.stool}{
   An ExpressionSet with 58 samples and 86,352 features specific to the stool body site
}

\subsection{Obregon-TitoAJ_2015.metaphlan_bugs_list.stool}{
   An ExpressionSet with 58 samples and 1,094 features specific to the stool body site
}

\subsection{Obregon-TitoAJ_2015.pathabundance_relab.stool}{
   An ExpressionSet with 58 samples and 9,801 features specific to the stool body site
}

\subsection{Obregon-TitoAJ_2015.pathcoverage.stool}{
   An ExpressionSet with 58 samples and 9,801 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Subsistence strategies in traditional societies distinguish gut microbiomes.
}

\subsection{Author}{
   Obregon-Tito AJ, Tito RY, Metcalf J, Sankaranarayanan K, Clemente JC, Ursell LK, Zech Xu Z, Van Treuren W, Knight R, Gaffney PM, Spicer P, Lawson P, Marin-Reyes L, Trujillo-Villarroel O, Foster M, Guija-Poma E, Troncoso-Corzo L, Warinner C, Ozga AT, Lewis CM
}

\subsection{Lab}{
   [1] 1] Department of Anthropology, University of Oklahoma, Dale Hall Tower, 521 Norman, Oklahoma 73019, USA [2] Universidad Cientifica del Sur, Lima 18, Peru [3] City of Hope, NCI-designated Comprehensive Cancer Center, Duarte, California 91010, USA., [2] 1] Department of Anthropology, University of Oklahoma, Dale Hall Tower, 521 Norman, Oklahoma 73019, USA [2] Universidad Cientifica del Sur, Lima 18, Peru, [3] City of Hope, NCI-designated Comprehensive Cancer Center, Duarte, California 91010, USA., [4] Department of Anthropology, University of Oklahoma, Dale Hall Tower, 521 Norman, Oklahoma 73019, USA., [5] Department of Chemistry and Biochemistry, University of Colorado, Boulder, Colorado 80309, USA., [6] Departments of Pediatrics and Computer Science &amp;Engineering University of California San Diego, La Jolla, CA 92093, USA., [7] Oklahoma Medical Research Foundation, Oklahoma City, Oklahoma 73104, USA., [8] Instituto Nacional de Salud, Lima 11, Peru, [9] Old Dominion University, Norfolk, Virginia 23529, USA., [10] Universidad Cientifica del Sur, Lima 18, Peru
}

\subsection{PMID}{
   25807110
}
}

\examples{
`Obregon-TitoAJ_2015.metaphlan_bugs_list.stool`()

}
