% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Obregon_TitoAJ_2015.pathcoverage.R
\name{Obregon_TitoAJ_2015.pathcoverage}
\alias{Obregon_TitoAJ_2015.pathcoverage}
\alias{Obregon_TitoAJ_2015.pathcoverage.stool}
\title{Pathcoverage data from the Obregon_TitoAJ_2015 dataset}
\source{
[1] Department of Anthropology, University of Oklahoma, Dale Hall Tower, 521 Norman, Oklahoma 73019, USA [2] Universidad Cientifica del Sur, Lima 18, Peru [3] City of Hope, NCI-designated Comprehensive Cancer Center, Duarte, California 91010, USA.
}
\description{
Pathcoverage data from the Obregon_TitoAJ_2015 dataset
}
\section{Datasets}{
 

\subsection{Obregon_TitoAJ_2015.pathcoverage.stool}{
   An ExpressionSet with 58 samples and 10,438 features specific to the stool bodysite
}
}

\examples{
Obregon_TitoAJ_2015.pathcoverage.stool()

}
\seealso{
\code{\link{Obregon_TitoAJ_2015.genefamilies_relab}}, \code{\link{Obregon_TitoAJ_2015.marker_abundance}}, \code{\link{Obregon_TitoAJ_2015.marker_presence}}, \code{\link{Obregon_TitoAJ_2015.metaphlan_bugs_list}}, \code{\link{Obregon_TitoAJ_2015.pathabundance_relab}}
}
