% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhJ_2014.marker_abundance.R
\name{OhJ_2014.marker_abundance}
\alias{OhJ_2014.marker_abundance}
\alias{OhJ_2014.marker_abundance.skin}
\title{Marker abundance data from the OhJ_2014 dataset}
\source{
Translational and Functional Genomics Branch, National Human Genome Research Institute, NIH, Bethesda, Maryland 20892, USA.
}
\description{
Marker abundance data from the OhJ_2014 dataset
}
\section{Datasets}{
 

\subsection{OhJ_2014.marker_abundance.skin}{
   An ExpressionSet with 291 samples and 202,658 features specific to the skin bodysite
}
}

\examples{
OhJ_2014.marker_abundance.skin()

}
\seealso{
\code{\link{OhJ_2014.genefamilies_relab}}, \code{\link{OhJ_2014.marker_presence}}, \code{\link{OhJ_2014.metaphlan_bugs_list}}, \code{\link{OhJ_2014.pathabundance_relab}}, \code{\link{OhJ_2014.pathcoverage}}
}
