% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhJ_2014.marker_presence.R
\name{OhJ_2014.marker_presence}
\alias{OhJ_2014.marker_presence}
\alias{OhJ_2014.marker_presence.skin}
\title{Marker presence data from the OhJ_2014 dataset}
\source{
Translational and Functional Genomics Branch, National Human Genome Research Institute, NIH, Bethesda, Maryland 20892, USA.
}
\description{
Marker presence data from the OhJ_2014 dataset
}
\section{Datasets}{
 

\subsection{OhJ_2014.marker_presence.skin}{
   An ExpressionSet with 291 samples and 184,915 features specific to the skin bodysite
}
}

\examples{
OhJ_2014.marker_presence.skin()

}
\seealso{
\code{\link{OhJ_2014.genefamilies_relab}}, \code{\link{OhJ_2014.marker_abundance}}, \code{\link{OhJ_2014.metaphlan_bugs_list}}, \code{\link{OhJ_2014.pathabundance_relab}}, \code{\link{OhJ_2014.pathcoverage}}
}
