% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhJ_2014.pathcoverage.R
\name{OhJ_2014.pathcoverage}
\alias{OhJ_2014.pathcoverage}
\alias{OhJ_2014.pathcoverage.skin}
\title{Pathcoverage data from the OhJ_2014 dataset}
\source{
Translational and Functional Genomics Branch, National Human Genome Research Institute, NIH, Bethesda, Maryland 20892, USA.
}
\description{
Pathcoverage data from the OhJ_2014 dataset
}
\section{Datasets}{
 

\subsection{OhJ_2014.pathcoverage.skin}{
   An ExpressionSet with 291 samples and 48,536 features specific to the skin bodysite
}
}

\examples{
OhJ_2014.pathcoverage.skin()

}
\seealso{
\code{\link{OhJ_2014.genefamilies_relab}}, \code{\link{OhJ_2014.marker_abundance}}, \code{\link{OhJ_2014.marker_presence}}, \code{\link{OhJ_2014.metaphlan_bugs_list}}, \code{\link{OhJ_2014.pathabundance_relab}}
}
