% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhJ_2014.R
\name{OhJ_2014}
\alias{OhJ_2014}
\alias{OhJ_2014.genefamilies_relab.skin}
\alias{OhJ_2014.marker_abundance.skin}
\alias{OhJ_2014.marker_presence.skin}
\alias{OhJ_2014.metaphlan_bugs_list.skin}
\alias{OhJ_2014.pathabundance_relab.skin}
\alias{OhJ_2014.pathcoverage.skin}
\title{Data from the OhJ_2014 study}
\description{
Data from the OhJ_2014 study
}
\section{Datasets}{
 

\subsection{OhJ_2014.genefamilies_relab.skin}{
   An ExpressionSet with 291 samples and 3,956,472 features specific to the skin body site
}

\subsection{OhJ_2014.marker_abundance.skin}{
   An ExpressionSet with 291 samples and 202,657 features specific to the skin body site
}

\subsection{OhJ_2014.marker_presence.skin}{
   An ExpressionSet with 291 samples and 184,914 features specific to the skin body site
}

\subsection{OhJ_2014.metaphlan_bugs_list.skin}{
   An ExpressionSet with 291 samples and 2,461 features specific to the skin body site
}

\subsection{OhJ_2014.pathabundance_relab.skin}{
   An ExpressionSet with 291 samples and 48,536 features specific to the skin body site
}

\subsection{OhJ_2014.pathcoverage.skin}{
   An ExpressionSet with 291 samples and 48,536 features specific to the skin body site
}
}

\section{Source}{
 

\subsection{Title}{
   Biogeography and individuality shape function in the human skin metagenome.
}

\subsection{Author}{
   Oh J, Byrd AL, Deming C, Conlan S, Kong HH, Segre JA, Barnabas B, Blakesley R, Bouffard G, Brooks S, Coleman H, Dekhtyar M, Gregory M, Guan X, Gupta J, Han J, Ho SL, Legaspi R, Maduro Q, Masiello C, Maskeri B, McDowell J, Montemayor C, Mullikin J, Park M, Riebow N, Schandler K, Schmidt B, Sison C, Stantripop M, Thomas J, Thomas P, Vemulapalli M, Young A
}

\subsection{Lab}{
   [1] Translational and Functional Genomics Branch, National Human Genome Research Institute, NIH, Bethesda, Maryland 20892, USA., [2] 1] Dermatology Branch, Center for Cancer Research, National Cancer Institute, NIH, Bethesda, Maryland 20892, USA [2]., [3] 1] Translational and Functional Genomics Branch, National Human Genome Research Institute, NIH, Bethesda, Maryland 20892, USA [2].
}

\subsection{PMID}{
   25279917
}
}

\examples{
`OhJ_2014.metaphlan_bugs_list.skin`()

}
