% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlmMR_2017.R
\name{OlmMR_2017}
\alias{OlmMR_2017}
\alias{OlmMR_2017.genefamilies_relab.oralcavity}
\alias{OlmMR_2017.genefamilies_relab.skin}
\alias{OlmMR_2017.genefamilies_relab.stool}
\alias{OlmMR_2017.marker_abundance.oralcavity}
\alias{OlmMR_2017.marker_abundance.skin}
\alias{OlmMR_2017.marker_abundance.stool}
\alias{OlmMR_2017.marker_presence.oralcavity}
\alias{OlmMR_2017.marker_presence.skin}
\alias{OlmMR_2017.marker_presence.stool}
\alias{OlmMR_2017.metaphlan_bugs_list.oralcavity}
\alias{OlmMR_2017.metaphlan_bugs_list.skin}
\alias{OlmMR_2017.metaphlan_bugs_list.stool}
\alias{OlmMR_2017.pathabundance_relab.oralcavity}
\alias{OlmMR_2017.pathabundance_relab.skin}
\alias{OlmMR_2017.pathabundance_relab.stool}
\alias{OlmMR_2017.pathcoverage.oralcavity}
\alias{OlmMR_2017.pathcoverage.skin}
\alias{OlmMR_2017.pathcoverage.stool}
\title{Data from the OlmMR_2017 study}
\description{
Data from the OlmMR_2017 study
}
\section{Datasets}{
 

\subsection{OlmMR_2017.genefamilies_relab.oralcavity}{
   An ExpressionSet with 4 samples and 138,859 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.genefamilies_relab.skin}{
   An ExpressionSet with 4 samples and 138,859 features specific to the skin body site
}

\subsection{OlmMR_2017.genefamilies_relab.stool}{
   An ExpressionSet with 37 samples and 138,859 features specific to the stool body site
}

\subsection{OlmMR_2017.marker_abundance.oralcavity}{
   An ExpressionSet with 4 samples and 10,965 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.marker_abundance.skin}{
   An ExpressionSet with 4 samples and 10,965 features specific to the skin body site
}

\subsection{OlmMR_2017.marker_abundance.stool}{
   An ExpressionSet with 37 samples and 10,965 features specific to the stool body site
}

\subsection{OlmMR_2017.marker_presence.oralcavity}{
   An ExpressionSet with 4 samples and 9,977 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.marker_presence.skin}{
   An ExpressionSet with 4 samples and 9,977 features specific to the skin body site
}

\subsection{OlmMR_2017.marker_presence.stool}{
   An ExpressionSet with 37 samples and 9,977 features specific to the stool body site
}

\subsection{OlmMR_2017.metaphlan_bugs_list.oralcavity}{
   An ExpressionSet with 4 samples and 201 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.metaphlan_bugs_list.skin}{
   An ExpressionSet with 4 samples and 201 features specific to the skin body site
}

\subsection{OlmMR_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 37 samples and 201 features specific to the stool body site
}

\subsection{OlmMR_2017.pathabundance_relab.oralcavity}{
   An ExpressionSet with 4 samples and 2,941 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.pathabundance_relab.skin}{
   An ExpressionSet with 4 samples and 2,941 features specific to the skin body site
}

\subsection{OlmMR_2017.pathabundance_relab.stool}{
   An ExpressionSet with 37 samples and 2,941 features specific to the stool body site
}

\subsection{OlmMR_2017.pathcoverage.oralcavity}{
   An ExpressionSet with 4 samples and 2,941 features specific to the oralcavity body site
}

\subsection{OlmMR_2017.pathcoverage.skin}{
   An ExpressionSet with 4 samples and 2,941 features specific to the skin body site
}

\subsection{OlmMR_2017.pathcoverage.stool}{
   An ExpressionSet with 37 samples and 2,941 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Identical bacterial populations colonize premature infant gut, skin, and oral microbiomes and exhibit different in situ growth rates.
}

\subsection{Author}{
   Olm MR, Brown CT, Brooks B, Firek B, Baker R, Burstein D, Soenjoyo K, Thomas BC, Morowitz M, Banfield JF
}

\subsection{Lab}{
   [1] Department of Plant and Microbial Biology, University of California, Berkeley, California 94720, USA., [2] Department of Surgery, University of Pittsburgh School of Medicine, Pittsburgh, Pennsylvania 15213, USA., [3] Division of Newborn Medicine, Children's Hospital of Pittsburgh and Magee-Womens Hospital of UPMC, Pittsburgh, Pennsylvania 15213, USA., [4] Department of Earth and Planetary Science, University of California, Berkeley, California 94709, USA., [5] Department of Environmental Science, Policy, and Management, University of California, Berkeley, California 94720, USA., [6] Earth Sciences Division, Lawrence Berkeley National Laboratory, Berkeley, California 94720, USA.
}

\subsection{PMID}{
   28073918
}
}

\examples{
OlmMR_2017.metaphlan_bugs_list.oralcavity()

}
