% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PehrssonE_2016.R
\name{PehrssonE_2016}
\alias{PehrssonE_2016}
\alias{PehrssonE_2016.genefamilies_relab.stool}
\alias{PehrssonE_2016.marker_abundance.stool}
\alias{PehrssonE_2016.marker_presence.stool}
\alias{PehrssonE_2016.metaphlan_bugs_list.stool}
\alias{PehrssonE_2016.pathabundance_relab.stool}
\alias{PehrssonE_2016.pathcoverage.stool}
\title{Data from the PehrssonE_2016 study}
\description{
Data from the PehrssonE_2016 study
}
\section{Datasets}{
 

\subsection{PehrssonE_2016.genefamilies_relab.stool}{
   An ExpressionSet with 191 samples and 1,197,185 features specific to the stool body site
}

\subsection{PehrssonE_2016.genefamilies_relab.stool}{
   An ExpressionSet with 191 samples and 1,197,185 features specific to the stool body site
}

\subsection{PehrssonE_2016.marker_abundance.stool}{
   An ExpressionSet with 191 samples and 76,834 features specific to the stool body site
}

\subsection{PehrssonE_2016.marker_abundance.stool}{
   An ExpressionSet with 191 samples and 76,834 features specific to the stool body site
}

\subsection{PehrssonE_2016.marker_presence.stool}{
   An ExpressionSet with 191 samples and 71,391 features specific to the stool body site
}

\subsection{PehrssonE_2016.marker_presence.stool}{
   An ExpressionSet with 191 samples and 71,391 features specific to the stool body site
}

\subsection{PehrssonE_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 191 samples and 921 features specific to the stool body site
}

\subsection{PehrssonE_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 191 samples and 921 features specific to the stool body site
}

\subsection{PehrssonE_2016.pathabundance_relab.stool}{
   An ExpressionSet with 191 samples and 7,096 features specific to the stool body site
}

\subsection{PehrssonE_2016.pathabundance_relab.stool}{
   An ExpressionSet with 191 samples and 7,096 features specific to the stool body site
}

\subsection{PehrssonE_2016.pathcoverage.stool}{
   An ExpressionSet with 191 samples and 7,096 features specific to the stool body site
}

\subsection{PehrssonE_2016.pathcoverage.stool}{
   An ExpressionSet with 191 samples and 7,096 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Interconnected microbiomes and resistomes in low-income human habitats.
}

\subsection{Author}{
   Pehrsson EC, Tsukayama P, Patel S, Mejía-Bautista M, Sosa-Soto G, Navarrete KM, Calderon M, Cabrera L, Hoyos-Arango W, Bertoli MT, Berg DE, Gilman RH, Dantas G
}

\subsection{Lab}{
   [1] Center for Genome Sciences and Systems Biology, Washington University School of Medicine, St Louis, Missouri 63110, USA., [2] Department of Pathology and Immunology, Washington University School of Medicine, St Louis, Missouri 63110, USA., [3] Facultad de Ciencias de la Salud &quot;Dr. Luis Edmundo Vasquez&quot;, Universidad Dr. Jose Matias Delgado, El Salvador., [4] Laboratorios de Investigacion y Desarrollo, Universidad Peruana Cayetano Heredia, San Martin de Porres, Lima 31, Peru., [5] Asociacion Benefica PRISMA, San Miguel, Lima 32, Peru., [6] Department of Molecular Microbiology, Washington University School of Medicine, St Louis, Missouri 63110, USA., [7] Department of Medicine, University of California San Diego, La Jolla, California 92093, USA., [8] Department of International Health, Johns Hopkins School of Public Health, Baltimore, Maryland 21205, USA., [9] Department of Biomedical Engineering, Washington University, St Louis, Missouri 63105, USA.
}

\subsection{PMID}{
   27172044
}
}

\examples{
PehrssonE_2016.metaphlan_bugs_list.stool()

}
