% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QinJ_2012.genefamilies_relab.R
\name{QinJ_2012.genefamilies_relab}
\alias{QinJ_2012.genefamilies_relab}
\alias{QinJ_2012.genefamilies_relab.stool}
\title{Relabeled genefamilies data from the QinJ_2012 dataset}
\source{
BGI-Shenzhen, Shenzhen 518083, China.
}
\description{
Relabeled genefamilies data from the QinJ_2012 dataset
}
\section{Datasets}{
 

\subsection{QinJ_2012.genefamilies_relab.stool}{
   An ExpressionSet with 363 samples and 1,690,774 features specific to the stool bodysite
}
}
\seealso{
\code{\link{QinJ_2012.marker_abundance}}, \code{\link{QinJ_2012.marker_presence}}, \code{\link{QinJ_2012.metaphlan_bugs_list}}, \code{\link{QinJ_2012.pathabundance_relab}}, \code{\link{QinJ_2012.pathcoverage}}
}

