% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QinJ_2012.marker_presence.R
\name{QinJ_2012.marker_presence}
\alias{QinJ_2012.marker_presence}
\alias{QinJ_2012.marker_presence.stool}
\title{Marker presence data from the QinJ_2012 dataset}
\source{
BGI-Shenzhen, Shenzhen 518083, China.
}
\description{
Marker presence data from the QinJ_2012 dataset
}
\section{Datasets}{
 

\subsection{QinJ_2012.marker_presence.stool}{
   An ExpressionSet with 363 samples and 125,127 features specific to the stool bodysite
}
}
\examples{
QinJ_2012.marker_presence.stool()

}
\seealso{
\code{\link{QinJ_2012.genefamilies_relab}}, \code{\link{QinJ_2012.marker_abundance}}, \code{\link{QinJ_2012.metaphlan_bugs_list}}, \code{\link{QinJ_2012.pathabundance_relab}}, \code{\link{QinJ_2012.pathcoverage}}
}

