% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QinJ_2012.R
\name{QinJ_2012}
\alias{QinJ_2012}
\alias{QinJ_2012.genefamilies_relab.stool}
\alias{QinJ_2012.marker_abundance.stool}
\alias{QinJ_2012.marker_presence.stool}
\alias{QinJ_2012.metaphlan_bugs_list.stool}
\alias{QinJ_2012.pathabundance_relab.stool}
\alias{QinJ_2012.pathcoverage.stool}
\title{Data from the QinJ_2012 study}
\description{
Data from the QinJ_2012 study
}
\section{Datasets}{
 

\subsection{QinJ_2012.genefamilies_relab.stool}{
   An ExpressionSet with 363 samples and 1,690,773 features specific to the stool body site
}

\subsection{QinJ_2012.marker_abundance.stool}{
   An ExpressionSet with 363 samples and 132,933 features specific to the stool body site
}

\subsection{QinJ_2012.marker_presence.stool}{
   An ExpressionSet with 363 samples and 125,126 features specific to the stool body site
}

\subsection{QinJ_2012.metaphlan_bugs_list.stool}{
   An ExpressionSet with 363 samples and 1,588 features specific to the stool body site
}

\subsection{QinJ_2012.pathabundance_relab.stool}{
   An ExpressionSet with 363 samples and 18,478 features specific to the stool body site
}

\subsection{QinJ_2012.pathcoverage.stool}{
   An ExpressionSet with 363 samples and 18,478 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   A metagenome-wide association study of gut microbiota in type 2 diabetes.
}

\subsection{Author}{
   Qin J, Li Y, Cai Z, Li S, Zhu J, Zhang F, Liang S, Zhang W, Guan Y, Shen D, Peng Y, Zhang D, Jie Z, Wu W, Qin Y, Xue W, Li J, Han L, Lu D, Wu P, Dai Y, Sun X, Li Z, Tang A, Zhong S, Li X, Chen W, Xu R, Wang M, Feng Q, Gong M, Yu J, Zhang Y, Zhang M, Hansen T, Sanchez G, Raes J, Falony G, Okuda S, Almeida M, LeChatelier E, Renault P, Pons N, Batto JM, Zhang Z, Chen H, Yang R, Zheng W, Li S, Yang H, Wang J, Ehrlich SD, Nielsen R, Pedersen O, Kristiansen K, Wang J
}

\subsection{Lab}{
   [1] BGI-Shenzhen, Shenzhen 518083, China.
}

\subsection{PMID}{
   23023125
}
}

\examples{
`QinJ_2012.metaphlan_bugs_list.stool`()

}
