% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QinN_2014.marker_abundance.R
\name{QinN_2014.marker_abundance}
\alias{QinN_2014.marker_abundance}
\alias{QinN_2014.marker_abundance.stool}
\title{Marker abundance data from the QinN_2014 dataset}
\source{
1] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, The First Affiliated Hospital, College of Medicine, Zhejiang University, 310003 Hangzhou, China [2] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, 310003 Hangzhou, China [3].
}
\description{
Marker abundance data from the QinN_2014 dataset
}
\section{Datasets}{
 

\subsection{QinN_2014.marker_abundance.stool}{
   An ExpressionSet with 237 samples and 132,775 features specific to the stool bodysite
}
}
\seealso{
\code{\link{QinN_2014.genefamilies_relab}}, \code{\link{QinN_2014.marker_presence}}, \code{\link{QinN_2014.metaphlan_bugs_list}}, \code{\link{QinN_2014.pathabundance_relab}}, \code{\link{QinN_2014.pathcoverage}}
}

