% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QinN_2014.R
\name{QinN_2014}
\alias{QinN_2014}
\alias{QinN_2014.genefamilies_relab.stool}
\alias{QinN_2014.marker_abundance.stool}
\alias{QinN_2014.marker_presence.stool}
\alias{QinN_2014.metaphlan_bugs_list.stool}
\alias{QinN_2014.pathabundance_relab.stool}
\alias{QinN_2014.pathcoverage.stool}
\title{Data from the QinN_2014 study}
\description{
Data from the QinN_2014 study
}
\section{Datasets}{
 

\subsection{QinN_2014.genefamilies_relab.stool}{
   An ExpressionSet with 237 samples and 1,747,533 features specific to the stool body site
}

\subsection{QinN_2014.marker_abundance.stool}{
   An ExpressionSet with 237 samples and 132,774 features specific to the stool body site
}

\subsection{QinN_2014.marker_presence.stool}{
   An ExpressionSet with 237 samples and 126,096 features specific to the stool body site
}

\subsection{QinN_2014.metaphlan_bugs_list.stool}{
   An ExpressionSet with 237 samples and 1,512 features specific to the stool body site
}

\subsection{QinN_2014.pathabundance_relab.stool}{
   An ExpressionSet with 237 samples and 19,418 features specific to the stool body site
}

\subsection{QinN_2014.pathcoverage.stool}{
   An ExpressionSet with 237 samples and 19,418 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Alterations of the human gut microbiome in liver cirrhosis.
}

\subsection{Author}{
   Qin N, Yang F, Li A, Prifti E, Chen Y, Shao L, Guo J, Le Chatelier E, Yao J, Wu L, Zhou J, Ni S, Liu L, Pons N, Batto JM, Kennedy SP, Leonard P, Yuan C, Ding W, Chen Y, Hu X, Zheng B, Qian G, Xu W, Ehrlich SD, Zheng S, Li L
}

\subsection{Lab}{
   [1] 1] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, The First Affiliated Hospital, College of Medicine, Zhejiang University, 310003 Hangzhou, China [2] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, 310003 Hangzhou, China [3]., [2] 1] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, The First Affiliated Hospital, College of Medicine, Zhejiang University, 310003 Hangzhou, China [2]., [3] 1] Metagenopolis, Institut National de la Recherche Agronomique, 78350 Jouy en Josas, France [2]., [4] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, The First Affiliated Hospital, College of Medicine, Zhejiang University, 310003 Hangzhou, China., [5] Metagenopolis, Institut National de la Recherche Agronomique, 78350 Jouy en Josas, France., [6] 1] State Key Laboratory for Diagnosis and Treatment of Infectious Disease, The First Affiliated Hospital, College of Medicine, Zhejiang University, 310003 Hangzhou, China [2] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, 310003 Hangzhou, China., [7] 1] Metagenopolis, Institut National de la Recherche Agronomique, 78350 Jouy en Josas, France [2] King's College London, Centre for Host-Microbiome Interactions, Dental Institute Central Office, Guy's Hospital, London Bridge, London SE1 9RT, UK., [8] 1] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, 310003 Hangzhou, China [2] Key Laboratory of Combined Multi-organ Transplantation, Ministry of Public Health, the First Affiliated Hospital, Zhejiang University, 310003 Hangzhou, China.
}

\subsection{PMID}{
   25079328
}
}

\examples{
`QinN_2014.metaphlan_bugs_list.stool`()

}
