% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RampelliS_2015.marker_presence.R
\name{RampelliS_2015.marker_presence}
\alias{RampelliS_2015.marker_presence}
\alias{RampelliS_2015.marker_presence.stool}
\title{Marker presence data from the RampelliS_2015 dataset}
\source{
Department of Pharmacy and Biotechnology, University of Bologna, Bologna 40126, Italy.
}
\description{
Marker presence data from the RampelliS_2015 dataset
}
\section{Datasets}{
 

\subsection{RampelliS_2015.marker_presence.stool}{
   An ExpressionSet with 38 samples and 47,456 features specific to the stool bodysite
}
}

\examples{
RampelliS_2015.marker_presence.stool()

}
\seealso{
\code{\link{RampelliS_2015.genefamilies_relab}}, \code{\link{RampelliS_2015.marker_abundance}}, \code{\link{RampelliS_2015.metaphlan_bugs_list}}, \code{\link{RampelliS_2015.pathabundance_relab}}, \code{\link{RampelliS_2015.pathcoverage}}
}
