% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RampelliS_2015.metaphlan_bugs_list.R
\name{RampelliS_2015.metaphlan_bugs_list}
\alias{RampelliS_2015.metaphlan_bugs_list}
\alias{RampelliS_2015.metaphlan_bugs_list.stool}
\title{Taxonomic abundance data from the RampelliS_2015 dataset}
\source{
Department of Pharmacy and Biotechnology, University of Bologna, Bologna 40126, Italy.
}
\description{
Taxonomic abundance data from the RampelliS_2015 dataset
}
\section{Datasets}{
 

\subsection{RampelliS_2015.metaphlan_bugs_list.stool}{
   An ExpressionSet with 38 samples and 727 features specific to the stool bodysite
}
}
\seealso{
\code{\link{RampelliS_2015.genefamilies_relab}}, \code{\link{RampelliS_2015.marker_abundance}}, \code{\link{RampelliS_2015.marker_presence}}, \code{\link{RampelliS_2015.pathabundance_relab}}, \code{\link{RampelliS_2015.pathcoverage}}
}

