% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RampelliS_2015.pathabundance_relab.R
\name{RampelliS_2015.pathabundance_relab}
\alias{RampelliS_2015.pathabundance_relab}
\alias{RampelliS_2015.pathabundance_relab.stool}
\title{Relabeled pathabundance data from the RampelliS_2015 dataset}
\source{
Department of Pharmacy and Biotechnology, University of Bologna, Bologna 40126, Italy.
}
\description{
Relabeled pathabundance data from the RampelliS_2015 dataset
}
\section{Datasets}{
 

\subsection{RampelliS_2015.pathabundance_relab.stool}{
   An ExpressionSet with 38 samples and 6,798 features specific to the stool bodysite
}
}
\examples{
RampelliS_2015.pathabundance_relab.stool()

}
\seealso{
\code{\link{RampelliS_2015.genefamilies_relab}}, \code{\link{RampelliS_2015.marker_abundance}}, \code{\link{RampelliS_2015.marker_presence}}, \code{\link{RampelliS_2015.metaphlan_bugs_list}}, \code{\link{RampelliS_2015.pathcoverage}}
}

