% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMetadata.R
\docType{data}
\name{sampleMetadata}
\alias{sampleMetadata}
\title{Manually Curated Sample Metadata}
\format{
An object of class \code{data.frame} with 20284 rows and 134 columns.
}
\usage{
sampleMetadata
}
\description{
The combined sample-level metadata of all studies in curatedMetagenomicData.
}
\keyword{datasets}
