% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchirmerM_2016.R
\name{SchirmerM_2016}
\alias{SchirmerM_2016}
\alias{SchirmerM_2016.genefamilies_relab.stool}
\alias{SchirmerM_2016.marker_abundance.stool}
\alias{SchirmerM_2016.marker_presence.stool}
\alias{SchirmerM_2016.metaphlan_bugs_list.stool}
\alias{SchirmerM_2016.pathabundance_relab.stool}
\alias{SchirmerM_2016.pathcoverage.stool}
\title{Data from the SchirmerM_2016 study}
\description{
Data from the SchirmerM_2016 study
}
\section{Datasets}{
 

\subsection{SchirmerM_2016.genefamilies_relab.stool}{
   An ExpressionSet with 471 samples and 1,396,085 features specific to the stool body site
}

\subsection{SchirmerM_2016.marker_abundance.stool}{
   An ExpressionSet with 471 samples and 104,930 features specific to the stool body site
}

\subsection{SchirmerM_2016.marker_presence.stool}{
   An ExpressionSet with 471 samples and 101,457 features specific to the stool body site
}

\subsection{SchirmerM_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 471 samples and 1,177 features specific to the stool body site
}

\subsection{SchirmerM_2016.pathabundance_relab.stool}{
   An ExpressionSet with 471 samples and 12,707 features specific to the stool body site
}

\subsection{SchirmerM_2016.pathcoverage.stool}{
   An ExpressionSet with 471 samples and 12,707 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Linking the Human Gut Microbiome to Inflammatory Cytokine Production Capacity.
}

\subsection{Author}{
   Schirmer M, Smeekens SP, Vlamakis H, Jaeger M, Oosting M, Franzosa EA, Horst RT, Jansen T, Jacobs L, Bonder MJ, Kurilshikov A, Fu J, Joosten LAB, Zhernakova A, Huttenhower C, Wijmenga C, Netea MG, Xavier RJ
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   27984736
}
}

\examples{
SchirmerM_2016.metaphlan_bugs_list.stool()

}
