% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmitsSA_2017.R
\name{SmitsSA_2017}
\alias{SmitsSA_2017}
\alias{SmitsSA_2017.genefamilies_relab.stool}
\alias{SmitsSA_2017.marker_abundance.stool}
\alias{SmitsSA_2017.marker_presence.stool}
\alias{SmitsSA_2017.metaphlan_bugs_list.stool}
\alias{SmitsSA_2017.pathabundance_relab.stool}
\alias{SmitsSA_2017.pathcoverage.stool}
\title{Data from the SmitsSA_2017 study}
\description{
Data from the SmitsSA_2017 study
}
\section{Datasets}{
 

\subsection{SmitsSA_2017.genefamilies_relab.stool}{
   An ExpressionSet with 40 samples and 400,325 features specific to the stool body site
}

\subsection{SmitsSA_2017.marker_abundance.stool}{
   An ExpressionSet with 40 samples and 21,684 features specific to the stool body site
}

\subsection{SmitsSA_2017.marker_presence.stool}{
   An ExpressionSet with 40 samples and 19,635 features specific to the stool body site
}

\subsection{SmitsSA_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 40 samples and 343 features specific to the stool body site
}

\subsection{SmitsSA_2017.pathabundance_relab.stool}{
   An ExpressionSet with 40 samples and 3,562 features specific to the stool body site
}

\subsection{SmitsSA_2017.pathcoverage.stool}{
   An ExpressionSet with 40 samples and 3,562 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Seasonal cycling in the gut microbiome of the Hadza hunter-gatherers of Tanzania.
}

\subsection{Author}{
   Smits SA, Leach J, Sonnenburg ED, Gonzalez CG, Lichtman JS, Reid G, Knight R, Manjurano A, Changalucha J, Elias JE, Dominguez-Bello MG, Sonnenburg JL
}

\subsection{Lab}{
   [1] Department of Microbiology and Immunology, Stanford University School of Medicine, Stanford, CA 94305, USA., [2] Human Food Project, 53600 Highway 118, Terlingua, TX 79852, USA., [3] The Department of Twin Research and Genetic Epidemiology, King's College London, St. Thomas' Hospital, Lambeth Palace Road, London SE1 7EH, UK., [4] Department of Chemical and Systems Biology, Stanford School of Medicine, Stanford University, Stanford, CA 94025, USA., [5] Lawson Health Research Institute and Western University, London, Ontario N6A 4V2, Canada., [6] Departments of Pediatrics and Computer Science and Engineering and Center for Microbiome Innovation, University of California, San Diego, CA 92093, USA., [7] National Institute for Medical Research, Mwanza 11101, Tanzania., [8] School of Medicine and Department of Anthropology, New York University, New York, NY, USA.
}

\subsection{PMID}{
   28839072
}
}

\examples{
SmitsSA_2017.metaphlan_bugs_list.stool()

}
