% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2016.marker_presence.R
\name{TettAJ_2016.marker_presence}
\alias{TettAJ_2016.marker_presence}
\alias{TettAJ_2016.marker_presence.left_ear}
\alias{TettAJ_2016.marker_presence.left_elbow}
\alias{TettAJ_2016.marker_presence.right_ear}
\alias{TettAJ_2016.marker_presence.right_elbow}
\title{Marker presence data from the TettAJ_2016 dataset}
\source{

}
\description{
Marker presence data from the TettAJ_2016 dataset
}
\section{Datasets}{
 

\subsection{TettAJ_2016.marker_presence.left_ear}{
   An ExpressionSet with 22 samples and 57,429 features specific to the left_ear bodysite
}

\subsection{TettAJ_2016.marker_presence.left_elbow}{
   An ExpressionSet with 26 samples and 57,429 features specific to the left_elbow bodysite
}

\subsection{TettAJ_2016.marker_presence.right_ear}{
   An ExpressionSet with 26 samples and 57,429 features specific to the right_ear bodysite
}

\subsection{TettAJ_2016.marker_presence.right_elbow}{
   An ExpressionSet with 23 samples and 57,429 features specific to the right_elbow bodysite
}
}

\examples{
TettAJ_2016.marker_presence.left_ear()

}
\seealso{
\code{\link{TettAJ_2016.genefamilies_relab}}, \code{\link{TettAJ_2016.marker_abundance}}, \code{\link{TettAJ_2016.metaphlan_bugs_list}}, \code{\link{TettAJ_2016.pathabundance_relab}}, \code{\link{TettAJ_2016.pathcoverage}}
}
