% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2016.pathabundance_relab.R
\name{TettAJ_2016.pathabundance_relab}
\alias{TettAJ_2016.pathabundance_relab}
\alias{TettAJ_2016.pathabundance_relab.left_ear}
\title{Relabeled pathabundance data from the TettAJ_2016 dataset}
\source{

}
\description{
Relabeled pathabundance data from the TettAJ_2016 dataset
}
\section{Datasets}{
 

\subsection{TettAJ_2016.pathabundance_relab.left_ear}{
   An ExpressionSet with 22 samples and 21,899 features specific to the left_ear bodysite
}
}
\seealso{
\code{\link{TettAJ_2016.genefamilies_relab}}, \code{\link{TettAJ_2016.genefamilies_relab}}, \code{\link{TettAJ_2016.genefamilies_relab}}, \code{\link{TettAJ_2016.genefamilies_relab}}, \code{\link{TettAJ_2016.marker_abundance}}, \code{\link{TettAJ_2016.marker_abundance}}, \code{\link{TettAJ_2016.marker_abundance}}, \code{\link{TettAJ_2016.marker_abundance}}, \code{\link{TettAJ_2016.marker_presence}}, \code{\link{TettAJ_2016.marker_presence}}, \code{\link{TettAJ_2016.marker_presence}}, \code{\link{TettAJ_2016.marker_presence}}, \code{\link{TettAJ_2016.metaphlan_bugs_list}}, \code{\link{TettAJ_2016.metaphlan_bugs_list}}, \code{\link{TettAJ_2016.metaphlan_bugs_list}}, \code{\link{TettAJ_2016.metaphlan_bugs_list}}, \code{\link{TettAJ_2016.pathcoverage}}, \code{\link{TettAJ_2016.pathcoverage}}, \code{\link{TettAJ_2016.pathcoverage}}, \code{\link{TettAJ_2016.pathcoverage}}
}

