% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2016.R
\name{TettAJ_2016}
\alias{TettAJ_2016}
\alias{TettAJ_2016.genefamilies_relab.skin}
\alias{TettAJ_2016.marker_abundance.skin}
\alias{TettAJ_2016.marker_presence.skin}
\alias{TettAJ_2016.metaphlan_bugs_list.skin}
\alias{TettAJ_2016.pathabundance_relab.skin}
\alias{TettAJ_2016.pathcoverage.skin}
\title{Data from the TettAJ_2016 study}
\description{
Data from the TettAJ_2016 study
}
\section{Datasets}{
 

\subsection{TettAJ_2016.genefamilies_relab.skin}{
   An ExpressionSet with 97 samples and 1,177,112 features specific to the skin body site
}

\subsection{TettAJ_2016.marker_abundance.skin}{
   An ExpressionSet with 97 samples and 64,205 features specific to the skin body site
}

\subsection{TettAJ_2016.marker_presence.skin}{
   An ExpressionSet with 97 samples and 59,833 features specific to the skin body site
}

\subsection{TettAJ_2016.metaphlan_bugs_list.skin}{
   An ExpressionSet with 97 samples and 1,002 features specific to the skin body site
}

\subsection{TettAJ_2016.pathabundance_relab.skin}{
   An ExpressionSet with 97 samples and 18,914 features specific to the skin body site
}

\subsection{TettAJ_2016.pathcoverage.skin}{
   An ExpressionSet with 97 samples and 18,914 features specific to the skin body site
}
}

\section{Source}{
 

\subsection{Title}{
   Unexplored diversity and strain-level structure of the skin microbiome associated with psoriasis.
}

\subsection{Author}{
   Tett A, Pasolli E, Farina S, Truong DT, Asnicar F, Zolfo M, Beghini F, Armanini F, Jousson O, De Sanctis V, Bertorelli R, Girolomoni G, Cristofolini M, Segata N
}

\subsection{Lab}{
   [1] Centre for Integrative Biology, University of Trento, Trento, Italy., [2] Istituto G.B. Mattei, Comano, Italy., [3] NGS Facility, Laboratory of Biomolecular Sequence and Structure Analysis for Health, Centre for Integrative Biology, University of Trento, Trento, Italy., [4] Department of Medicine, Section of Dermatology, University of Verona, Verona, Italy.
}

\subsection{PMID}{
   28649415
}
}

\examples{
`TettAJ_2016.metaphlan_bugs_list.skin`()

}
