% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2019_a.R
\name{TettAJ_2019_a}
\alias{TettAJ_2019_a}
\alias{TettAJ_2019_a.genefamilies_relab.stool}
\alias{TettAJ_2019_a.marker_abundance.stool}
\alias{TettAJ_2019_a.marker_presence.stool}
\alias{TettAJ_2019_a.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_a.pathabundance_relab.stool}
\alias{TettAJ_2019_a.pathcoverage.stool}
\title{Data from the TettAJ_2019_a study}
\description{
Data from the TettAJ_2019_a study
}
\section{Datasets}{
 

\subsection{TettAJ_2019_a.genefamilies_relab.stool}{
   An ExpressionSet with 68 samples and 1,173,786 features specific to the stool body site
}

\subsection{TettAJ_2019_a.marker_abundance.stool}{
   An ExpressionSet with 68 samples and 75,815 features specific to the stool body site
}

\subsection{TettAJ_2019_a.marker_presence.stool}{
   An ExpressionSet with 68 samples and 73,480 features specific to the stool body site
}

\subsection{TettAJ_2019_a.metaphlan_bugs_list.stool}{
   An ExpressionSet with 68 samples and 960 features specific to the stool body site
}

\subsection{TettAJ_2019_a.pathabundance_relab.stool}{
   An ExpressionSet with 68 samples and 11,310 features specific to the stool body site
}

\subsection{TettAJ_2019_a.pathcoverage.stool}{
   An ExpressionSet with 68 samples and 11,310 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
TettAJ_2019_a.metaphlan_bugs_list.stool()

}
