% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2019_b.R
\name{TettAJ_2019_b}
\alias{TettAJ_2019_b}
\alias{TettAJ_2019_b.genefamilies_relab.stool}
\alias{TettAJ_2019_b.marker_abundance.stool}
\alias{TettAJ_2019_b.marker_presence.stool}
\alias{TettAJ_2019_b.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_b.pathabundance_relab.stool}
\alias{TettAJ_2019_b.pathcoverage.stool}
\title{Data from the TettAJ_2019_b study}
\description{
Data from the TettAJ_2019_b study
}
\section{Datasets}{
 

\subsection{TettAJ_2019_b.genefamilies_relab.stool}{
   An ExpressionSet with 44 samples and 1,004,301 features specific to the stool body site
}

\subsection{TettAJ_2019_b.marker_abundance.stool}{
   An ExpressionSet with 44 samples and 70,004 features specific to the stool body site
}

\subsection{TettAJ_2019_b.marker_presence.stool}{
   An ExpressionSet with 44 samples and 67,823 features specific to the stool body site
}

\subsection{TettAJ_2019_b.metaphlan_bugs_list.stool}{
   An ExpressionSet with 44 samples and 849 features specific to the stool body site
}

\subsection{TettAJ_2019_b.pathabundance_relab.stool}{
   An ExpressionSet with 44 samples and 10,531 features specific to the stool body site
}

\subsection{TettAJ_2019_b.pathcoverage.stool}{
   An ExpressionSet with 44 samples and 10,531 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
TettAJ_2019_b.metaphlan_bugs_list.stool()

}
