% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TettAJ_2019_c.R
\name{TettAJ_2019_c}
\alias{TettAJ_2019_c}
\alias{TettAJ_2019_c.genefamilies_relab.stool}
\alias{TettAJ_2019_c.marker_abundance.stool}
\alias{TettAJ_2019_c.marker_presence.stool}
\alias{TettAJ_2019_c.metaphlan_bugs_list.stool}
\alias{TettAJ_2019_c.pathabundance_relab.stool}
\alias{TettAJ_2019_c.pathcoverage.stool}
\title{Data from the TettAJ_2019_c study}
\description{
Data from the TettAJ_2019_c study
}
\section{Datasets}{
 

\subsection{TettAJ_2019_c.genefamilies_relab.stool}{
   An ExpressionSet with 50 samples and 1,023,154 features specific to the stool body site
}

\subsection{TettAJ_2019_c.marker_abundance.stool}{
   An ExpressionSet with 50 samples and 62,479 features specific to the stool body site
}

\subsection{TettAJ_2019_c.marker_presence.stool}{
   An ExpressionSet with 50 samples and 60,773 features specific to the stool body site
}

\subsection{TettAJ_2019_c.metaphlan_bugs_list.stool}{
   An ExpressionSet with 50 samples and 866 features specific to the stool body site
}

\subsection{TettAJ_2019_c.pathabundance_relab.stool}{
   An ExpressionSet with 50 samples and 9,250 features specific to the stool body site
}

\subsection{TettAJ_2019_c.pathcoverage.stool}{
   An ExpressionSet with 50 samples and 9,250 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
TettAJ_2019_c.metaphlan_bugs_list.stool()

}
