% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThomasAM_2018a.R
\name{ThomasAM_2018a}
\alias{ThomasAM_2018a}
\alias{ThomasAM_2018a.genefamilies_relab.stool}
\alias{ThomasAM_2018a.marker_abundance.stool}
\alias{ThomasAM_2018a.marker_presence.stool}
\alias{ThomasAM_2018a.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018a.pathabundance_relab.stool}
\alias{ThomasAM_2018a.pathcoverage.stool}
\title{Data from the ThomasAM_2018a study}
\description{
Data from the ThomasAM_2018a study
}
\section{Datasets}{
 

\subsection{ThomasAM_2018a.genefamilies_relab.stool}{
   An ExpressionSet with 80 samples and 1,324,452 features specific to the stool body site
}

\subsection{ThomasAM_2018a.marker_abundance.stool}{
   An ExpressionSet with 80 samples and 88,476 features specific to the stool body site
}

\subsection{ThomasAM_2018a.marker_presence.stool}{
   An ExpressionSet with 80 samples and 86,326 features specific to the stool body site
}

\subsection{ThomasAM_2018a.metaphlan_bugs_list.stool}{
   An ExpressionSet with 80 samples and 1,162 features specific to the stool body site
}

\subsection{ThomasAM_2018a.pathabundance_relab.stool}{
   An ExpressionSet with 80 samples and 8,227 features specific to the stool body site
}

\subsection{ThomasAM_2018a.pathcoverage.stool}{
   An ExpressionSet with 80 samples and 8,227 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
ThomasAM_2018a.metaphlan_bugs_list.stool()

}
