% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThomasAM_2018b.R
\name{ThomasAM_2018b}
\alias{ThomasAM_2018b}
\alias{ThomasAM_2018b.genefamilies_relab.stool}
\alias{ThomasAM_2018b.marker_abundance.stool}
\alias{ThomasAM_2018b.marker_presence.stool}
\alias{ThomasAM_2018b.metaphlan_bugs_list.stool}
\alias{ThomasAM_2018b.pathabundance_relab.stool}
\alias{ThomasAM_2018b.pathcoverage.stool}
\title{Data from the ThomasAM_2018b study}
\description{
Data from the ThomasAM_2018b study
}
\section{Datasets}{
 

\subsection{ThomasAM_2018b.genefamilies_relab.stool}{
   An ExpressionSet with 60 samples and 1,401,350 features specific to the stool body site
}

\subsection{ThomasAM_2018b.marker_abundance.stool}{
   An ExpressionSet with 60 samples and 93,190 features specific to the stool body site
}

\subsection{ThomasAM_2018b.marker_presence.stool}{
   An ExpressionSet with 60 samples and 91,404 features specific to the stool body site
}

\subsection{ThomasAM_2018b.metaphlan_bugs_list.stool}{
   An ExpressionSet with 60 samples and 1,079 features specific to the stool body site
}

\subsection{ThomasAM_2018b.pathabundance_relab.stool}{
   An ExpressionSet with 60 samples and 8,107 features specific to the stool body site
}

\subsection{ThomasAM_2018b.pathcoverage.stool}{
   An ExpressionSet with 60 samples and 8,107 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   NA
}

\subsection{Author}{
   NA
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   NA
}
}

\examples{
ThomasAM_2018b.metaphlan_bugs_list.stool()

}
