% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{uploads_dir}
\alias{uploads_dir}
\title{create a 'globally usable' temporary folder with name 'uploads'}
\usage{
uploads_dir(path = cmg_tmpdir())
}
\arguments{
\item{path}{character(1) path to a temporary folder}
}
\description{
create a 'globally usable' temporary folder with name 'uploads'
}
