% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VatanenT_2016.R
\name{VatanenT_2016}
\alias{VatanenT_2016}
\alias{VatanenT_2016.genefamilies_relab.stool}
\alias{VatanenT_2016.marker_abundance.stool}
\alias{VatanenT_2016.marker_presence.stool}
\alias{VatanenT_2016.metaphlan_bugs_list.stool}
\alias{VatanenT_2016.pathabundance_relab.stool}
\alias{VatanenT_2016.pathcoverage.stool}
\title{Data from the VatanenT_2016 study}
\description{
Data from the VatanenT_2016 study
}
\section{Datasets}{
 

\subsection{VatanenT_2016.genefamilies_relab.stool}{
   An ExpressionSet with 785 samples and 1,719,634 features specific to the stool body site
}

\subsection{VatanenT_2016.marker_abundance.stool}{
   An ExpressionSet with 785 samples and 135,979 features specific to the stool body site
}

\subsection{VatanenT_2016.marker_presence.stool}{
   An ExpressionSet with 785 samples and 131,625 features specific to the stool body site
}

\subsection{VatanenT_2016.metaphlan_bugs_list.stool}{
   An ExpressionSet with 785 samples and 1,584 features specific to the stool body site
}

\subsection{VatanenT_2016.pathabundance_relab.stool}{
   An ExpressionSet with 785 samples and 19,236 features specific to the stool body site
}

\subsection{VatanenT_2016.pathcoverage.stool}{
   An ExpressionSet with 785 samples and 19,236 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Variation in Microbiome LPS Immunogenicity Contributes to Autoimmunity in Humans.
}

\subsection{Author}{
   Vatanen T, Kostic AD, d'Hennezel E, Siljander H, Franzosa EA, Yassour M, Kolde R, Vlamakis H, Arthur TD, Hämäläinen AM, Peet A, Tillmann V, Uibo R, Mokurov S, Dorshakova N, Ilonen J, Virtanen SM, Szabo SJ, Porter JA, Lähdesmäki H, Huttenhower C, Gevers D, Cullen TW, Knip M, Xavier RJ
}

\subsection{Lab}{
   NA
}

\subsection{PMID}{
   27259157
}
}

\examples{
`VatanenT_2016.metaphlan_bugs_list.stool`()

}
