% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WenC_2017.R
\name{WenC_2017}
\alias{WenC_2017}
\alias{WenC_2017.genefamilies_relab.stool}
\alias{WenC_2017.marker_abundance.stool}
\alias{WenC_2017.marker_presence.stool}
\alias{WenC_2017.metaphlan_bugs_list.stool}
\alias{WenC_2017.pathabundance_relab.stool}
\alias{WenC_2017.pathcoverage.stool}
\title{Data from the WenC_2017 study}
\description{
Data from the WenC_2017 study
}
\section{Datasets}{
 

\subsection{WenC_2017.genefamilies_relab.stool}{
   An ExpressionSet with 97 samples and 1,160,257 features specific to the stool body site
}

\subsection{WenC_2017.marker_abundance.stool}{
   An ExpressionSet with 97 samples and 89,660 features specific to the stool body site
}

\subsection{WenC_2017.marker_presence.stool}{
   An ExpressionSet with 97 samples and 83,333 features specific to the stool body site
}

\subsection{WenC_2017.metaphlan_bugs_list.stool}{
   An ExpressionSet with 97 samples and 1,075 features specific to the stool body site
}

\subsection{WenC_2017.pathabundance_relab.stool}{
   An ExpressionSet with 97 samples and 11,713 features specific to the stool body site
}

\subsection{WenC_2017.pathcoverage.stool}{
   An ExpressionSet with 97 samples and 11,713 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Quantitative metagenomics reveals unique gut microbiome biomarkers in ankylosing spondylitis.
}

\subsection{Author}{
   Wen C, Zheng Z, Shao T, Liu L, Xie Z, Le Chatelier E, He Z, Zhong W, Fan Y, Zhang L, Li H, Wu C, Hu C, Xu Q, Zhou J, Cai S, Wang D, Huang Y, Breban M, Qin N, Ehrlich SD
}

\subsection{Lab}{
   [1] Institute of Basic Research in Clinical Medicine, College of Basic Medical Science, Zhejiang Chinese Medical University, Hangzhou, 310053, China. wengcp@163.com., [2] Realbio Genomics Institute, Shanghai, 200123, China., [3] Institute of Basic Research in Clinical Medicine, College of Basic Medical Science, Zhejiang Chinese Medical University, Hangzhou, 310053, China., [4] State Key Laboratory for Diagnosis and Treatment of Infectious Diseases, Department of Infectious Diseases, the First Affiliated College of Medicine, Zhejiang University, Hangzhou, 310003, China., [5] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, Hangzhou, 310003, China., [6] INRA, Institut National de la Recherche Agronomique, Metagenopolis, Jouy en Josas, 78350, France., [7] Rheumatology Division, Ambroise-Pare Hospital, AP-HP, 9, avenue Charles-de-Gaulle, 92100, Boulogne-Billancourt, France., [8] Realbio Genomics Institute, Shanghai, 200123, China. qinnan001@126.com., [9] State Key Laboratory for Diagnosis and Treatment of Infectious Diseases, Department of Infectious Diseases, the First Affiliated College of Medicine, Zhejiang University, Hangzhou, 310003, China. qinnan001@126.com., [10] Collaborative Innovation Center for Diagnosis and Treatment of Infectious Diseases, Zhejiang University, Hangzhou, 310003, China. qinnan001@126.com., [11] INRA, Institut National de la Recherche Agronomique, Metagenopolis, Jouy en Josas, 78350, France. dusko.ehrlich@jouy.inra.fr., [12] King's College London, Centre for Host-Microbiome Interactions, Dental Institute Central Office, Guy's Hospital, London Bridge, London, SE1 9RT, UK. dusko.ehrlich@jouy.inra.fr.
}

\subsection{PMID}{
   28750650
}
}

\examples{
WenC_2017.metaphlan_bugs_list.stool()

}
