% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZellerG_2014.genefamilies_relab.R
\name{ZellerG_2014.genefamilies_relab}
\alias{ZellerG_2014.genefamilies_relab}
\alias{ZellerG_2014.genefamilies_relab.stool}
\title{Relabeled genefamilies data from the ZellerG_2014 dataset}
\source{
Structural and Computational Biology Unit, European Molecular Biology Laboratory, Heidelberg, Germany.
}
\description{
Relabeled genefamilies data from the ZellerG_2014 dataset
}
\section{Datasets}{
 

\subsection{ZellerG_2014.genefamilies_relab.stool}{
   An ExpressionSet with 156 samples and 1,796,274 features specific to the stool bodysite
}
}
\examples{
ZellerG_2014.genefamilies_relab.stool()

}
\seealso{
\code{\link{ZellerG_2014.marker_abundance}}, \code{\link{ZellerG_2014.marker_presence}}, \code{\link{ZellerG_2014.metaphlan_bugs_list}}, \code{\link{ZellerG_2014.pathabundance_relab}}, \code{\link{ZellerG_2014.pathcoverage}}
}

