% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZellerG_2014.R
\name{ZellerG_2014}
\alias{ZellerG_2014}
\alias{ZellerG_2014.genefamilies_relab.stool}
\alias{ZellerG_2014.marker_abundance.stool}
\alias{ZellerG_2014.marker_presence.stool}
\alias{ZellerG_2014.metaphlan_bugs_list.stool}
\alias{ZellerG_2014.pathabundance_relab.stool}
\alias{ZellerG_2014.pathcoverage.stool}
\title{Data from the ZellerG_2014 study}
\description{
Data from the ZellerG_2014 study
}
\section{Datasets}{
 

\subsection{ZellerG_2014.genefamilies_relab.stool}{
   An ExpressionSet with 199 samples and 1,847,962 features specific to the stool body site
}

\subsection{ZellerG_2014.marker_abundance.stool}{
   An ExpressionSet with 199 samples and 138,412 features specific to the stool body site
}

\subsection{ZellerG_2014.marker_presence.stool}{
   An ExpressionSet with 199 samples and 133,484 features specific to the stool body site
}

\subsection{ZellerG_2014.metaphlan_bugs_list.stool}{
   An ExpressionSet with 199 samples and 1,585 features specific to the stool body site
}

\subsection{ZellerG_2014.pathabundance_relab.stool}{
   An ExpressionSet with 199 samples and 18,579 features specific to the stool body site
}

\subsection{ZellerG_2014.pathcoverage.stool}{
   An ExpressionSet with 199 samples and 18,579 features specific to the stool body site
}
}

\section{Source}{
 

\subsection{Title}{
   Potential of fecal microbiota for early-stage detection of colorectal cancer.
}

\subsection{Author}{
   Zeller G, Tap J, Voigt AY, Sunagawa S, Kultima JR, Costea PI, Amiot A, Böhm J, Brunetti F, Habermann N, Hercog R, Koch M, Luciani A, Mende DR, Schneider MA, Schrotz-King P, Tournigand C, Tran Van Nhieu J, Yamada T, Zimmermann J, Benes V, Kloor M, Ulrich CM, von Knebel Doeberitz M, Sobhani I, Bork P
}

\subsection{Lab}{
   [1] Structural and Computational Biology Unit, European Molecular Biology Laboratory, Heidelberg, Germany., [2] Structural and Computational Biology Unit, European Molecular Biology Laboratory, Heidelberg, Germany Department of Gastroenterology and LIC-EA4393-EC2M3, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [3] Structural and Computational Biology Unit, European Molecular Biology Laboratory, Heidelberg, Germany Department of Applied Tumor Biology, Institute of Pathology University Hospital Heidelberg, Heidelberg, Germany Clinical Cooperation Unit Applied Tumor Biology, German Cancer Research Center (DKFZ), Heidelberg, Germany Molecular Medicine Partnership Unit (MMPU), University Hospital Heidelberg and European Molecular Biology Laboratory, Heidelberg, Germany., [4] Department of Gastroenterology and LIC-EA4393-EC2M3, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [5] Division of Preventive Oncology, National Center for Tumor Diseases (NCT) Heidelberg, Heidelberg, Germany German Cancer Research Center (DKFZ), Heidelberg, Germany., [6] Department of Surgery, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [7] Genomics Core Facility, European Molecular Biology Laboratory, Heidelberg, Germany., [8] Department of General, Visceral and Transplantation Surgery, University Hospital Heidelberg, Heidelberg, Germany., [9] Department of Radiology, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [10] Department of Medical Oncology, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [11] Department of Pathology and LIC-EA4393-EC2M3, APHP and UPEC Universite Paris-Est Creteil, Creteil, France., [12] Department of Biological Information, Tokyo Institute of Technology, Tokyo, Japan., [13] Department of Applied Tumor Biology, Institute of Pathology University Hospital Heidelberg, Heidelberg, Germany Clinical Cooperation Unit Applied Tumor Biology, German Cancer Research Center (DKFZ), Heidelberg, Germany Molecular Medicine Partnership Unit (MMPU), University Hospital Heidelberg and European Molecular Biology Laboratory, Heidelberg, Germany., [14] Division of Preventive Oncology, National Center for Tumor Diseases (NCT) Heidelberg, Heidelberg, Germany German Cancer Research Center (DKFZ), Heidelberg, Germany Fred Hutchinson Cancer Research Center (FHCRC), Seattle, WA, USA., [15] Department of Gastroenterology and LIC-EA4393-EC2M3, APHP and UPEC Universite Paris-Est Creteil, Creteil, France iradj.sobhani@hmn.aphp.fr bork@embl.de., [16] Structural and Computational Biology Unit, European Molecular Biology Laboratory, Heidelberg, Germany Molecular Medicine Partnership Unit (MMPU), University Hospital Heidelberg and European Molecular Biology Laboratory, Heidelberg, Germany Max Delbruck Centre for Molecular Medicine, Berlin, Germany iradj.sobhani@hmn.aphp.fr bork@embl.de.
}

\subsection{PMID}{
   25432777
}
}

\examples{
`ZellerG_2014.metaphlan_bugs_list.stool`()

}
