% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_metadata.R
\docType{data}
\name{combined_metadata}
\alias{combined_metadata}
\title{Start Using 'sampleMetadata' Instead}
\format{
An object of class \code{data.frame} with 20283 rows and 130 columns.
}
\usage{
combined_metadata
}
\description{
Start using 'sampleMetadata' instead, 'combined_metadata' will be removed in the next release.
}
\seealso{
\link{sampleMetadata}
}
\keyword{datasets}
