% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_metadata.R
\docType{data}
\name{combined_metadata}
\alias{combined_metadata}
\title{Manually Curated Sample Metadata}
\format{
An object of class \code{data.frame} with 20283 rows and 134 columns.
}
\usage{
combined_metadata
}
\description{
Manually curated sample metadata for all samples in curatedMetagenomicData.
}
\keyword{datasets}
