% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData.R
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\title{Title}
\usage{
curatedMetagenomicData(x)
}
\arguments{
\item{x}{description}
}
\value{
description
}
\description{
Description
}
