% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData.R
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\title{Access Curated Metagenomic Data}
\usage{
curatedMetagenomicData(pattern, dryrun = TRUE, counts = FALSE)
}
\arguments{
\item{pattern}{description}

\item{dryrun}{description}

\item{counts}{description}
}
\value{
description
}
\description{
Description
}
