% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData-package.R
\docType{package}
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\alias{curatedMetagenomicData-package}
\title{curatedMetagenomicData.}
\description{
curatedMetagenomicData.
}

