% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData.R
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\title{Download Metadata/Data for Any/All Studies in curatedMetagenomicData}
\usage{
curatedMetagenomicData(x = "", dryrun = TRUE, counts = FALSE)
}
\arguments{
\item{x}{regular expression(s) to be matched against curatedMetagenomicData;
if multiple, they are combined by logical \code{&}}

\item{dryrun}{\code{logical} indicating if the user would like to return metadata
(the default); when set to \code{FALSE} a \code{SummarizedExperiment} is returned}

\item{counts}{\code{logical} indicating if the user would like to return data as
relative abundances (the default); or return data as counts}
}
\value{
a \code{data.frame} of ExperimentHub metadata when \code{dryrun = TRUE}, and a
\code{SummarizedExperiment} when \code{dryrun = FALSE}
}
\description{
\code{curatedMetagenomicData()} is a simplified interface to what was previously a
rather complicated way to access data in the package. Users should first look
through the \code{metadata}, which has information about every sample currently
available in \code{curatedMetagenomicData} (and those that are forthcoming), to
"shop" for a study they would like to download. Then, a query to access data
can be perfected while \code{dryrun = TRUE} because it only returns metadata.
Finally, \code{dryrun} can be set to \code{FALSE} and a \code{SummarizedExperiment} object
will be returned.
}
\examples{
curatedMetagenomicData(x = "AsnicarF_2017.marker_presence")

}
