% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData-package.r
\docType{package}
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\alias{curatedMetagenomicData-package}
\title{Curated Metagenomic Data of the Human Microbiome}
\description{
The curatedMetagenomicData package provides taxonomic, functional, and gene
marker abundance for samples collected from different bodysites. It provides
data from aproximately 3000 human microbiome samples that has been highly
processed, refined, and curated such that analysis that might otherwise
require a computing cluster can be done on an ordianary laptop.
}

