% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData-package.R,
%   R/curatedMetagenomicData.R
\docType{package}
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\alias{curatedMetagenomicData-package}
\alias{curatedMetagenomicData}
\title{Curated Metagenomic Data of the Human Microbiome}
\usage{
curatedMetagenomicData(x = "*", dryrun = TRUE, counts = FALSE,
  bugs.as.phyloseq = FALSE, x.is.glob = TRUE)
}
\arguments{
\item{x}{A character vector of dataset names, regexes, or globs, that will be matched
to available datasets. If x.is.glob is TRUE (default), wildcards such as "*"
and "?" are supported (see ?glob2rx), otherwise, regexes are supported (see
?grep)}

\item{counts}{= FALSE
If TRUE, relative abundances will be multiplied by read depth, then rounded to
the nearest integer.}

\item{bugs.as.phyloseq}{= FALSE
If TRUE, tables of taxonomic abundance (metaphlan datasets) will be converted
to phyloseq objects for use with the phyloseq package.}

\item{x.is.glob}{= TRUE
Set to FALSE to actually download the datasets}
}
\value{
A list of ExpressionSet and/or phyloseq objects
}
\description{
The curatedMetagenomicData package provides taxonomic, functional, and gene
marker abundance for samples collected from different bodysites. It provides
data from aproximately 3000 human microbiome samples that has been highly
processed, refined, and curated such that analysis that might otherwise
require a computing cluster can be done on an ordianary laptop.
}
\examples{
curatedMetagenomicData()
curatedMetagenomicData("ZellerG*")
curatedMetagenomicData("ZellerG.+marker", x.is.glob=FALSE)
curatedMetagenomicData("ZellerG_2014.metaphlan_bugs_list.stool", dryrun=FALSE)
curatedMetagenomicData("ZellerG_2014.metaphlan_bugs_list.stool",
        counts=TRUE, dryrun=FALSE, bugs.as.phyloseq=TRUE)
}
