% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curatedMetagenomicData.R
\name{curatedMetagenomicData}
\alias{curatedMetagenomicData}
\title{Curated Metagenomic Data of the Human Microbiome}
\description{
The curatedMetagenomicData package provides taxonomic, functional, and gene
marker abundance for samples collected from different bodysites.
}

