\name{curatedMetagenomicData-defunct}
\alias{Castro_NallarE_2015.genefamilies_relab.oralcavity}
\alias{Castro_NallarE_2015.marker_abundance.oralcavity}
\alias{Castro_NallarE_2015.marker_presence.oralcavity}
\alias{Castro_NallarE_2015.metaphlan_bugs_list.oralcavity}
\alias{Castro_NallarE_2015.pathabundance_relab.oralcavity}
\alias{Castro_NallarE_2015.pathcoverage.oralcavity}
\alias{Heitz_BuschartA_2016.genefamilies_relab.stool}
\alias{Heitz_BuschartA_2016.marker_abundance.stool}
\alias{Heitz_BuschartA_2016.marker_presence.stool}
\alias{Heitz_BuschartA_2016.metaphlan_bugs_list.stool}
\alias{Heitz_BuschartA_2016.pathabundance_relab.stool}
\alias{Heitz_BuschartA_2016.pathcoverage.stool}
\alias{Obregon_TitoAJ_2015.genefamilies_relab.stool}
\alias{Obregon_TitoAJ_2015.marker_abundance.stool}
\alias{Obregon_TitoAJ_2015.marker_presence.stool}
\alias{Obregon_TitoAJ_2015.metaphlan_bugs_list.stool}
\alias{Obregon_TitoAJ_2015.pathabundance_relab.stool}
\alias{Obregon_TitoAJ_2015.pathcoverage.stool}
\title{Defunct functions in package \sQuote{curatedMetagenomicData}}
\description{These functions are defunct and no longer available.}

\details{
  Defunct functions and arguments include any functions or arguments referencing: \code{Bengtsson_PalmeJ}, \code{Castro_NallarE}, \code{Heitz_BuschartA}, and \code{Obregon_TitoAJ}. Instead use \code{Bengtsson-PalmeJ}, \code{Castro-NallarE}, \code{Heitz-BuschartA}, and \code{Obregon-TitoAJ}, respectively.
}
