\name{curatedMetagenomicData-defunct}
\alias{}
\title{Defunct functions in package \sQuote{curatedMetagenomicData}}
\description{These functions are defunct and no longer available.}

\details{
  Defunct functions and arguments include any functions or arguments referencing: \code{Bengtsson_PalmeJ}, \code{Castro_NallarE}, \code{Heitz_BuschartA}, and \code{Obregon_TitoAJ}. Instead use \code{Bengtsson-PalmeJ}, \code{Castro-NallarE}, \code{Heitz-BuschartA}, and \code{Obregon-TitoAJ}, respectively.
}
