% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionSet2phyloseq.R
\name{ExpressionSet2phyloseq}
\alias{ExpressionSet2phyloseq}
\title{Convert an ExpressionSet object to a phyloseq object}
\usage{
ExpressionSet2phyloseq(ExpressionSet)
}
\arguments{
\item{ExpressionSet}{An ExpressionSet object}
}
\value{
A phyloseq object
}
\description{
Convert an ExpressionSet object to a phyloseq object
}

