% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMP_2012.pathcoverage.R
\name{HMP_2012.pathcoverage}
\alias{HMP_2012.pathcoverage}
\alias{HMP_2012.pathcoverage.anterior_nares}
\alias{HMP_2012.pathcoverage.buccal_mucosa}
\alias{HMP_2012.pathcoverage.hard_palate}
\alias{HMP_2012.pathcoverage.keratinized_gingiva}
\alias{HMP_2012.pathcoverage.l_retroauricular_crease}
\alias{HMP_2012.pathcoverage.mid_vagina}
\alias{HMP_2012.pathcoverage.palatine_tonsils}
\alias{HMP_2012.pathcoverage.posterior_fornix}
\alias{HMP_2012.pathcoverage.r_retroauricular_crease}
\alias{HMP_2012.pathcoverage.saliva}
\alias{HMP_2012.pathcoverage.stool}
\alias{HMP_2012.pathcoverage.subgingival_plaque}
\alias{HMP_2012.pathcoverage.supragingival_plaque}
\alias{HMP_2012.pathcoverage.throat}
\alias{HMP_2012.pathcoverage.tongue_dorsum}
\alias{HMP_2012.pathcoverage.vaginal_introitus}
\title{Pathcoverage data from the HMP_2012 dataset}
\source{
Department of Psychology, Abdul Haq Campus, Federal Urdu University for Arts, Science and Technology, Karachi, Pakistan. shahiq_psy@yahoo.com
}
\description{
Pathcoverage data from the HMP_2012 dataset
}
\section{Datasets}{
 

\subsection{HMP_2012.pathcoverage.anterior_nares}{
   An ExpressionSet with 93 samples and 23,271 features specific to the anterior_nares bodysite
}

\subsection{HMP_2012.pathcoverage.buccal_mucosa}{
   An ExpressionSet with 119 samples and 23,271 features specific to the buccal_mucosa bodysite
}

\subsection{HMP_2012.pathcoverage.hard_palate}{
   An ExpressionSet with 1 sample and 23,271 features specific to the hard_palate bodysite
}

\subsection{HMP_2012.pathcoverage.keratinized_gingiva}{
   An ExpressionSet with 6 samples and 23,271 features specific to the keratinized_gingiva bodysite
}

\subsection{HMP_2012.pathcoverage.l_retroauricular_crease}{
   An ExpressionSet with 9 samples and 23,271 features specific to the l_retroauricular_crease bodysite
}

\subsection{HMP_2012.pathcoverage.mid_vagina}{
   An ExpressionSet with 2 samples and 23,271 features specific to the mid_vagina bodysite
}

\subsection{HMP_2012.pathcoverage.palatine_tonsils}{
   An ExpressionSet with 6 samples and 23,271 features specific to the palatine_tonsils bodysite
}

\subsection{HMP_2012.pathcoverage.posterior_fornix}{
   An ExpressionSet with 62 samples and 23,271 features specific to the posterior_fornix bodysite
}

\subsection{HMP_2012.pathcoverage.r_retroauricular_crease}{
   An ExpressionSet with 18 samples and 23,271 features specific to the r_retroauricular_crease bodysite
}

\subsection{HMP_2012.pathcoverage.saliva}{
   An ExpressionSet with 5 samples and 23,271 features specific to the saliva bodysite
}

\subsection{HMP_2012.pathcoverage.stool}{
   An ExpressionSet with 147 samples and 23,271 features specific to the stool bodysite
}

\subsection{HMP_2012.pathcoverage.subgingival_plaque}{
   An ExpressionSet with 7 samples and 23,271 features specific to the subgingival_plaque bodysite
}

\subsection{HMP_2012.pathcoverage.supragingival_plaque}{
   An ExpressionSet with 128 samples and 23,271 features specific to the supragingival_plaque bodysite
}

\subsection{HMP_2012.pathcoverage.throat}{
   An ExpressionSet with 7 samples and 23,271 features specific to the throat bodysite
}

\subsection{HMP_2012.pathcoverage.tongue_dorsum}{
   An ExpressionSet with 136 samples and 23,271 features specific to the tongue_dorsum bodysite
}

\subsection{HMP_2012.pathcoverage.vaginal_introitus}{
   An ExpressionSet with 3 samples and 23,271 features specific to the vaginal_introitus bodysite
}
}
\examples{
HMP_2012.pathcoverage.anterior_nares()

}
\seealso{
\code{\link{HMP_2012.genefamilies_relab}}, \code{\link{HMP_2012.marker_abundance}}, \code{\link{HMP_2012.marker_presence}}, \code{\link{HMP_2012.metaphlan_bugs_list}}, \code{\link{HMP_2012.pathabundance_relab}}
}

