% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge a list of curatedMetagenomicData datasets}
\usage{
mergeData(obj, sampledelim = ":", studycolname = "studyID")
}
\arguments{
\item{obj}{A list or SimpleList containing a SummarizedExperiment-class object in each element.}

\item{sampledelim}{If a character vector of length one is provided, for example ":" (default) then
sample names in the merged SummarizedExperiment-class object will combine study
identifier with sample identifier in the form studyID:sampleID.
If not a character vector of length one, then sample names from the original
studies will be preserved. Can be set to NULL to keep the sample names of the
original studies.}

\item{studycolname}{If a character vector of length one is provided (default: studyID), a column with this
name will be added to the colData, containing study IDs taken from
the names of the SummarizedExperiment-class object.}
}
\value{
a SummarizedExperiment-class object
}
\description{
\code{mergeData} merges a list of SummarizedExperiment-class objects produced by the
\code{curatedMetagenomicData} function into a single SummarizedExperiment-class object. It is
recommended to use this functions only on a list of datasets of the same
data type (for example, all relative_abundance datasets).
}
\examples{
library(curatedMetagenomicData)
library(SummarizedExperiment)

resources <- curatedMetagenomicData(pattern = "AsnicarF.*relative_abundance", dryrun = TRUE)
data_sets <- lapply(resources, function(x) curatedMetagenomicData(x, print = FALSE, dryrun = FALSE))
names(data_sets) <- gsub("^.*\\\\.(.*)\\\\..*$", "\\\\1", resources)

merged_data_sets <- mergeData(data_sets)
merged_data_sets

relative_abundance <- assays(merged_data_sets)$relative_abundance
relative_abundance[1:10, 1:10]

metadata <- colData(merged_data_sets)
metadata
}
