% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Title}
\usage{
mergeData(x)
}
\arguments{
\item{x}{a list}
}
\value{
a \code{SummarizedExperiment} is returned
}
\description{
Description
}
\examples{
curatedMetagenomicData("AsnicarF_20.+.pathway_abundance", print = FALSE) |>
    sapply(curatedMetagenomicData, dryrun = FALSE) |>
    mergeData()

curatedMetagenomicData("AsnicarF_20.+.relative_abundance", print = FALSE) |>
    sapply(curatedMetagenomicData, dryrun = FALSE) |>
    mergeData()

}
