% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnSamples.R
\name{returnSamples}
\alias{returnSamples}
\title{Title}
\usage{
returnSamples(sampleMetadata, dataType, counts = FALSE)
}
\arguments{
\item{sampleMetadata}{description}

\item{dataType}{description}

\item{counts}{description}
}
\value{
description
}
\description{
Description
}
