% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMetadata.R
\docType{data}
\name{sampleMetadata}
\alias{sampleMetadata}
\title{Manually Curated Sample Metadata}
\format{
An object of class \code{data.frame} with 20534 rows and 136 columns.
}
\usage{
sampleMetadata
}
\description{
Manually curated sample metadata for all samples in curatedMetagenomicData.
}
\keyword{datasets}
